% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{all.same}
\alias{all.same}
\title{Test if all items in a vector or a list are identical.}
\usage{
\method{all}{same}(x)
}
\arguments{
\item{x}{a vector or a list}
}
\value{
\code{TRUE} if all elements of \code{x} are identical to each other.
}
\description{
Test if all items in a vector or a list are identical.
}
\examples{

stopifnot(!all.same(1:3))

stopifnot(all.same(list("a", "a", "a")))
}
\seealso{
\code{\link[base:identical]{base::identical()}}
}
