% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{all_same}
\alias{all_same}
\title{Test if all items in a vector or a list are identical.}
\usage{
all_same(x)
}
\arguments{
\item{x}{a vector or a list}
}
\value{
\code{TRUE} if all elements of \code{x} are identical to each other.
}
\description{
Test if all items in a vector or a list are identical.
}
\examples{

stopifnot(!all_same(1:3))

stopifnot(all_same(list("a", "a", "a")))
}
\seealso{
\code{\link[base:identical]{base::identical()}}
}
