% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.hotelling.diff.test.R
\name{ar.yw.mcmc.list}
\alias{ar.yw.mcmc.list}
\title{An ar.yw method for \code{\link{mcmc.list}} objects.}
\usage{
\method{ar.yw}{mcmc.list}(x, aic = TRUE, order.max = NULL,
  na.action = na.fail, demean = TRUE, series = NULL, ...)
}
\arguments{
\item{x}{an \code{\link{mcmc.list}} object.}

\item{aic, order.max, na.action, demean, series, ...}{See \code{\link[=ar]{ar()}}.}
}
\description{
An ar.yw method for \code{\link{mcmc.list}} objects.
}
