% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlebdm.R
\name{as.rlebdm.ergm}
\alias{as.rlebdm.ergm}
\title{Extract dyad-level ERGM constraint information from an \code{\link{ergm}} fit
into an \code{\link{rlebdm}} object}
\usage{
\method{as.rlebdm}{ergm}(x, ...)
}
\arguments{
\item{x}{an \code{\link{ergm}} fit.}

\item{...}{additional arguments passed on to
\code{\link[=as.rlebdm.ergm_conlist]{as.rlebdm.ergm_conlist()}}; note, in particular, the \verb{which=}
argument.}
}
\description{
This is a thin wrapper around \code{\link[=as.rlebdm.ergm_conlist]{as.rlebdm.ergm_conlist()}}.
}
\seealso{
\code{\link[=as.rlebdm.ergm_conlist]{as.rlebdm.ergm_conlist()}}
}
\keyword{internal}
