% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{asymmetric-ergmTerm}
\alias{asymmetric-ergmTerm}
\alias{InitErgmTerm.asymmetric}
\title{Asymmetric dyads}
\usage{
# binary: asymmetric(attr=NULL, diff=FALSE, keep=NULL, levels=NULL)
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.) If specified, only symmetric pairs that match on the vertex attribute are counted.}

\item{diff}{Used in the same way as for the \code{nodematch} term. (See \code{nodematch} (\code{ergmTerm?nodematch}) for details.)}

\item{keep}{deprecated}

\item{level}{Used in the same way as for the \code{nodematch} term. (See \code{nodematch} (\code{ergmTerm?nodematch}) for details.)}
}
\description{
This term adds one network statistic to the model equal to the
number of pairs of actors for which exactly one of
\eqn{(i{\rightarrow}j)}{(i,j)} or \eqn{(j{\rightarrow}i)}{(j,i)} exists.
}
\note{
This term can only be used with directed networks.

The argument \code{keep} is retained for backwards compatibility and may be
removed in a future version. When both \code{keep} and \code{levels} are passed,
\code{levels} overrides \code{keep}.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "asymmetric", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{triad-related}
