% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.diversity.R
\name{b1covrange-ergmTerm}
\alias{b1covrange-ergmTerm}
\alias{InitErgmTerm.b1covrange}
\title{Range of covariate values for neighbors of a mode-1 node}
\usage{
# binary: nodecovrange(attr)
}
\arguments{
\item{attr}{a vertex attribute specification (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}
}
\description{
This term adds a single network statistic equalling
the sum over the nodes of the range over of its neighbors'
values.
}
\details{
This is a network analogue of the statistic introduced by
\insertCite{HoBl23m;textual}{ergm}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "b1covrange", "subsection")}
}
\concept{bipartite}
\concept{quantitative nodal attribute}
