% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.degfactorial.R
\name{b1degm1lfactorial-ergmTerm}
\alias{b1degm1lfactorial-ergmTerm}
\alias{InitErgmTerm.b1degm1lfactorial}
\title{Sum of log-factorials of degrees of Mode-1 non-isolates minus 1}
\usage{
# binary: b1degm1lfactorial
}
\description{
This term adds one network statistic to the model,
equalling \eqn{\sum_{i=1}^b \mathbb{I}(d_i>0) \log(d_i - 1)!}.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "b1degm1lfactorial", "subsection")}
}
\concept{bipartite}
\concept{undirected}
