% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{b1sociality-ergmTerm}
\alias{b1sociality-ergmTerm}
\alias{InitErgmTerm.b1sociality}
\alias{InitWtErgmTerm.b1sociality}
\title{Degree}
\usage{
# binary: b1sociality(nodes=-1)

# valued: b1sociality(nodes=-1, form="sum")
}
\arguments{
\item{node}{for level specification. The attribute and the sorted
unique values are the vector of vertex indices \code{1:nb1} , where
\code{nb1} is the size of the first bipartition. By default, \code{nodes=-1} means that the statistic for the
first node will be omitted, but this argument may be changed to control
which statistics are included. See Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}
}
\description{
Degree
}
\details{
This term adds one network statistic for each node in the first bipartition, equal to the number of
ties of that node. This term can only be used with bipartite networks. For directed networks, see \code{sender} and
\code{receiver}. For unipartite networks, see \code{sociality}.
}
\concept{bipartite}
\concept{dyad-independent}
\concept{undirected}
