% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{b1twostar-ergmTerm}
\alias{b1twostar-ergmTerm}
\alias{InitErgmTerm.b1twostar}
\title{Two-star census for central nodes centered on the first mode of a bipartite network}
\usage{
# binary: b1twostar(b1attr, b2attr, base=NULL, b1levels=NULL, b2levels=NULL, levels2=NULL)
}
\arguments{
\item{b1attr}{b1 nodes (actors in some contexts) (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details)}

\item{b2attr}{b2 nodes (events in some contexts). If \code{b2attr} is not passed, it is assumed to be the same as \code{b1attr} .}

\item{base}{This argument is retained for backwards compatibility and may be
removed in a future version. When both \code{base} and \code{levels2} are passed,
\code{levels2} overrides \code{base}.}
}
\description{
Two-star census for central nodes centered on the first mode of a bipartite network
}
\details{
This term takes two nodal attributes. Assuming that there are
\eqn{n_1} values of \code{b1attr} among the b1 nodes and \eqn{n_2}
values of \code{b2attr} among the b2 nodes, then the total number of
distinct categories of two stars according to these two attributes is
\eqn{n_1(n_2)(n_2+1)/2}. By default, this model term creates a distinct statistic
counting each of these categories. The \code{b1levels} , \code{b2levels} , \code{base} , and \code{levels2} arguments
may be used to leave some of these categories out (seefor details).
}
\concept{bipartite}
\concept{categorical nodal attribute}
\concept{undirected}
