% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{b2factor-ergmTerm}
\alias{b2factor-ergmTerm}
\alias{InitErgmTerm.b2factor}
\alias{InitWtErgmTerm.b2factor}
\title{Factor attribute effect for the second mode in a bipartite (aka two-mode) network}
\usage{
# binary: b2factor(attr, base=1, levels=-1)

# valued: b2factor(attr, base=1, levels=-1, form="sum")
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{base}{deprecated}

\item{levels}{this optional argument controls which levels of the attribute
should be included and which should be excluded. (See Specifying Vertex
attributes and Levels (\code{?nodal_attributes}) for details.) For example, if
the "fruit" attribute has levels "orange", "apple", "banana", and "pear",
then to include just two levels, one for "apple" and one for "pear", use
any of \code{b2factor("fruit", levels=-(2:3))}, \code{b2factor("fruit", levels=c(1,4))}, and
\code{b2factor("fruit", levels=c("apple", "pear"))}. Note: if you are using numeric
values to specify the levels of a character variable, the levels will
correspond to the alphabetically sorted character levels.}

\item{form}{character how to aggregate tie values in a valued ERGM}
}
\description{
Factor attribute effect for the second mode in a bipartite (aka two-mode) network
}
\details{
This term adds multiple network statistics to the model, one for each of (a subset of) the
unique values of the \code{attr} attribute. Each of these statistics
gives the number of times a node with that attribute in the second mode of
the network appears in an edge. The second mode of a bipartite network
object is sometimes known as the "event" mode.
}
\note{
To include all attribute values is usually not a good idea, because
the sum of all such statistics equals the number of edges and hence a linear
dependency would arise in any model also including \code{edges}. The default,
\code{levels=-1}, is therefore to omit the first (in lexicographic order)
attribute level. To include all levels, pass either \code{levels=TRUE}
(i.e., keep all levels) or \code{levels=NULL} (i.e., do not filter levels).

The argument \code{base} is retained for backwards compatibility and may be
removed in a future version. When both \code{base} and \code{levels} are passed,
\code{levels} overrides \code{base}.

This term can only be used with undirected bipartite networks.
}
\concept{bipartite}
\concept{categorical nodal attribute}
\concept{dyad-independent}
\concept{frequently-used}
\concept{undirected}
