% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.bipartite.degree.R
\name{b2mindegree-ergmTerm}
\alias{b2mindegree-ergmTerm}
\alias{InitErgmTerm.b2mindegree}
\title{Minimum degree for the second mode in a bipartite network}
\usage{
# binary: b2mindegree(d)
}
\arguments{
\item{d}{a vector of distinct integers}
}
\description{
Minimum degree for the second mode in a bipartite (aka two-mode) network
}
\details{
This term adds one network statistic to the model for
each element in \code{d} ; the \eqn{i} th such statistic equals the number of
nodes in the second mode of a bipartite network with at least degree \code{d[i]} .
The second mode of a bipartite network object is sometimes known as the "event" mode.
}
\note{
This term can only be used with undirected bipartite networks.
}
\concept{bipartite}
\concept{undirected}
