% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{b2starmix-ergmTerm}
\alias{b2starmix-ergmTerm}
\alias{InitErgmTerm.b2starmix}
\title{Mixing matrix for k-stars centered on the second mode of a bipartite network}
\usage{
# binary: b2starmix(k, attr, base=NULL, diff=TRUE)
}
\description{
Mixing matrix for k-stars centered on the second mode of a bipartite network
}
\details{
This term is exactly the same as \code{b1starmix} except that the roles of
b1 and b2 are reversed.
}
\concept{bipartite}
\concept{categorical nodal attribute}
\concept{undirected}
