% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{b2twostar-ergmTerm}
\alias{b2twostar-ergmTerm}
\alias{InitErgmTerm.b2twostar}
\title{Two-star census for central nodes centered on the second mode of a bipartite network}
\usage{
# binary: b2twostar(b1attr, b2attr, base=NULL, b1levels=NULL, b2levels=NULL, levels2=NULL)
}
\description{
Two-star census for central nodes centered on the second mode of a bipartite network
}
\details{
This term is exactly the same as \code{b1twostar} except that the
roles of b1 and b2 are reversed.
}
\concept{bipartite}
\concept{categorical nodal attribute}
\concept{undirected}
