% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{balance-ergmTerm}
\alias{balance-ergmTerm}
\alias{InitErgmTerm.balance}
\title{Balanced triads}
\usage{
# binary: balance
}
\description{
This term adds one network statistic to the model equal to the number of
triads in the network that are balanced. The balanced triads are those of
type \code{102} or \code{300} in the categorization of Davis and Leinhardt (1972). For details on the 16 possible triad types, see
\code{?triad.classify} in the \code{{sna}} package. For an undirected
network, the balanced triads are those with an odd number of ties (i.e., 1
and 3).
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{triad-related}
\concept{undirected}
