% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{bd-ergmConstraint}
\alias{bd-ergmConstraint}
\alias{InitErgmConstraint.bd}
\title{Constrain maximum and minimum vertex degree}
\usage{
# bd(attribs, maxout, maxin, minout, minin)
}
\arguments{
\item{attribs}{a matrix of logicals with dimension \verb{(n_nodes, attrcount)} for the attributes on which we are
conditioning, where \code{attrcount} is the number of distinct attributes values to condition on.}

\item{maxout, maxin, minout, minin}{matrices of alter attributes with the same dimension as \code{attribs} when used
in conjunction with \code{attribs}. Otherwise, vectors of integers specifying the relevant limits.
If the vector is of length 1, the limit is applied to all nodes. If an individual entry is \code{NA},
then there is no restriction of that kind is applied.}
}
\description{
Condition on the number of inedge or outedges posessed by a node.
See Placing Bounds on Degrees section for more information. (\code{\link[=ergmConstraint]{?ergmConstraint}})
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{directed}
\concept{undirected}
