% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{BDStratTNT-ergmProposal}
\alias{BDStratTNT-ergmProposal}
\alias{InitErgmProposal.BDStratTNT}
\title{TNT proposal with degree bounds}
\description{
Implements a TNT proposal with any subset of the following features:
\enumerate{
\item upper on degree, specified via the \code{\link[=bd-ergmConstraint]{bd}} constraint's \code{maxout}, \code{maxin}, and \code{attribs} arguments
\item stratification of proposals according to mixing type on a vertex attribute, specified via the \code{\link[=strat-ergmHint]{strat}} hint;
\item fixation of specified mixing types on a(nother) vertex attribute, specified via the \code{\link[=blocks-ergmConstraint]{blocks}} constraint.
}
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="BDStratTNT"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="BDStratTNT"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="BDStratTNT"))}}
}
\keyword{internal}
