% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_model.R, R/ergm_proposal.R
\name{call.ErgmTerm}
\alias{call.ErgmTerm}
\alias{call.ErgmConstraint}
\alias{call.ErgmReference}
\title{Locate and call an ERGM term and term-alike initialization functions.}
\usage{
call.ErgmTerm(term, env, nw, ..., term.options = list())

call.ErgmConstraint(term, env, nw, ..., term.options = list(), consign = +1L)

call.ErgmReference(term, env, nw, ..., term.options = list())
}
\arguments{
\item{term}{A term from an \code{\link[=ergm]{ergm()}} model, constraint, or reference
formula: typically a \code{\link{name}} or a \code{\link{call}}.}

\item{env}{Environment in which it is to be evaluated.}

\item{nw}{A \code{\link[network:network]{network}} object.}

\item{...}{Additional term options.}

\item{term.options}{A list of optional settings such as calculation
tuning options to be passed to the appropriate \verb{InitErgm*} or
\verb{InitWtErgm*} functions.}

\item{consign}{the sign of the constraint in the formula.}
}
\value{
The list returned by the the \verb{InitErgm*} or \verb{InitWtErgm*}
function with appropriate metadata set.
}
\description{
This is a family of helper functions that searche attached and
loaded packages for an initialization function with the specified
name, calls it with the specified arguments, and returns the
result.
}
\keyword{internal}
