% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ergm.R
\name{coef.ergm}
\alias{coef.ergm}
\alias{coefficients.ergm}
\alias{vcov.ergm}
\title{Extract Model Fit Coefficients and Uncertainty Estimates}
\usage{
\method{coef}{ergm}(object, ...)

\method{coefficients}{ergm}(object, ...)

\method{vcov}{ergm}(object, sources = c("all", "model", "estimation"), ...)
}
\arguments{
\item{object}{{an object for which the extraction of model coefficients is
meaningful.}}

\item{...}{{other arguments.}}

\item{sources}{{Specify whether to return the covariance matrix
from the ERGM model, the estimation process, or both combined.}}
}
\value{
Coefficients extracted from the model object \code{object}.
}
\description{
\code{coef} extracts model coefficients from objects returned by
the \code{\link{ergm}} function.

\code{coefficients} is an \emph{alias} for \coef{ergm}.

\code{vcov} extracts the variance-covariance matrix of parameter
estimates.
}
\examples{
data(florentine)
fit <- ergm(flomarriage ~ edges + concurrent)
coef(fit)

vcov(fit, sources="model")
vcov(fit, sources="estimation")
vcov(fit, sources="all") # the default
}
\seealso{
\code{\link{fitted.values}} and \code{\link{residuals}} for related methods;
\code{\link{glm}}, \code{\link{lm}} for model fitting.
}
\keyword{models}
\keyword{regression}
