% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_model.utils.R
\name{coef.sublength.model}
\alias{coef.sublength.model}
\alias{coef.length.model}
\title{Extract Number of parameters in ergm Model}
\usage{
\method{coef}{sublength.model}(object, ...)

\method{coef}{length.model}(object, ...)
}
\arguments{
\item{object}{an ergm model object}

\item{\dots}{other arguments.}
}
\value{
\code{coef.sublength.model} returns a vector containing the number
of model parameters corresponding to each model term.

\code{coef.length.model} returns the length of the
parameter vector of the model.
}
\description{
\code{coef.sublength} and \code{coef.length} are methods that extract the
numbers of parameters for ergm model objects.
}
\note{
These are \emph{not} methods at this time. This may change in the future.
}
\keyword{models}
