% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.model.utils.R
\name{coef.names.model}
\alias{coef.names.model}
\title{Parameters names of an initialized ergm model}
\usage{
\method{coef}{names.model}(object, canonical)
}
\arguments{
\item{object}{an \code{ergm.model}.}

\item{canonical}{whether the canonical (natural) parameters or the
model (curved) parameters are wanted.}
}
\value{
a character vector of parameter names.
}
\description{
Parameters names of an initialized ergm model
}
