% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.diagnostics.ergm.R
\name{colMeans.mcmc.list}
\alias{colMeans.mcmc.list}
\alias{sweep.mcmc.list}
\title{Utility operations for \code{\link{mcmc.list}} objects}
\usage{
colMeans.mcmc.list(x, ...)

sweep.mcmc.list(x, STATS, FUN = "-", check.margin = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{mcmc.list}} object.}

\item{\dots}{additional arguments to \code{\link{colMeans}} or
\code{\link{sweep}}.}

\item{STATS, FUN, check.margin}{See help for \code{\link[=sweep]{sweep()}}.}
}
\value{
\code{colMeans.mcmc} returns a vector with length equal to
the number of mcmc chains in \code{x} with the mean value for
each chain.

\code{sweep.mcmc.lists} returns an appropriately modified
version of \code{x}
}
\description{
\code{colMeans.mcmc.list} is a "method" for (non-generic) \code{\link[=colMeans]{colMeans()}} applicable to \code{\link{mcmc.list}} objects.

\code{sweep.mcmc.list} is a "method" for (non-generic)
\code{\link[=sweep]{sweep()}} applicable to \code{\link{mcmc.list}} objects.
}
\seealso{
\code{\link[=colMeans]{colMeans()}}

\code{\link[=sweep]{sweep()}}
}
