% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{concurrent-ergmTerm}
\alias{concurrent-ergmTerm}
\alias{InitErgmTerm.concurrent}
\title{Concurrent node count}
\usage{
# binary: concurrent(by=NULL, levels=NULL)
}
\arguments{
\item{by}{this optional argument specifies a vertex attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)
It functions just like the \code{by} argument of the \code{degree} term.}
}
\description{
This term adds one network statistic to the model, equal to the number of
nodes in the network with degree 2 or higher.
This term can only be used with undirected
networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{categorical nodal attribute}
\concept{undirected}
