% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.extra.R
\name{concurrentties-ergmTerm}
\alias{concurrentties-ergmTerm}
\alias{InitErgmTerm.concurrentties}
\title{Concurrent tie count}
\usage{
# binary: concurrentties(by=NULL, levels=NULL)
}
\arguments{
\item{by}{a vertex attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.);
it functions just like the \code{by} argument of the \code{degree} term}

\item{levels}{TODO (See Specifying Vertex
attributes and Levels (\code{\link[=nodal_attributes]{?nodal_attributes}}) for details.)}
}
\description{
Concurrent tie count
}
\details{
This term adds one network statistic to the model, equal to the number of
ties incident on each actor beyond the first.
This term can only be used with undirected networks.
}
\concept{categorical nodal attribute}
\concept{undirected}
