% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondB1Degree-ergmProposal}
\alias{CondB1Degree-ergmProposal}
\alias{InitErgmProposal.CondB1Degree}
\title{MHp for b1degree constraints}
\description{
MHp for \code{constraints= ~b1degrees}. For bipartite networks, randomly select an edge \eqn{(B_{1i},B_{2j})}
and an empty dyad with the same node B1i, \eqn{(B_{1i},B_{2k})}, and propose to toggle both \eqn{(B_{1i},B_{2j})} and \eqn{(B_{1i},B_{2k})}.
This ensures that the degrees of individual nodes in mode 1 are preserved.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondB1Degree"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondB1Degree"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondB1Degree"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "CondB1Degree", "subsection")}
}
\keyword{internal}
