% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondB2Degree-ergmProposal}
\alias{CondB2Degree-ergmProposal}
\alias{InitErgmProposal.CondB2Degree}
\title{MHp for b2degree constraints}
\description{
MHp for \eqn{constraints= ~b2degrees}. For bipartite networks, randomly select an edge {B1j, B2i}
and an empty dyad with the same node B2i, {B1k, B2i}, and propose to toggle both {B1j, B2i} and {B1k, B2i}.
This ensures that the degrees of individual nodes in mode 2 are preserved.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondB2Degree"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondB2Degree"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondB2Degree"))}}
}
