% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondDegreeDist-ergmProposal}
\alias{CondDegreeDist-ergmProposal}
\alias{InitErgmProposal.CondDegreeDist}
\title{MHp for degreedist constraints}
\description{
MHp for \eqn{constraints= ~degreedist}. Randomly select a node (T) and its edge (E).  If the head
node of the edge (H) has 1 degree more than another randomly select node (A), and A is disconnected to both
T and H, then propose to toggle E and the dyad between T and A.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondDegreeDist"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondDegreeDist"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondDegreeDist"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "CondDegreeDist", "subsection")}
}
\keyword{internal}
