% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondDegreeMix-ergmProposal}
\alias{CondDegreeMix-ergmProposal}
\alias{InitErgmProposal.CondDegreeMix}
\title{MHp for degree mix constraints}
\description{
MHp for \eqn{constraints= ~degreesmix}. Similar to \code{InitErgmProposal.CondDegree}, except that
the toggle is proposed only if the mixing matrix of degrees is preserved before and after the toggle.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondDegreeMix"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondDegreeMix"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondDegreeMix"))}}
}
