% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondInDegree-ergmProposal}
\alias{CondInDegree-ergmProposal}
\alias{InitErgmProposal.CondInDegree}
\title{MHp for idegree constraints}
\description{
MHp for \code{constraints= ~idegrees}. For directed networks, randomly select two dyads with a
common head node, one having an edge one not, and propose to swap the tie from one tail to the other.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondInDegree"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondInDegree"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondInDegree"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "CondInDegree", "subsection")}
}
\keyword{internal}
