% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondInDegreeDist-ergmProposal}
\alias{CondInDegreeDist-ergmProposal}
\alias{InitErgmProposal.CondInDegreeDist}
\title{MHp for idegreedist constraints}
\description{
MHp for \code{constraints= ~idegreedist}. For directed networks, similar to
\code{InitErgmProposal.CondDegreeDist}, except for indegree case
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondInDegreeDist"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondInDegreeDist"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondInDegreeDist"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "CondInDegreeDist", "subsection")}
}
\keyword{internal}
