% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondInDegreeDist-ergmProposal}
\alias{CondInDegreeDist-ergmProposal}
\alias{InitErgmProposal.CondInDegreeDist}
\title{MHp for idegreedist constraints}
\description{
MHp for \eqn{constraints= ~idegreedist}. For directed networks, similar to
\code{InitErgmProposal.CondDegreeDist}, except for indegree case
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.parseProposal(proposal="CondInDegreeDist"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.parseProposal(proposal="CondInDegreeDist"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.parseProposal(proposal="CondInDegreeDist"))}}
}
