% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondOutDegree-ergmProposal}
\alias{CondOutDegree-ergmProposal}
\alias{InitErgmProposal.CondOutDegree}
\title{MHp for odegree constraints}
\description{
MHp for \code{constraints= ~odegrees}. For directed networks, randomly select two dyads with a
common tail node, one having an edge and one not, and propose to swap the tie from one head to the other.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondOutDegree"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondOutDegree"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondOutDegree"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "CondOutDegree", "subsection")}
}
\keyword{internal}
