% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{ConstantEdges-ergmProposal}
\alias{ConstantEdges-ergmProposal}
\alias{InitErgmProposal.ConstantEdges}
\title{MHp for edges constraints}
\description{
MHp for \code{constraints= ~edges}. Propose pairs of toggles that keep number of edges the same.
This is done by:
\enumerate{
\item choosing an existing edge at random;
\item choosing a dyad at random that does not have an edge; and
\item proposing toggling both these dyads.
}
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="ConstantEdges"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="ConstantEdges"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="ConstantEdges"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "ConstantEdges", "subsection")}
}
\keyword{internal}
