% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.transitiveties.R,
%   R/InitWtErgmTerm.R
\name{cyclicalties-ergmTerm}
\alias{cyclicalties-ergmTerm}
\alias{InitErgmTerm.cyclicalties}
\alias{InitWtErgmTerm.cyclicalties}
\title{Cyclical ties}
\usage{
# binary: cyclicalties(attr=NULL, levels=NULL)

# valued: cyclicalties(threshold=0)
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.) If set, all three nodes involved ( \eqn{i} , \eqn{j} , and the node on the two-path) must match
on this attribute in order for \eqn{i\rightarrow j}{i-->j} to be counted.}

\item{levels}{TODO}
}
\description{
Cyclical ties
}
\details{
This term adds one statistic, equal to the number of ties
\eqn{i\rightarrow j}{i-->j} such that there exists a two-path from
\eqn{j} to \eqn{i} . (Related to the \code{ttriple} term.)
}
\concept{directed}
\concept{undirected}
