% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{degcor-ergmTerm}
\alias{degcor-ergmTerm}
\alias{InitErgmTerm.degcor}
\title{Degree Correlation}
\usage{
# binary: degcor
}
\description{
Degree Correlation
}
\details{
This term adds one network statistic equal to the correlation
of the degrees of all pairs of nodes in the network which are tied.
Only coded for undirected networks.
}
\concept{undirected}
