% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{degcrossprod-ergmTerm}
\alias{degcrossprod-ergmTerm}
\alias{InitErgmTerm.degcrossprod}
\title{Degree Cross-Product}
\usage{
# binary: degcrossprod
}
\description{
Degree Cross-Product
}
\details{
This term adds one network statistic equal to the mean of the cross-products
of the degrees of all pairs of nodes in the network which are tied.
Only coded for undirected networks.
}
\concept{undirected}
