% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{degree1.5-ergmTerm}
\alias{degree1.5-ergmTerm}
\alias{InitErgmTerm.degree1.5}
\title{Degree to the 3/2 power}
\usage{
# binary: degree1.5
}
\description{
This term adds one network statistic to the model equaling the sum over
the actors of each actor's degree taken to the 3/2 power (or,
equivalently, multiplied by its square root). This term is an
undirected analog to the terms of \insertCite{SnVa10i;textual}{ergm}, equations
(11) and (12). This term can only be used with undirected networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "degree1.5", "subsection")}
}
\concept{undirected}
