% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{degreedist-ergmConstraint}
\alias{degreedist-ergmConstraint}
\alias{InitErgmConstraint.degreedist}
\title{Preserve the degree distribution of the given network}
\usage{
# degreedist
}
\description{
Only networks
whose degree distributions are the same as those in the network passed
in the model formula have non-zero probability.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{directed}
\concept{undirected}
