% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{degreepopularity-ergmTerm}
\alias{degreepopularity-ergmTerm}
\alias{InitErgmTerm.degreepopularity}
\title{Degree popularity (deprecated)}
\usage{
# binary: degreepopularity
}
\description{
Degree popularity (deprecated)
}
\details{
see \code{degree1.5} .
}
\concept{deprecated}
\concept{undirected}
