% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{density-ergmTerm}
\alias{density-ergmTerm}
\alias{InitErgmTerm.density}
\title{Density}
\usage{
# binary: density
}
\description{
This term adds one network statistic equal to the density of the network.
For undirected networks, \code{density} equals \code{kstar(1)} or
\code{edges} divided by \eqn{n(n-1)/2} ; for directed networks,
\code{density} equals \code{edges} or \code{istar(1)} or \code{ostar(1)}
divided by \eqn{n(n-1)} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
