% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{direct.network}
\alias{direct.network}
\title{Returns a directed version of an undirected binary network}
\usage{
direct.network(x, rule = c("both", "upper", "lower"))
}
\arguments{
\item{x}{a \code{\link{network}} object.}

\item{rule}{a string specifying how the network is to be
constructed.}
}
\description{
Returns a directed version of an undirected binary network
}
