% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{DiscUnif-ergmProposal}
\alias{DiscUnif-ergmProposal}
\alias{InitWtErgmProposal.DiscUnif}
\title{TODO}
\description{
TODO
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.parseProposal(proposal="DiscUnif"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.parseProposal(proposal="DiscUnif"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.parseProposal(proposal="DiscUnif"))}}
}
