% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{DiscUnif-ergmReference}
\alias{DiscUnif-ergmReference}
\alias{InitErgmReference.DiscUnif}
\title{Discrete Uniform reference}
\usage{
# DiscUnif(a,b)
}
\arguments{
\item{a, b}{minimum and maximum to the baseline discrete uniform distribution, both inclusive. Both values must be finite.}
}
\description{
Specifies each dyad's baseline distribution to be discrete uniform
between \code{a} and \code{b} (both inclusive): \eqn{h(y)=1} , with
the support being
\code{a}, \code{a+1}, \ldots, \code{b-1}, \code{b}.
}
\seealso{
\code{\link{ergmReference}} for index of reference distributions currently visible to the package.
}
\concept{discrete}
\concept{finite}
