% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{DiscUnifNonObserved-ergmProposal}
\alias{DiscUnifNonObserved-ergmProposal}
\alias{InitWtErgmProposal.DiscUnifNonObserved}
\title{TODO}
\description{
TODO
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="DiscUnifNonObserved"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="DiscUnifNonObserved"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="DiscUnifNonObserved"))}}
}
