% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{Dist-ergmProposal}
\alias{Dist-ergmProposal}
\alias{InitWtErgmProposal.Dist}
\title{Random toggle for a number of reference distributions}
\description{
Implements the \code{DyadGen} API and the
\ergmConstraint{ergm}{ChangeStats}{()} API.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="Dist"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="Dist"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="Dist"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "Dist", "subsection")}
}
\keyword{internal}
