% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_attr.R
\name{dyad_attr}
\alias{dyad_attr}
\alias{set.dyad.attribute}
\alias{set.dyad.attribute.network}
\alias{\%d\%<-}
\alias{get.dyad.attribute}
\alias{get.dyad.attribute.network}
\alias{\%d\%}
\title{Get and set dyad attribute matrices robust to transformations}
\usage{
set.dyad.attribute(x, attrname, value, ...)

\method{set.dyad.attribute}{network}(x, attrname, value, ...)

`\%d\%`(x, attrname) <- value

get.dyad.attribute(x, attrname)

\method{get.dyad.attribute}{network}(x, attrname)

x \%d\% attrname
}
\arguments{
\item{x}{an object capable of having vertex and network attributes.}

\item{attrname}{name of an attribute; uses the "namespace" of network attributes.}

\item{value}{matrix of appropriate dimension.}

\item{...}{additional arguments to methods.}
}
\description{
Get and set dyad attribute matrices robust to transformations
}
\section{Functions}{
\itemize{
\item \code{set.dyad.attribute()}: Set a dyad attribute.

\item \code{`\%d\%`(x, attrname) <- value}: Alias for \code{set.dyad.attribute(x, attrname, value)}.

\item \code{get.dyad.attribute()}: Get a dyad attribute, adjusting for any transformations such as taking of a subgraph.

\item \code{x \%d\% attrname}: Alias for \code{get.dyad.attribute(x, attrname)}.

}}
