% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{dyadnoise-ergmConstraint}
\alias{dyadnoise-ergmConstraint}
\alias{InitErgmConstraint.dyadnoise}
\title{A soft constraint to adjust the sampled distribution for
dyad-level noise with known perturbation probabilities}
\usage{
# dyadnoise(p01, p10)
}
\arguments{
\item{p01, p10}{can both be scalars or both be adjacency matrices of the same dimension as that of the
LHS network giving these probabilities.}
}
\description{
It is assumed that the observed LHS network is a noisy observation of
some unobserved true network, with \code{p01} giving the dyadwise
probability of erroneously observing a tie where the true network
had a non-tie and \code{p10} giving the dyadwise probability of
erroneously observing a nontie where the true network had a tie.
}
\note{
See \insertCite{KaKr16s;textual}{ergm} for an application.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "dyadnoise", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{soft}
\concept{undirected}
