\name{ecoli}
\docType{data}
\title{Goodreau's Faux Magnolia High School as a network object}
\alias{ecoli}
\alias{ecoli1}
\alias{ecoli2}
\description{
This network data set comprises two versions of a biological
network in which the nodes are operons in \emph{Escherichia Coli}
and a directed edge from one node to another indicates that 
the first encodes the transcription factor that regulates the second.

}

\usage{data(ecoli)}


\source{
The data set is based on the RegulonDB network (Salgado et al, 2001)
and was modified by Shen-Orr et al (2002).  

}


\references{

Salgado et al (2001)

Shen-Orr et al (2002)

Saul and Filkov (2007)

Hummel et al (2010)

}

\keyword{datasets}
\section{Licenses and Citation}{
When publishing results obtained using this data set, the original authors 
(Salgado et al, 2001; Shen-Orr et al, 2002) should
be cited, along with this \code{R} package.
}
