% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{edges-ergmConstraint}
\alias{edges-ergmConstraint}
\alias{InitErgmConstraint.edges}
\title{Preserve the edge count of the given network}
\usage{
# edges
}
\description{
Only networks
having the same number of edges as the network passed
in the model formula have non-zero probability.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "edges", "subsection")}
}
