% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{edges-ergmConstraint}
\alias{edges-ergmConstraint}
\alias{InitErgmConstraint.edges}
\title{Preserve the edge count of the given network}
\usage{
# edges
}
\description{
Preserve the edge count of the given network
}
\details{
Only networks
having the same number of edges as the network passed
in the model formula have non-zero probability.
}
