% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{edges-ergmTerm}
\alias{edges-ergmTerm}
\alias{InitErgmTerm.edges}
\alias{InitWtErgmTerm.nonzero}
\alias{InitWtErgmTerm.edges}
\title{Edges}
\usage{
# binary: edges

# valued: nonzero

# valued: edges
}
\description{
Number of edges in the network.
}
\details{
This term adds one network statistic equal to the number of
edges (i.e. nonzero values) in the network. For undirected networks, \code{edges}
is equal to \code{kstar(1)}; for directed networks, edges is equal to both
\code{ostar(1)} and \code{istar(1)}.
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
