% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/ergm-disambiguation.R
\name{edges}
\alias{edges}
\alias{edges2}
\title{Edges}
\usage{
# edges
}
\description{
This term adds one network statistic equal to the number of
edges (i.e. nonzero values) in the network. For undirected networks, \code{edges}
is equal to \code{kstar(1)}; for directed networks, edges is equal to both
\code{ostar(1)} and \code{istar(1)}.

\code{edges} may refer to:
\itemize{
\item \link[=edges-term]{An ERGM statistic} (\code{help("edges-term")})
\item \link[=edges-constraint]{An ERGM sample space constraint} (\code{help("edges-constraint")})
}
}
\seealso{
Other binary: 
\code{\link{absdiff}}

Other dyad-independent: 
\code{\link{absdiff}}

Other directed: 
\code{\link{absdiff}}

Other undirected: 
\code{\link{absdiff}}
}
\concept{binary}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
\concept{valued}
