% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.update.R
\name{empty_network}
\alias{empty_network}
\title{Create an empty copy of a network object}
\usage{
empty_network(nw, ignore.nattr = c("bipartite", "directed", "hyper",
  "loops", "mnext", "multiple", "n"), ignore.vattr = c())
}
\arguments{
\item{nw}{a \code{\link{network}} object}

\item{ignore.nattr}{character vector of the names of network-level
attributes to ignore when updating network objects (defaults to
standard network properties)}

\item{ignore.vattr}{character vector of the names of vertex-level
attributes to ignore when updating network objects}
}
\description{
Initializes an empty network with the same vertex and network
attributes as the original network, but no edges.
}
