%  File man/ergm-defunct.Rd in package ergm, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2017 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{ergm-defunct}
\alias{ergm-defunct}
\alias{delete.isolates}
\alias{central.network}
\alias{largest.components}
\alias{sociality.default}
\alias{sociality.ergm}
\alias{sociality.formula}
\alias{sociality.network}
\alias{rspartnerdist}
\alias{espartnerdist}
\alias{dspartnerdist}
\alias{twopathdist}
\alias{ostar2deg}
\alias{drawpie}
\alias{invert.network}
\alias{is.invertible}
\alias{mvmodel}
\alias{mvmodel.default}
\alias{mvmodel.ergm}
\alias{mvmodel.formula}
\alias{ergm.mahalanobis}
\title{
Functions that have been removed from this package
}
\description{
Functions that have been removed after a period of deprecation.

\itemize{
\item{delete.isolates}
\item{central.network}
\item{largest.components}
\item{sociality.default}
\item{sociality.ergm}
\item{sociality.formula}
\item{sociality.network}
\item{rspartnerdist}
\item{espartnerdist}
\item{dspartnerdist}
\item{twopathdist}
\item{ostar2deg}
\item{drawpie}
\item{invert.network}
\item{is.invertible}
\item{mvmodel}
\item{mvmodel.default}
\item{mvmodel.ergm}
\item{mvmodel.formula}
\item{ergm.mahalanobis}
}
}
