\name{ergm-internal}
\title{Internal ergm Objects}
% terms
\alias{ergm.checkargs}
\alias{ergm.checkdirected}
\alias{ergm.checkdegeneracy}
\alias{ergm.checkbipartite}
%
\alias{ConstraintImplications}
\alias{InitConstraint.atleast}
\alias{InitConstraint.atmost}
\alias{InitConstraint.bd}
\alias{InitConstraint.degreedist}
\alias{InitConstraint.degreesHexad}
\alias{InitConstraint.degreessimple}
\alias{InitConstraint.degreesTetrad}
\alias{InitConstraint.degrees}
\alias{InitConstraint.edges}
\alias{InitConstraint.hamming}
\alias{InitConstraint.indegreedist}
\alias{InitConstraint.outdegreedist}
\alias{InitConstraint.nodedegrees}
\alias{InitConstraint.observed}
\alias{InitConstraint.ranks}
%

\alias{InitErgm.D.dyad}
\alias{InitErgm.D.edge}
\alias{InitErgm.D.off}
\alias{InitErgm.altistar}
\alias{InitErgm.altostar}
\alias{InitErgm.b1kappa}
\alias{InitErgm.b1share}
\alias{InitErgm.b2kappa}
\alias{InitErgm.b2share}
\alias{InitErgm.bichange}
\alias{InitErgm.biduration}
\alias{InitErgm.biendure}
\alias{InitErgm.bounded.degree}
\alias{InitErgm.bounded.idegree}
\alias{InitErgm.bounded.istar}
\alias{InitErgm.bounded.kstar}
\alias{InitErgm.bounded.odegree}
\alias{InitErgm.bounded.ostar}
\alias{InitErgm.bounded.triangle}
\alias{InitErgm.degreep}
\alias{InitErgm.dissolve}
\alias{InitErgm.duration}
\alias{InitErgm.factor}
\alias{InitErgm.formation}
\alias{InitErgm.geodegree}
\alias{InitErgm.gwb1706}
\alias{InitErgm.gwb1share}
\alias{InitErgm.gwb2706}
\alias{InitErgm.gwb2share}
\alias{InitErgm.gwdegree706}
\alias{InitErgm.gwdegreealpha}
\alias{InitErgm.gwdegreelambda}
\alias{InitErgm.hamming.weighted}
\alias{InitErgm.hammingmix.constant}
\alias{InitErgm.heideriandynamic}
\alias{InitErgm.hiertriad}
\alias{InitErgm.hiertriaddegree}
\alias{InitErgm.icvar}
\alias{InitErgm.idc}
\alias{InitErgm.intransitivedynamic}
\alias{InitErgm.intransitivity}
\alias{InitErgm.kappa}
\alias{InitErgm.monopolymixmat}
\alias{InitErgm.simmeliandynamic}
\alias{InitErgm.spatial}
\alias{InitErgm.transitivedynamic}
\alias{InitErgm.transitivity}

\alias{check.ErgmTerm}
\alias{assignvariables}
\alias{InitErgmTerm.absdiff}
\alias{InitErgmTerm.absdiffcat}
\alias{InitErgmTerm.adegcor}
\alias{InitErgmTerm.altkstar}
\alias{InitErgmTerm.asymmetric}
\alias{InitErgmTerm.b1degree}   
\alias{InitErgmTerm.b1star}
\alias{InitErgmTerm.b1starmix}
\alias{InitErgmTerm.b1twostar}
\alias{InitErgmTerm.b2degree}
\alias{InitErgmTerm.b2star}
\alias{InitErgmTerm.b2starmix}
\alias{InitErgmTerm.b2twostar}
\alias{InitErgmTerm.cycle}
\alias{InitErgmTerm.degcor}
\alias{InitErgmTerm.degcrossprod}
\alias{InitErgmTerm.edgecov}
\alias{InitErgmTerm.hamming}
\alias{InitErgmTerm.isolates}
\alias{InitErgmTerm.match}
\alias{InitErgmTerm.mutual}
\alias{InitErgmTerm.nodefactor}
\alias{InitErgmTerm.nodeifactor}
\alias{InitErgmTerm.nodematch}
\alias{InitErgmTerm.nodemix}
\alias{InitErgmTerm.nodeofactor}
\alias{InitErgmTerm.pdegcor}
\alias{InitErgmTerm.rdegcor}
\alias{InitErgmTerm.threepath}

\alias{InitErgmTerm.b1concurrent}
\alias{InitErgmTerm.b1degree.edgecov}
\alias{InitErgmTerm.b1factor}
\alias{InitErgmTerm.b1mindegree}
\alias{InitErgmTerm.b1mindegree.edgecov}
\alias{InitErgmTerm.b2concurrent}
\alias{InitErgmTerm.b2degree.edgecov}
\alias{InitErgmTerm.b2factor}
\alias{InitErgmTerm.b2mindegree}
\alias{InitErgmTerm.b2mindegree.edgecov}
\alias{InitErgmTerm.balance}
\alias{InitErgmTerm.coincidence}
\alias{InitErgmTerm.concurrent}
\alias{InitErgmTerm.concurrentties}
\alias{InitErgmTerm.concurrentties2}
\alias{InitErgmTerm.conddegmple}
\alias{InitErgmTerm.ctriad}
\alias{InitErgmTerm.ctriple}
\alias{InitErgmTerm.cyclicalties}
\alias{InitErgmTerm.cyclicalties2}
\alias{InitErgmTerm.degree}
\alias{InitErgmTerm.density}
\alias{InitErgmTerm.dsp}
\alias{InitErgmTerm.dyadcov}
\alias{InitErgmTerm.edges}
\alias{InitErgmTerm.esp}
\alias{InitErgmTerm.gwb1degree}
\alias{InitErgmTerm.gwb2degree}
\alias{InitErgmTerm.gwdegree}
\alias{InitErgmTerm.gwdsp}
\alias{InitErgmTerm.gwesp}
\alias{InitErgmTerm.gwidegree}
\alias{InitErgmTerm.gwnsp}
\alias{InitErgmTerm.gwodegree}
\alias{InitErgmTerm.hammingmix}
\alias{InitErgmTerm.homoproportion}
\alias{InitErgmTerm.idegree}
\alias{InitErgmTerm.indegreepopularity}
\alias{InitErgmTerm.intransitive}
\alias{InitErgmTerm.istar}
\alias{InitErgmTerm.kstar}
\alias{InitErgmTerm.localtriangle}
\alias{InitErgmTerm.m2star}
\alias{InitErgmTerm.meandeg}
\alias{InitErgmTerm.nearsimmelian}
\alias{InitErgmTerm.nodecov}
\alias{InitErgmTerm.nodeicov}
\alias{InitErgmTerm.nodemain}
\alias{InitErgmTerm.nodeocov}
\alias{InitErgmTerm.nsp}
\alias{InitErgmTerm.odegree}
\alias{InitErgmTerm.opentriad}
\alias{InitErgmTerm.ostar}
\alias{InitErgmTerm.outdegreepopularity}
\alias{InitErgmTerm.receiver}
\alias{InitErgmTerm.sender}
\alias{InitErgmTerm.simmelian}
\alias{InitErgmTerm.simmelianties}
\alias{InitErgmTerm.smalldiff}
\alias{InitErgmTerm.sociality}
\alias{InitErgmTerm.transitive}
\alias{InitErgmTerm.transitiveties}
\alias{InitErgmTerm.transitiveties2}
\alias{InitErgmTerm.triadcensus}
\alias{InitErgmTerm.triangle}
\alias{InitErgmTerm.triangles}
\alias{InitErgmTerm.tripercent}
\alias{InitErgmTerm.ttriad}
\alias{InitErgmTerm.ttriple}
\alias{InitErgmTerm.twopath}

\alias{InitWtErgmTerm.CMP}
\alias{InitWtErgmTerm.absdiff} 
\alias{InitWtErgmTerm.absdiffcat}
\alias{InitWtErgmTerm.atleast}
\alias{InitWtErgmTerm.cyclicalties}
\alias{InitWtErgmTerm.cyclicalweights}
\alias{InitWtErgmTerm.deference}
\alias{InitWtErgmTerm.edgecov}
\alias{InitWtErgmTerm.greaterthan}
\alias{InitWtErgmTerm.inconsistency}
\alias{InitWtErgmTerm.ininterval}
\alias{InitWtErgmTerm.mutual}
\alias{InitWtErgmTerm.nodecov}
\alias{InitWtErgmTerm.nodecovar}
\alias{InitWtErgmTerm.nodefactor}
\alias{InitWtErgmTerm.nodeicov}
\alias{InitWtErgmTerm.nodeicovar}
\alias{InitWtErgmTerm.nodeifactor}
\alias{InitWtErgmTerm.nodeisqrtcovar}
\alias{InitWtErgmTerm.nodemain}
\alias{InitWtErgmTerm.nodeocov}
\alias{InitWtErgmTerm.nodeocovar}
\alias{InitWtErgmTerm.nodeofactor}
\alias{InitWtErgmTerm.nodeosqrtcovar}
\alias{InitWtErgmTerm.nodesqrtcovar}
\alias{InitWtErgmTerm.nonconformity}
\alias{InitWtErgmTerm.nonzero}
\alias{InitWtErgmTerm.sum}
\alias{InitWtErgmTerm.transitiveties}
\alias{InitWtErgmTerm.transitiveweights}

\alias{check.ErgmTerm.summarystats}
\alias{ergm.getglobalstats}
\alias{get.InitErgm.fname}
\alias{updatemodel.ErgmTerm}
\alias{copy.named}
\alias{ergm.update.formula}
\alias{term.list.formula}
\alias{model.matrix.pen.glm}
\alias{zerowarnings}
\alias{extremewarnings}

\alias{InitMHP.CondDegree}
\alias{InitMHP.CondDegreeDist}
\alias{InitMHP.CondInDegreeDist}
\alias{InitMHP.CondOutDegreeDist}
\alias{InitMHP.ConstantEdges}
\alias{InitMHP.dissolution}
\alias{InitMHP.formation}
\alias{InitMHP.formationTNT}
\alias{InitMHP.HammingConstantEdges}
\alias{InitMHP.HammingTNT}
\alias{InitMHP.nobetweengroupties}
\alias{InitMHP.randomtoggle}
\alias{InitMHP.randomtoggleNonObserved}
\alias{InitMHP.TNT}
\alias{InitMHP.TNT10}
\alias{InitMHP.CondDegreeHexad}
\alias{InitMHP.CondDegreeTetrad}
\alias{InitMHP.CondDegreeSimple}
\alias{InitMHP.dissolutionMLE}
\alias{InitMHP.dissolutionNonObservedMLE}
\alias{InitMHP.formationMLE}
\alias{InitMHP.formationMLETNT}
\alias{InitMHP.formationNonObservedMLE}

\alias{InitWtMHP.DescRank}
\alias{InitWtMHP.DescRankEquivalent}
\alias{InitWtMHP.Poisson}
\alias{InitWtMHP.PoissonNonObserved}
\alias{InitWtMHP.StdNormal}
\alias{InitWtMHP.StdNormalRank}
\alias{InitWtMHP.ZIPoisson}

\alias{MHproposal}
\alias{MHproposal.character}
\alias{MHproposal.ergm}
\alias{MHproposal.formula}
\alias{MHproposal.MHproposal}
\alias{MHproposal.NULL}
\alias{MHproposals}
%



\alias{append.rhs.formula}
\alias{autoboot.ergm}
\alias{autoboot.se.ergm}
\alias{central.network}
\alias{coef.length.model}
\alias{coef.sublength.model}
\alias{compress.data.frame}
\alias{control.ergm.bridge}
\alias{control.ergm.toplevel}
\alias{control.godfather}
\alias{control.logLik.ergm}
\alias{control.logLik.stergm}
\alias{control.simulate.ergm.toplevel}
\alias{degree.mix.matrix}
\alias{delete.isolates}
\alias{duration.matrix}
\alias{ergm.Cprepare.el}
\alias{ergm.Cprepare.miss}
\alias{ergm.MCMCse.old}
\alias{ergm.MCMLE}
\alias{ergm.MPIcluster.started}
\alias{ergm.bridge.0.llk}
\alias{ergm.bridge.hammingstart.llk}
\alias{ergm.bridge.preproc}
\alias{ergm.checkconstraints.model}
\alias{ergm.checkextreme.model}
\alias{ergm.coefficient.degeneracy}
\alias{ergm.getCluster}
\alias{ergm.godfather}
\alias{ergm.mahalanobis}
\alias{ergm.mapl}
\alias{ergm.maple}
\alias{ergm.plinfo}
\alias{ergm.reviseinit}
\alias{ergm.sample.tomcmc}
\alias{ergm.stopCluster}
\alias{ergm.stopCluster.MPIcluster}
\alias{ergm.stopCluster.default}
\alias{init.methods}
\alias{largest.components}
\alias{llik.fun.obs}
\alias{llik.fun.obs.robust}
\alias{llik.grad.obs}
\alias{llik.hessian.obs}
\alias{logLik.stergm.CMLE}
\alias{logp.rank.ergm}
\alias{model.transform.formula}
\alias{mvmodel}
\alias{mvmodel.default}
\alias{mvmodel.ergm}
\alias{mvmodel.formula}
\alias{myLibLoc}
\alias{nologLik.message}
\alias{overlap.matrix}
\alias{prevalence}
\alias{print.summary.stergm}
\alias{simulate.formula.stergm}
\alias{simulate.network}
\alias{sociality.default}
\alias{sociality.ergm}
\alias{sociality.formula}
\alias{sociality.network}
\alias{sort.data.frame}
\alias{stergm.CMLE}
\alias{stergm.EGMoME}
\alias{stergm.EGMoME.initialfit}
\alias{stergm.RM}
\alias{stergm.SPSA}
\alias{stergm.SPSA.C}
\alias{stergm.getMCMCsample}
\alias{stergm.phase12.C}
\alias{vector.namesmatch}
\alias{colMeans.mcmc.list}
\alias{control.simulate.formula.ergm}
\alias{control.simulate.network}
\alias{ergm.sample.eta2theta}
\alias{plot.mcmc.list.ergm}
\alias{set.offset.formula}
\alias{stergm.EGMoME.RM.Phase2.C}
\alias{stergm.RM2}
\alias{sweep.mcmc.list}





\alias{degreedistfactor}
\alias{drawcircle}
\alias{drawpie}
\alias{dspartnerdist}
\alias{ergm.compute.degeneracy}
\alias{ergm.curved}
\alias{ergm.curved.statsmatrix}
\alias{ergm.curved.update}
\alias{ergm.design}
\alias{ergm.estimate}
\alias{ergm.estimate.only}
\alias{ergm.eta}
\alias{ergm.etagrad}
\alias{ergm.etagradmult}
\alias{ergm.etamap}
\alias{ergm.geodesicmatrix}
\alias{ergm.geodesicmatrix.edgelist}
\alias{ergm.geodesicmatrix.old}
\alias{ergm.geodist}
\alias{ergm.geodistdist}
\alias{ergm.geodistn}
\alias{ergm.getMCMCsample}
\alias{ergm.mcmcslave}
\alias{ergm.getnetwork}
\alias{ergm.independencemodel}
\alias{ergm.initialfit}
\alias{ergm.logisticdeviance}
\alias{ergm.logitreg}
\alias{ergm.mainfitloop}
\alias{ergm.marquardt}
\alias{ergm.marquardt2}
\alias{ergm.pl}
\alias{ergm.nodegeodesics}
\alias{ergm.pairgeodesic}
\alias{ergm.phase12}
\alias{ergm.raftery.diag}
\alias{ergm.revisetheta0}
\alias{ergm.rhs.formula} 
\alias{ergm.stepping}
\alias{ergm.stocapprox}
\alias{ergm.t.summary}
\alias{espartnerdist}
\alias{get.node.attr}
\alias{is.ergm}
\alias{is.matrixnetwork}
\alias{llik.fun}
\alias{llik.fun.miss}
\alias{llik.fun2}
\alias{llik.fun3}
\alias{llik.grad}
\alias{llik.grad.miss}
\alias{llik.grad2}
\alias{llik.grad3}
\alias{llik.hessian}
\alias{llik.hessian.miss} 
\alias{llik.hessian2}
\alias{llik.info3}
\alias{llik.mcmcvar3}
\alias{llik.fun.EF}
\alias{llik.fun.median}
\alias{llik.fun.miss.robust}
\alias{llik.hessian.naive}
\alias{mcmc.diagnostics.default}
\alias{midarrow}
\alias{newnw.extract}
\alias{plot.mcmc.ergm}
\alias{print.raftery.diag.ergm}
\alias{robust.inverse}
\alias{rspartnerdist}
\alias{set.mfrow}
\alias{traceplot.ergm}
\alias{twopathdist}
%
\alias{ergm.sufftoprob}
\alias{ergm.Cprepare}
\alias{ergm.getmodel}
\alias{ergm.getterms}
\alias{ergm.gettermnames}
\alias{ergm.mple}
\alias{ergm.statseval}
\alias{fulldistdist}
\alias{fullgcount}
\alias{statnet.edit}
%\alias{get.node.attr}
%\alias{geodist}
%\alias{has.loops}
%\alias{is.directed}
%\alias{is.hyper}
%\alias{lpz.dist}
%\alias{mahal}
%\alias{mlpY}
%\alias{mlpY.grad}
%\alias{mlpY0}
%\alias{mlpY0.grad}
%\alias{mlpYmdsZ}
%\alias{mlpYmdsZ.grad}
%\alias{mlpYnull}
%\alias{prepmahal}
%\alias{symmetrize}
%
\alias{ergm.MCMCse}
\alias{ergm.MCMCse.lognormal}
%\alias{ergm.latentplot}
%\alias{mlpY.grad.plot}
%\alias{mlpY.plot}
%\alias{mlpYmdsZ.grad.plot}
%\alias{mlpYmdsZ.plot}
\alias{ergm.MCMCacf}
\alias{ergm.pen.glm}
\alias{logistftest}
%
\alias{ergm.mainloop}
\alias{ergm.robmon}
%

\alias{as.edgelist.compressed}
\alias{as.network.uncompressed}

%
% These need to be documented!
%

\alias{as.edgelist}
\alias{ergm.degeneracy} % At least this needs to be documented in summary.ergm and/or mcmc.diagnostics.ergm
\alias{print.network.series}
\alias{summary.statsmatrix.ergm}
\alias{summary.network.series}
\alias{license.statnet}
\alias{license.statnet1}
\alias{statnetShowDoc}
\alias{statnetbrowseURL}
\alias{ergm.bounddeg}
\alias{is.invertible}
\alias{ostar2deg}
\alias{sufftoprob}
\alias{degreedist}
\alias{mixingmatrix}
\alias{print.mixingmatrix}
\alias{print.summary.ergm} 
\alias{Summary.ergm.future}

\description{
  Internal ergm functions.}

%\usage{
%anova.ergm(object, \dots)
%}
\details{
  Most of these are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).}

%\arguments{
% \item{\dots}{further arguments passed to or used by methods.}
%}
\seealso{ergm, statnet-package}
\keyword{internal}
