% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHproposal.R
\name{ergm.ConstraintImplications}
\alias{ergm.ConstraintImplications}
\title{Set up the implied constraints from the current constraint}
\usage{
ergm.ConstraintImplications(implier, implies)
}
\arguments{
\item{implier}{The current constraint specified in the model.  For the list
of constraints, see \code{\link{ergm-constraints}}}

\item{implies}{Implied constraints from the current constraint (based on the
user's knowledge).}
}
\description{
This is a low-level function not intended to be called directly by
end users. For information on constraints, see the
\code{\link{ergm-constraints}} page.  Calling this function with
arguments adds the specified constraint implication to the
list. Calling it without arguments returns the implication list so
far.
}
