% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.design.R
\name{ergm.design}
\alias{ergm.design}
\title{Obtain the set of informative dyads based on the network structure. Note that \verb{model=} argument is not needed and will be removed in a future release.}
\usage{
ergm.design(nw, verbose = FALSE)
}
\value{
\code{ergm.design} returns a \code{\link{rlebdm}} of
informative (non-missing, non fixed) dyads.
}
\description{
Obtain the set of informative dyads based on the network structure. Note that \verb{model=} argument is not needed and will be removed in a future release.
}
