% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.getglobalstats.R
\name{ergm.emptynwstats.model}
\alias{ergm.emptynwstats.model}
\title{Empty network statistics for an ergm model}
\usage{
ergm.emptynwstats.model(m)
}
\arguments{
\item{m}{an \code{ergm.model}.}
}
\value{
a numeric vector equal in length to the number of
statistics of the model containing the statistics evaluated on an empt network.
}
\description{
Given an ergm model, return a vector of statistics of that model evaluated on an empty network.
}
