% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{ergm.eta}
\alias{ergm.eta}
\alias{ergm.etagrad}
\alias{ergm.etagradmult}
\alias{ergm.etagradmultt}
\title{Operations to map curved \code{\link[=ergm]{ergm()}} parameters onto canonical parameters}
\usage{
ergm.eta(theta, etamap)

ergm.etagrad(theta, etamap)

ergm.etagradmult(theta, v, etamap)

ergm.etagradmultt(theta, s, etamap)
}
\arguments{
\item{theta}{a curved model parameter \eqn{q}-vector}

\item{etamap}{the list of values that describes the \eqn{\theta\to\eta}{theta -> eta}
mapping, usually attached as \verb{$etamap} element of an \code{\link{ergm_model}}
object. At this time, it is a list with the following elements:
\describe{
\item{\code{canonical}}{ a numeric vector whose \code{i}th entry specifies whether the \code{i}th component of theta is canonical (via non-negative integers) or curved (via zeroes)}
\item{\code{offsetmap}}{ a logical vector whose \code{i}th entry tells whether the ith coefficient of the canonical parameterization was "offset", i.e fixed}
\item{\code{offset}}{ a logical vector whose ith entry tells whether the ith model term was offset/fixed}
\item{\code{offsettheta}}{ a logical vector whose ith entry tells whether the ith curved theta coeffient was offset/fixed;}
\item{\code{curved}}{ a list with one component per curved EF term in
the model containing \describe{
\item{\code{from}}{ the indices of the curved theta parameter that are to be mapped from}
\item{\code{to}}{ the indices of the canonical eta parameters to be mapped to}
\item{\code{map}}{ the map provided by \code{\link{InitErgmTerm}}}
\item{\code{gradient}}{ the gradient function provided by \code{\link{InitErgmTerm}}}
\item{\code{cov}}{ optional additional covariates to be passed to the map and the gradient functions }
\item{\code{etalength}}{ the length of the eta vector}
}
}
}}

\item{v}{a \eqn{p}-vector the vector or a \eqn{p\times r}{p*r}
matrix}

\item{s}{a \eqn{r\times p} matrix}
}
\value{
For \code{ergm.eta}, the canonical \eqn{\eta} parameters as mapped
from \eqn{\theta}.
}
\description{
This family of functions concerns mapping model parameter
\eqn{q}-vector \eqn{\theta} using to the canonical parameter
\eqn{p}-vector \eqn{\eta} based on the \code{etamap} object, usually
attached as the \verb{$etamap} element of an \code{\link{ergm_model}} object.
}
\details{
These functions are mainly important in the case of curved exponential family
models, i.e., those in which the parameter of interest (\eqn{\theta}) is not a
linear function of the natural parameters (\eqn{\eta}) in the exponential-family
model. In non-curved models, we may assume without loss of generality that
\eqn{\eta(\theta)=\theta}.

A succinct description of how eta(theta) is incorporated into an ERGM is
given by equation (5) of \insertCite{Hu07c;textual}{ergm}.  See \insertCite{HuHa06i;textual}{ergm} and
\insertCite{Hu07c;textual}{ergm} for further details about how eta and its derivatives are used
in the estimation process.
}
\section{Functions}{
\itemize{
\item \code{ergm.eta()}: Evaluate \eqn{\eta(\theta):
\mathbb{R}^q\to\mathbb{R}^p}{eta(theta): R^q -> R^p}.

\item \code{ergm.etagrad()}: Evaluate \eqn{\eta'(\theta):
  \mathbb{R}^q\to\mathbb{R}^{q\times p}}{deta(theta): R^q ->
  R^{q*p}}. If the gradient is only intended to be a multiplier for
some vector or matrix, the more efficient \code{ergm.etagradmult}
is recommended.

\item \code{ergm.etagradmult()}: Evaluate \eqn{\eta'(\theta) V: \mathbb{R}^q \times \mathbb{R}^{p\times r}\to\mathbb{R}^{q\times
r}}{deta(theta)\%*\%V: R^q * R^{p*r} -> R^{q*r}}.

\item \code{ergm.etagradmultt()}: A convenience function to evaluate
\eqn{\{\eta'(\theta) S^{\top}\}^{\top} = S
  \eta'(\theta)^{\top}}{t(deta(theta)\%*\%t(S)) =
  S\%*\%t(deta(theta))}.

}}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ergmTerm}}
}
\keyword{internal}
