% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.getglobalstats.R
\name{ergm.getglobalstats}
\alias{ergm.getglobalstats}
\title{Internal function to return global statistics for a given network; (DEPRECATED: use \code{\link[=summary.ergm_model]{summary.ergm_model()}}.)}
\usage{
ergm.getglobalstats(nw, m, response = NULL)
}
\arguments{
\item{nw}{a \code{\link{network}} object}

\item{m}{the model in use with network \code{nw}, as returned by
\code{\link{ergm_model}}}

\item{response}{character name of an edge attribute to be used (for weighted
ergm models)}
}
\value{
returns a vector of the global statistics
}
\description{
The \code{ergm.getglobalstats} function is a low-level function not normally
called by the user. It calculates and returns the global statistics for a
given network and model.
}
\details{
Calculates and returns the global statistics for a given network via
\code{\link{ergm.Cprepare}} and \code{network_stats_wrapper.C} or
\code{wt_network_stats_wrapper.C} if the model is weighted. It is called by
\code{\link{summary_formula.network}} which is generally the better way
to access the functionality.
}
\seealso{
\code{\link{summary.formula}}
}
