% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.getnetwork.R
\name{ergm.getnetwork}
\alias{ergm.getnetwork}
\title{Acquire and verify the network from the LHS of an \code{ergm} formula
and verify that it is a valid network.}
\usage{
ergm.getnetwork(formula, loopswarning = getOption("ergm.warn_loops"))
}
\arguments{
\item{formula}{a two-sided formula whose LHS is a \code{\link[network:network]{network}}, an object
that can be coerced to a \code{\link[network:network]{network}}, or an expression that evaluates
to one.}

\item{loopswarning}{whether warnings about loops should be printed
(\code{TRUE} or \code{FALSE}); defaults to value of option \code{ergm.warn_loops}
or \code{TRUE}.}
}
\value{
A \code{\link[network:network]{network}} object constructed by evaluating the LHS of
the model formula in the formula's environment.
}
\description{
The function function ensures that the network in a given formula
is valid; if so, the network is returned; if not, execution is
halted with warnings.
}
