% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHproposal.R
\name{ergm.MHP.table}
\alias{ergm.MHP.table}
\title{Table mapping reference,constraints, etc. to Metropolis Hastings Proposals
(MHP)}
\usage{
ergm.MHP.table(Class, Reference, Constraints, Priority, Weights, MHP)
}
\arguments{
\item{Class}{default to "c"}

\item{Reference}{The reference measure used in the model. For the list of
reference measures, see \code{\link{ergm-references}}}

\item{Constraints}{The constraints used in the model. For the list of
constraints, see \code{\link{ergm-constraints}}}

\item{Priority}{On existence of multiple qualifying MHPs, specifies the
priority (\code{-1},\code{0},\code{1}, etc.) of MHPs to be used.}

\item{Weights}{The sampling weights on selecting toggles (random, TNT, etc).}

\item{MHP}{The matching MHP from the previous arguments.}
}
\description{
This is a low-level function not intended to be called directly by
end users. For information on Metropolis-Hastings proposal methods,
\link{ergm_MH_proposals}. This function sets up the table mapping
constraints, references, etc. to MHproposals. Calling it with
arguments adds a new row to this table. Calling it without
arguments returns the table so far.
}
\note{
The arguments \code{Class}, \code{Reference}, and \code{Constraints} can
have length greater than 1. If this is the case, the rows added
to the table are a \emph{Cartesian product} of their elements.
}
