% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_proposal.R
\name{ergm_conlist}
\alias{ergm_conlist}
\alias{ergm_conlist.ergm_conlist}
\alias{ergm_conlist.NULL}
\alias{ergm_conlist.formula}
\alias{ergm_conlist.term_list}
\alias{c.ergm_conlist}
\alias{[.ergm_conlist}
\title{Internal representation of ERGM sample space constraints and hints}
\usage{
ergm_conlist(object, ...)

\method{ergm_conlist}{ergm_conlist}(object, ...)

\method{ergm_conlist}{`NULL`}(object, ...)

\method{ergm_conlist}{formula}(object, nw, ...)

\method{ergm_conlist}{term_list}(object, nw, ..., term.options = list())

\method{c}{ergm_conlist}(...)

\method{[}{ergm_conlist}(x, ...)
}
\arguments{
\item{object}{a data type}

\item{nw}{a \code{\link[network:network]{network}} object.}

\item{term.options}{A list of additional arguments to be passed to term initializers. See \code{\link[=term.options]{? term.options}}.}
}
\description{
See \code{\link{ergmConstraint}} for details on specifying proposals and
constraints.
}
\section{Methods (by class)}{
\itemize{
\item \code{ergm_conlist(ergm_conlist)}: identity method.

\item \code{ergm_conlist(`NULL`)}: handling for \code{NULL}, to simplify other code.

\item \code{ergm_conlist(formula)}: initialize from \code{\link{formula}}, such as in
\code{\link[=ergm]{ergm()}} \code{constraints} and \code{obs.constraints} arguments: preserve
the LHS, convert to \code{\link[statnet.common:term_list]{term_list}}, and use the nxt method.

\item \code{ergm_conlist(term_list)}: initialize from \code{\link[statnet.common:term_list]{term_list}}.

}}
\section{Methods (by generic)}{
\itemize{
\item \code{c(ergm_conlist)}: a concatenation method.

\item \code{[}: a subsetting method.

}}
\keyword{internal}
