% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.operator.R
\name{ergm_constrain_changestats}
\alias{ergm_constrain_changestats}
\title{Helper function for proposals implementing the \ergmConstraint{ergm}{ChangeStats}{()} constraint}
\usage{
ergm_constrain_changestats(
  arguments,
  aux_before = EMPTY_TERM_LIST,
  aux_after = EMPTY_TERM_LIST
)
}
\arguments{
\item{arguments}{\code{arguments} argument of a call to
\verb{Init*ErgmProposal.*()} function.}

\item{aux_before, aux_after}{either \code{\link{formula}}s or a \code{\link[statnet.common:term_list]{term_list}}s
for other auxiliaries of the proposal.}
}
\value{
a list containing two elements: \code{auxiliaries} with the
concatenated \code{\link[statnet.common:term_list]{term_list}} and \code{ChangeStat_pos}, giving the
position of the \code{.submodel(terms)} auxiliary or \code{NULL} if none
was specified.
}
\description{
This function inserts a \code{.submodel(terms)} auxiliary among the
proposal's other auxiliaries and records its position, where
\code{terms} are the statistics passed to the \ergmConstraint{ergm}{ChangeStats}{()}
constraint.
}
\keyword{internal}
