% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{ergm_cutoff_message}
\alias{ergm_cutoff_message}
\title{Helper function for constructing \verb{gw*} cutoff error messages}
\usage{
ergm_cutoff_message(cutoff, term, stat, arg = NULL, opt = NULL)
}
\arguments{
\item{cutoff}{the maximum value for the statistic of interest.}

\item{term}{the name of the term.}

\item{stat}{the name of the statistic of interest.}

\item{arg}{the name of the term argument (if any) that controls the cutoff.}

\item{opt}{the name of the term option (if any) that controls the cutoff.}
}
\value{
A character string with the error message.
}
\description{
Helper function for constructing \verb{gw*} cutoff error messages
}
\keyword{internal}
