\name{ergm_deprecated}
\alias{delete.isolates}
\alias{central.network}
\alias{largest.components}
\alias{degreedistfactor}
\alias{sociality.default}
\alias{sociality.ergm}
\alias{sociality.formula}
\alias{sociality.network}
\alias{rspartnerdist}
\alias{espartnerdist}
\alias{dspartnerdist}
\alias{twopathdist}
\alias{ostar2deg}
\alias{drawpie}
\alias{invert.network}
\alias{is.invertible}
\alias{mvmodel}
\alias{mvmodel.default}
\alias{mvmodel.ergm}
\alias{mvmodel.formula}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Functions that will no longer be supported in future releases of the package
}
\description{
Functions that have been superceed, were never documented, or were removed from the package for ther reasons
}
