% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal.R
\name{ergm_flatten_conterm_list}
\alias{ergm_flatten_conterm_list}
\title{Convert a list of constraint formulas or terms to a flat term list}
\usage{
ergm_flatten_conterm_list(l)
}
\arguments{
\item{l}{a \code{\link{formula}}, a \code{\link[statnet.common:term_list]{term_list}}, a \code{\link{character}}, \code{\link{NULL}},
or a (potentially heterogeneous) list thereof.}
}
\value{
a \code{\link[statnet.common:term_list]{term_list}}.
}
\description{
This helper function processes the usual constraint specifications
(\code{\link{formula}}s, \code{\link[statnet.common:term_list]{term_list}}s, \code{\link{character}} strings) or a list
thereof into a flat \code{\link[statnet.common:term_list]{term_list}} of distinct values.
}
\examples{
z <- 5
ergm_flatten_conterm_list(list(
  NULL,
  ~a(x) - b,
  statnet.common::base_env(~c),
  NULL,
  statnet.common::term_list(call("f", z), env = baseenv(), sign = -1)
))

}
