% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_keyword.R
\name{ergm_keyword}
\alias{ergm_keyword}
\title{Dynamic ERGM keyword registry}
\usage{
ergm_keyword(
  name = NULL,
  short = NULL,
  description = NULL,
  popular = NULL,
  package = NULL
)
}
\arguments{
\item{name}{full name of the keyword}

\item{short}{abbreviation of the keyword name}

\item{description}{description of the keyword}

\item{popular}{logical to indicate if a keyword is popular}

\item{package}{package the keyword is first defined in}
}
\value{
Returns a dataframe with the following columns:
\itemize{
\item name
\item short
\item description
\item popular
\item package
}
}
\description{
A function to manage dynamic ERGM keywords. To register a keyword, call the function with all parameters
provided. To fetch all registered keywords, call the function with no parameters specified.
}
\keyword{internal}
\keyword{models}
