% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_model.R
\name{ergm_model}
\alias{ergm_model}
\title{Internal representation of an \code{ergm} network model}
\usage{
ergm_model(object, nw, response = NULL, silent = FALSE, role = "static",
  ...)
}
\arguments{
\item{object}{See specific method documentation.}

\item{nw}{the network of interest}

\item{response}{Name of the edge attribute whose value is to be
modeled in the valued ERGM framework. Defaults to \code{NULL} for
simple presence or absence, modeled via a binary ERGM.}

\item{silent}{logical, whether to print the warning messages from the
initialization of each model term.}

\item{role}{A hint about how the model will be used. Used primarily for
dynamic network models.}

\item{\dots}{additional parameters for model formulation}
}
\value{
\code{ergm_model} returns an  \code{ergm_model} object as a list
containing:
\item{ formula}{the formula inputted to
\code{\link{ergm_model}}}
\item{coef.names}{a vector of coefficient names}
\item{offset}{a logical vector of whether each term was "offset", i.e.
fixed}
\item{terms}{a list of terms and 'term components' initialized by the
appropriate \code{InitErgmTerm.X} function.}
\item{network.stats0}{NULL always??}
\item{etamap}{the theta -> eta mapping as a list returned from
<ergm.etamap>}
}
\description{
These methods are generally not called directly by users, but may
be employed by other depending packages.
A model object is constructed from an \code{\link[=ergm]{ergm()}}
formula of the form \code{network ~ model.term(s)} or \code{~
model.term(s)}, with the network passed separately. Each term is
initialized via the \code{InitErgmTerm} functions to create a
\code{ergm_model} object.
}
\note{
This API is not to be considered fixed and may change between versions. However, an effort will be made to ensure that the methods of this class remain stable.
}
