% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_mplecov.R
\name{ergm_mplecov}
\alias{ergm_mplecov}
\title{Approximate MPLE standard errors in dyad-dependent models}
\usage{
ergm_mplecov(pl, nw, fd, m, theta.mple, invHess, control, verbose)
}
\arguments{
\item{pl}{An \code{\link{ergm.pl}} object.
@param nw response network.}

\item{fd}{An \code{\link{rlebdm}} with informative dyads.}

\item{m}{the model, as returned by \code{\link{ergm_model}}}

\item{theta.mple}{the MPLE of a given model}

\item{invHess}{the inverse Hessian matrix obtained from glm()}

\item{control}{a list of MCMC related parameters; recognized
components include: samplesize : the number of networks to sample
Clist.miss : see 'Clist.miss' above; some of the code uses this
Clist.miss,}

\item{verbose}{whether this and the C routines should be verbose (T
or F); default=FALSE}
}
\value{
\code{ergm_mplecov} returns a list either
containing a Godambe covariance matrix or a diagonal matrix with bootstrap variances.
}
\description{
Function to approximate the MPLE covariance matrix in a dyad dependence model
using the Godambe matrix as described in Schmid and Hunter (2020) or by parametric bootstrap
as described by Schmid and Desmarais (2017).
}
\references{
Schmid CS and Desmarais BA (2017). "Exponential random graph
models with big networks: Maximum pseudolikelihood estimation and the parametric bootstrap"
\emph{IEEE International Conference on Big Data (Big Data)}, pp. 116-121.

Schmid CS and Hunter DR (2020).  "Accounting for Model Misspecification When Using
Pseudolikelihood for ERGMs" \emph{Working Paper}.
}
