% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.diagnostics.ergm.R
\name{ergm_plot.mcmc.list}
\alias{ergm_plot.mcmc.list}
\title{Plot MCMC list using \code{lattice} package graphics}
\usage{
ergm_plot.mcmc.list(x, main = NULL, vars.per.page = 3, ...)
}
\arguments{
\item{x}{an \code{\link{mcmc.list}} object containing the mcmc diagnostic
samples.}

\item{main}{character, main plot heading title.}

\item{vars.per.page}{Number of rows (one variable per row) per
plotting page.  Ignored if \code{latticeExtra} package is not
installed.}

\item{...}{additional arguments, currently unused.}
}
\description{
Plot MCMC list using \code{lattice} package graphics
}
\note{
This is not a method at this time.
}
