% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_proposal.R
\name{ergm_proposal_table}
\alias{ergm_proposal_table}
\title{Table mapping reference,constraints, etc. to ERGM Metropolis-Hastings proposals}
\usage{
ergm_proposal_table(
  Class,
  Reference,
  Constraints,
  Priority,
  Weights,
  Proposal,
  Package = NULL
)
}
\arguments{
\item{Class}{default to "c"}

\item{Reference}{The reference measure used in the model. For the
list of reference measures, see \code{\link{ergmReference}}. Use \code{"*"} to
indicate that the proposal can be used with \emph{any} reference
distribution, for example if it's an proposal "operator".}

\item{Constraints}{The constraints used in the model. For the list
of constraints, see \code{\link{ergmConstraint}}. They are specified as a
single string of text, with each contrast prefixed by either \code{&}
for constraints that the proposal \emph{always} enforces or \code{|} for
constraints that the proposal \emph{can} enforce if
needed. Placeholder \code{*} can be used with proposal "operators" to
indicate that they will initialise sub-proposals that may be able
to accommodate these constraints and hints.}

\item{Priority}{On existence of multiple qualifying proposals, specifies the
priority (\code{-1},\code{0},\code{1}, etc.) of proposals to be used.}

\item{Weights}{The sampling weights on selecting toggles (random, TNT, etc).}

\item{Proposal}{The matching proposal from the previous arguments.}

\item{Package}{The package in which the proposal is implemented;
it's normally autodetected based on the package to which the
calling function belongs.}
}
\description{
This is a low-level function not intended to be called directly by
end users. For information on Metropolis-Hastings proposal methods,
\link{ergm-proposals}. This function sets up the table mapping
constraints, references, etc. to \code{ergm_proposals}. Calling it with
arguments adds a new row to this table. Calling it without
arguments returns the table so far.
}
\details{
The first time a particular package calls
\code{ergm_proposal_table()}, it also sets a call-back to remove all
of its proposals from the table should the package be unloaded.
}
\note{
The arguments \code{Class}, \code{Reference}, and \code{Constraints} can
have length greater than 1. If this is the case, the rows added
to the table are a \emph{Cartesian product} of their elements.
}
\keyword{internal}
