% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_state.R
\name{ergm_state_receive}
\alias{ergm_state_receive}
\alias{ergm_state_receive.ergm_state}
\alias{ergm_state_receive.ergm_state_full}
\title{A nonce class representing an \code{\link{ergm_state}} without \code{nw0}, \code{model}, or \code{proposal}.}
\usage{
ergm_state_receive(x, ...)

\method{ergm_state_receive}{ergm_state}(x, ...)

\method{ergm_state_receive}{ergm_state_full}(x, ...)
}
\arguments{
\item{x}{Typically an \code{\link{ergm_state}}.}
}
\description{
A nonce class representing an \code{\link{ergm_state}} without \code{nw0}, \code{model}, or \code{proposal}.
}
