% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_state.R
\name{ergm_state_send}
\alias{ergm_state_send}
\alias{ergm_state_send.ergm_state_send}
\alias{ergm_state_send.ergm_state_full}
\alias{update.ergm_state_send}
\title{A nonce class representing an \code{\link{ergm_state}} without \code{nw0}.}
\usage{
ergm_state_send(x, ...)

\method{ergm_state_send}{ergm_state_send}(x, ...)

\method{ergm_state_send}{ergm_state_full}(x, ...)

\method{update}{ergm_state_send}(object, state, ...)
}
\description{
A nonce class representing an \code{\link{ergm_state}} without \code{nw0}.
}
