% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-package.R
\docType{package}
\name{ergmProposals}
\alias{ergmProposals}
\alias{ergm-proposals}
\alias{proposals-ergm}
\alias{ergm.proposals}
\alias{proposals.ergm}
\title{Proposals for Exponential-Family Random Graph Models}
\description{
TODO
}
\section{Implemented proposals for ergm models}{

\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.parseProposal(), keepProposal=TRUE)}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.parseProposal(), keepProposal=TRUE)}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.parseProposal(), keepProposal=TRUE)}}
}

\references{
TODO
}
\seealso{
TODO
}
\keyword{models}
