% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-terms-index.R
\docType{package}
\name{ergmReference}
\alias{ergmReference}
\title{An index of Ergm reference}
\description{
An index of Ergm references
}
\section{Reference index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmReference"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmReference"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmReference"))}}
}

