% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R,
%   R/InitWtErgmTerm.operator.R
\name{Exp-ergmTerm}
\alias{Exp-ergmTerm}
\alias{InitErgmTerm.Exp}
\alias{InitWtErgmTerm.Exp}
\title{Exponentiate a network's statistic}
\usage{
# binary: Exp(formula)

# valued: Exp(formula)
}
\arguments{
\item{formula}{formula to be evaluated}
}
\description{
Exponentiate a network's statistic
}
\details{
Evaluate the terms specified in \code{formula} and exponentiates them with base \eqn{e} .
}
\concept{operator}
