%  File man/faux.magnolia.high.Rd in package ergm, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2013 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{faux.magnolia.high.withnumericattributes}
\docType{data}
\title{Goodreau's Faux magnolia High School as a network object}
\alias{faux.magnolia.high.withnumericattributes}
\description{
This is the \code{faux.magnolia.high} dataset, with the attributes \code{Race} and \code{Sex}
coded numerically.  See the documentation for \code{\link{faux.magnolia.high}}.
}

\usage{
 data(faux.magnolia.high.withnumericattributes)}
 

}

\keyword{datasets}

\section{Licenses and Citation}{
If the source of the data set does not specified otherwise,
this data set is protected by the Creative Commons License
\url{http://creativecommons.org/licenses/by-nc-nd/2.5/}.

When publishing results obtained using this data set, the original authors 
(Resnick et al, 1997) should
be cited. In addition this package should be cited as:

Mark S. Handcock, David R. Hunter, Carter T. Butts, Steven M. Goodreau,
and Martina Morris. 2003 
\emph{statnet: Software tools for the Statistical Modeling of Network Data}
\cr
\url{statnet.org}.
}



