% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{fixedas-ergmConstraint}
\alias{fixedas-ergmConstraint}
\alias{InitErgmConstraint.fixedas}
\title{Preserve and preclude edges}
\usage{
# fixedas(present, absent)
}
\arguments{
\item{present, absent}{a two-column edge list or a \code{\link{network}}}
}
\description{
Preserve the edges in 'present' and preclude the edges in 'absent'.
}
\note{
The current implementation of \code{fixedas()} simply fixes the dyads found in the \code{present} and \code{absent} lists, regardless of their status. That is, if a dyad in the \code{present} edge list does not already have an edge, no edge is created, and analogously with the \code{absent} edge list. Thus, it is up to the user to ensure that this is the case in their LHS network. This may change in the future.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "fixedas", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
