% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{For-ergmTerm}
\alias{For-ergmTerm}
\alias{InitErgmTerm.For}
\title{A \code{\link{for}} operator for terms}
\usage{
# binary: For(...)
}
\arguments{
\item{...}{in any order, \itemize{

\item one \emph{unnamed} one-sided \code{\link[=ergm]{ergm()}}-style formula with the
terms to be evaluated, containing one or more placeholders
\var{VAR} \emph{and}

\item one or more \emph{named} expressions of the form \code{\var{VAR}
  = \var{SEQ}} specifying the placeholder and its range. See
Details below.

}}
}
\description{
This operator evaluates the formula given to it,
substituting the specified loop counter variable with each
element in a sequence.
}
\details{
Placeholders are specified in the style of \code{foreach}, as
\code{\var{VAR} = \var{SEQ}}. \var{VAR} can be any valid \R
variable name, and \var{SEQ} can be a vector, a list, a function
of one argument, or a one-sided formula.  The vector or list will
be used directly, whereas a function will be called with the
network as its argument to produced the list, and the formula
will be used analogously to \code{\link[purrr:as_mapper]{purrr::as_mapper()}}, its RHS
evaluated in an environment in which the network itself will be
accessible as \code{.}.

If more than one named expression is given, they will be expanded
as one would expect in a nested \code{\link{for}} loop: earlier expressions
will form the outer loops and latter expressions the inner loops.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{operator}
