% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{ForEach-ergmTerm}
\alias{ForEach-ergmTerm}
\alias{InitErgmTerm.ForEach}
\title{A \code{\link{for}} operator for terms}
\usage{
# binary: ForEach(formula, counter, list)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated; it should contain a placeholder indicated by \code{counter}}

\item{counter}{a string giving the name of the counter}

\item{list}{a \code{\link{list}} of values to substitute for the counter, or a function with one argument or a one sided \code{\link{formula}} to construct the list from the network}
}
\description{
This operator evaluates the formula given to it for each value of the counter.
}
\details{
\code{list} can be a function of one argument or a one-sided formula. If the former, it will be called with the network as its argument, and if the latter, the expression on the RHS of the formula will be
evaluated in an environment in which the network itself will be accessible as \code{.}.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{operator}
