% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{get.vertex.attributes}
\alias{get.vertex.attributes}
\title{Get multiple vertex attributes at once and paste them together.}
\usage{
get.vertex.attributes(x, attrnames, na.omit = FALSE, null.na = TRUE,
  unlist = TRUE, sep = NULL)
}
\arguments{
\item{x, na.omit, null.na, unlist}{see \code{\link[network:get.vertex.attribute]{network::get.vertex.attribute()}}.}

\item{attrnames}{a character vector of one or more vertex attribute
names; see \code{\link[network:get.vertex.attribute]{network::get.vertex.attribute()}}.}

\item{sep}{an optional character vector of length 1 to use as a
separator for attribute values.}
}
\value{
If \code{sep} is \code{NULL}, a list with an element for each element
of \code{attrnames} containing the vertex attribute vector for that
attribute. Otherwise, if a character vector, convert the
attribute values to strings and join them with \code{sep} as the
separator.
}
\description{
Get multiple vertex attributes at once and paste them together.
}
