% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofN.R
\name{gofN}
\alias{gofN}
\alias{plot.gofN}
\title{Linear model diagnostics for multinetwork linear models}
\usage{
gofN(object, GOF, subset = TRUE, control = control.gof.ergm(), ...)

\method{plot}{gofN}(x, against = NULL, which = 1:2, col = 1, pch = 1,
  cex = 1, ..., ask = dev.interactive())
}
\arguments{
\item{object}{an \code{\link{ergm}} object.}

\item{GOF}{a one-sided \code{\link{ergm}} formula specifying network statistics whose goodness of fit to test.}

\item{subset}{argument for the \code{\link[=ergm-terms]{N}} term.}

\item{\dots}{additional arguments to \code{simulate.ergm()} and \code{summary.ergm_model()}.}

\item{against}{vector of values, network attribute, or a formula whose RHS gives an expression in terms of network attributes to plot against; if \code{NULL} (default), plots against fitted values.}

\item{which}{which to plot (\code{1} for residuals plot, \code{2} for \eqn{\sqrt{|R_i|}} scale plot).}

\item{col, pch, cex}{vector of values (wrapped in \code{\link[=I]{I()}}), network attribute, or a formula whose RHS gives an expression in terms of network attributes to plot against.}

\item{See}{\code{\link[=control.gof.ergm]{control.gof.ergm()}}.}
}
\description{
Linear model diagnostics for multinetwork linear models
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: A plotting method, making residuals and scale-location plots.
}}

