% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{greaterthan-ergmTerm}
\alias{greaterthan-ergmTerm}
\alias{InitWtErgmTerm.greaterthan}
\title{Number of dyads with values strictly greater than a threshold}
\usage{
# valued: greaterthan(threshold=0)
}
\arguments{
\item{threshold}{a vector of numerical values}
}
\description{
Adds the number of statistics equal to the length of \code{threshold}
equaling to the number of dyads whose values exceed the
corresponding element of \code{threshold} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "greaterthan", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
