% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{gwb1dsp-ergmTerm}
\alias{gwb1dsp-ergmTerm}
\alias{InitErgmTerm.gwb1dsp}
\title{Geometrically weighted dyadwise shared partner distribution for dyads in the first bipartition}
\usage{
# binary: gwb1dsp(decay=0, fixed=FALSE, cutoff=30)
}
\arguments{
\item{decay, fixed}{This value non-negative which should be non-negative. The value supplied for
this parameter may be fixed (if \code{fixed=TRUE} ),
or it may be used instead as the starting value for the estimation of \code{decay}
in a curved exponential family model (when \code{fixed=FALSE} , the default) (see Hunter and Handcock, 2006).}

\item{cutoff}{This optional argument sets the number of underlying b1dsp terms
to use in computing the statistics when \code{fixed=FALSE}, in order to reduce
the computational burden.  Its default value can also be controlled by the
\code{gw.cutoff} term option control parameter. (See \code{?control.ergm}.)}
}
\description{
This term adds one network statistic to the model equal to the geometrically
weighted dyadwise shared partner distribution for dyads in the first bipartition with decay parameter
\code{decay} parameter, which should be non-negative. This term can only be used with bipartite networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{bipartite}
\concept{curved}
\concept{undirected}
