% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{gwdsp-ergmTerm}
\alias{gwdsp-ergmTerm}
\alias{InitErgmTerm.gwdsp}
\title{Geometrically weighted dyadwise shared partner distribution}
\usage{
# binary: gwdsp(decay, fixed=FALSE, cutoff=30)
}
\arguments{
\item{fixed}{optional argument indicating
whether the \code{decay} parameter is fixed at the given value, or is to be fit as a curved
exponential-family model (see Hunter and Handcock, 2006). The
default is \code{FALSE} , which means the scale parameter is not
fixed and thus the model is a CEF model.}

\item{cutoff}{This optional argument sets the number of underlying DSP terms
to use in computing the statistics when \code{fixed=FALSE}, in order to reduce
the computational burden.  Its default value can also be controlled by the
\code{gw.cutoff} term option control parameter. (See \code{?control.ergm}.)}
}
\description{
Geometrically weighted dyadwise shared partner distribution
}
\details{
This term adds one network statistic to the model equal to the geometrically
weighted dyadwise shared partner distribution with decay parameter
\code{decay} parameter, which should be non-negative.
This term can be used with directed and undirected networks.
}
\note{
For directed networks, only outgoing two-path ("OTP") shared partners
are counted.  In other words, for a (directed) dyad \verb{(i,j)} in a directed graph, the number of
shared partners counted by \code{gwdsp} is the number of nodes \code{k} that have edges
\code{i -> k -> j}.  (These may also be called homogeneous shared partners.)  To
count other types of shared partners instead, see \code{dgwdsp}.
}
\concept{curved}
\concept{directed}
\concept{undirected}
