% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{hamming-ergmConstraint}
\alias{hamming-ergmConstraint}
\alias{InitErgmConstraint.hamming}
\title{Preserve the hamming distance to the given network (BROKEN: Do NOT Use)}
\usage{
# hamming
}
\description{
This constraint is currently broken. Do not use.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "hamming", "subsection")}
}
\concept{directed}
\concept{undirected}
