% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{hamming-ergmConstraint}
\alias{hamming-ergmConstraint}
\alias{InitErgmConstraint.hamming}
\title{TODO}
\usage{
# hamming
}
\description{
TODO
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{directed}
\concept{undirected}
