% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.operator.R
\name{I-ergmConstraint}
\alias{I-ergmConstraint}
\alias{InitErgmConstraint.I}
\title{Substitute a formula into the constraints formula}
\usage{
# I(formula)
}
\arguments{
\item{formula}{a constraints formula}
}
\description{
This is a convenience operator that can be used to
paste terms constructed elsewhere into a formula.
}
\note{
\code{formula} can also be a \link[statnet.common:term_list]{term_list}.
}
\seealso{
\code{\link[base:AsIs]{base::I()}} (a.k.a. \code{AsIs})

\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "I", "subsection")}
}
\concept{operator}
