% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.degfactorial.R
\name{idegm1lfactorial-ergmTerm}
\alias{idegm1lfactorial-ergmTerm}
\alias{InitErgmTerm.idegm1lfactorial}
\title{Sum of log-factorials of indegrees of non-in-isolates minus 1}
\usage{
# binary: idegm1lfactorial
}
\description{
This term adds one network statistic to the model,
equalling \eqn{\sum_{i=1}^n \mathbb{I}(id_i>0) \log(id_i - 1)!}.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "idegm1lfactorial", "subsection")}
}
\concept{directed}
