% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{idegreedist-ergmConstraint}
\alias{idegreedist-ergmConstraint}
\alias{InitErgmConstraint.idegreedist}
\title{Preserve the indegree distribution}
\usage{
# idegreedist
}
\description{
Preserve the indegree distribution of the given network.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "idegreedist", "subsection")}
}
\concept{directed}
