% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{idegreepopularity-ergmTerm}
\alias{idegreepopularity-ergmTerm}
\alias{InitErgmTerm.idegreepopularity}
\title{In-degree popularity (deprecated)}
\usage{
# binary: idegreepopularity
}
\description{
In-degree popularity (deprecated)
}
\details{
see \code{idegree1.5} .
}
\concept{deprecated}
\concept{directed}
