% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{idegrees-ergmConstraint}
\alias{idegrees-ergmConstraint}
\alias{InitErgmConstraint.idegrees}
\title{Preserve indegree for directed networks}
\usage{
# idegrees
}
\description{
For directed networks, preserve the indegree of each vertex of the given
network, while allowing outdegree to vary
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{directed}
