% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{ininterval-ergmTerm}
\alias{ininterval-ergmTerm}
\alias{InitWtErgmTerm.ininterval}
\title{Number of dyads whose values are in an interval}
\usage{
# valued: ininterval(lower=-Inf, upper=+Inf, open=c(TRUE,TRUE))
}
\arguments{
\item{lower}{defaults to -Inf}

\item{upper}{defaults to +Inf}

\item{open}{a \code{logical} vector of length 2 that controls whether
the interval is open (exclusive) on the lower and on the upper
end, respectively. \code{open} can also be specified as one of
\code{"[]"} , \code{"(]"} , \code{"[)"} , and \code{"()"} .}
}
\description{
Adds one statistic equaling to the number of dyads whose values
are between \code{lower} and \code{upper} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "ininterval", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
