% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{intransitive-ergmTerm}
\alias{intransitive-ergmTerm}
\alias{InitErgmTerm.intransitive}
\title{Intransitive triads}
\usage{
# binary: intransitive
}
\description{
This term adds one statistic to the model, equal to the number of triads in
the network that are intransitive. The intransitive triads are those of type
\verb{111D} , \code{201} , \verb{111U} , \verb{021C} , or \verb{030C} in the
categorization of Davis and Leinhardt (1972). For details on the 16 possible
triad types, see \code{triad.classify} in the
\code{sna} package. Note the distinction from the \code{ctriple}
term.
}
\note{
This term can only be used with directed networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{triad-related}
