% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{isolatededges-ergmTerm}
\alias{isolatededges-ergmTerm}
\alias{InitErgmTerm.isolatededges}
\title{Isolated edges}
\usage{
# binary: isolatededges
}
\description{
This term adds one statistic to the
model equal to the number of isolated edges in the network, i.e., the number
of edges each of whose endpoints has degree 1. This term can only be used
with undirected networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{bipartite}
\concept{undirected}
