% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{isolates-ergmTerm}
\alias{isolates-ergmTerm}
\alias{InitErgmTerm.isolates}
\title{Isolates}
\usage{
# binary: isolates
}
\description{
This term adds one statistic to the
model equal to the number of isolates in the network. For an undirected
network, an isolate is defined to be any node with degree zero. For a
directed network, an isolate is any node with both in-degree and out-degree
equal to zero.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{frequently-used}
\concept{undirected}
