% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{istar-ergmTerm}
\alias{istar-ergmTerm}
\alias{InitErgmTerm.istar}
\title{In-stars}
\usage{
# binary: istar(k, attr=NULL, levels=NULL)
}
\arguments{
\item{k}{a vector of distinct integers}

\item{attr}{a vertex attribute specification (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{levels}{TODO (See Specifying Vertex
attributes and Levels (\code{\link[=nodal_attributes]{?nodal_attributes}}) for details.)}
}
\description{
This term adds one network statistic to the
model for each element in \code{k} . The \eqn{i} th such statistic counts the
number of distinct \code{k[i]} -instars in the network, where a
\eqn{k} -instar is defined to be a node \eqn{N} and a set of \eqn{k}
different nodes \eqn{\{O_1, \dots, O_k\}}{\{O[1], ..., O[k]\}} such that the ties
\eqn{(O_j{\rightarrow}N)}{(O_j, N)} exist for \eqn{j=1, \dots, k} . If \code{attr} is specified
then the count is over the number of \eqn{k} -instars where all nodes have
the same value of the attribute. This term can only be used for directed
networks; for undirected networks see \code{kstar} . Note that
\code{istar(1)} is equal to both \code{ostar(1)} and \code{edges} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "istar", "subsection")}
}
\concept{categorical nodal attribute}
\concept{directed}
