% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{kstar-ergmTerm}
\alias{kstar-ergmTerm}
\alias{InitErgmTerm.kstar}
\title{k-Stars}
\usage{
# binary: kstar(k, attr=NULL, levels=NULL)
}
\arguments{
\item{k}{a vector of distinct integers}

\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{levels}{TODO (See Specifying Vertex
attributes and Levels (\code{\link[=nodal_attributes]{?nodal_attributes}}) for details.)}
}
\description{
This term adds one
network statistic to the model for each element in \code{k} . The \eqn{i} th
such statistic counts the number of distinct \code{k[i]} -stars in the
network, where a \eqn{k} -star is defined to be a node \eqn{N} and a set of
\eqn{k} different nodes \eqn{\{O_1, \dots, O_k\}}{\{O[1], ..., O[k]\}} such that the ties
\eqn{\{N, O_i\}}{\{N, O[i]\}} exist for \eqn{i=1, \dots, k} . If this is specified then the count is over
the number of \eqn{k} -stars where all nodes have the same value of the
attribute. This term can only be used for undirected networks; for directed
networks, see \code{istar} , \code{ostar} , \code{twopath} and \code{m2star} .
Note that \code{kstar(1)} is equal to \code{edges} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{categorical nodal attribute}
\concept{undirected}
