% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R,
%   R/InitWtErgmTerm.operator.R
\name{Label-ergmTerm}
\alias{Label-ergmTerm}
\alias{InitErgmTerm.Label}
\alias{InitWtErgmTerm.Label}
\title{Modify terms' coefficient names}
\usage{
# binary: Label(formula, label, pos)

# valued: Label(formula, label, pos)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}

\item{label}{either a character vector specifying the label for the terms or a function through which term names are mapped (or a \code{as_mapper} -style formula).}

\item{pos}{controls how \code{label} modifies the term names: one of \code{"prepend"} , \code{"replace"} , \code{"append"} , or \code{"("} , with the latter wrapping the term names in parentheses like a function call with name specified by \code{label} .}
}
\description{
This operator evaluates \code{formula} without modification, but modifies its coefficient and/or parameter names based on \code{label} and \code{pos} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "Label", "subsection")}
}
\concept{operator}
