% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{Layer}
\alias{Layer}
\title{A multilayer network representation.}
\usage{
Layer(...)
}
\arguments{
\item{...}{layer specification, in one of three formats:
\enumerate{
\item An (optionally named) list of identically-dimensioned and
directed networks.
\item Several networks as (optionally named) arguments.
\item A single network and a chararacter vector. Then, the layers are
values of the named edge attributes.
}}
}
\value{
A network object with layer metadata.
}
\description{
A function for specifying the LHS of a multilayer (a.k.a. multiplex
a.k.a. multirelational a.k.a. multivariate) ERGM.
}
\examples{

data(florentine)

# Method 1: list of networks
flo <- Layer(list(m = flomarriage, b = flobusiness))
ergm(flo ~ L(~edges, ~m)+L(~edges, ~b))

# Method 2: networks as arguments
flo <- Layer(m = flomarriage, b = flobusiness)
ergm(flo ~ L(~edges, ~m)+L(~edges, ~b))

# Method 3: edge attributes:
flo <- flomarriage
flo <- flomarriage | flobusiness
flo[,, names.eval="m"] <- as.matrix(flomarriage)
flo[,, names.eval="b"] <- as.matrix(flobusiness)
flo <- Layer(flo, c("m","b"))
ergm(flo ~ L(~edges, ~m)+L(~edges, ~b))

}
\seealso{
\link[=ergm-terms]{Help on model specification} for specific terms.
}
