% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.ergm.R
\name{logLikNull}
\alias{logLikNull}
\alias{logLikNull.ergm}
\title{Calculate the null model likelihood}
\usage{
logLikNull(object, ...)

\method{logLikNull}{ergm}(object, control = control.logLik.ergm(), ...)
}
\arguments{
\item{object}{a fitted model.}

\item{...}{further arguments to lower-level functions.

\code{logLikNull} computes, when possible the log-probability of
the data under the null model (reference distribution).}

\item{control}{A list of control parameters for algorithm tuning,
typically constructed with \code{\link[=control.logLik.ergm]{control.logLik.ergm()}}. Its documentation
gives the the list of recognized control parameters and their
meaning. The more generic utility \code{\link[=snctrl]{snctrl()}} (StatNet ConTRoL)
also provides argument completion for the available control
functions and limited argument name checking.}
}
\value{
\code{logLikNull} returns an object of type \code{\link{logLik}} if it is
able to compute the null model probability, and \code{NA} otherwise.
}
\description{
Calculate the null model likelihood
}
\section{Methods (by class)}{
\itemize{
\item \code{logLikNull(ergm)}: A method for \code{\link{ergm}} fits; currently only
implemented for binary ERGMs with dyad-independent sample-space
constraints.

}}
