% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{loop-ergmTerm}
\alias{loop-ergmTerm}
\alias{InitErgmTerm.loop}
\alias{InitWtErgmTerm.loop}
\title{Number of self-loops in the network}
\usage{
# binary: loop

# valued: loop(form="sum")
}
\arguments{
\item{form}{how to aggregate tie values in a valued ERGM: \code{"sum"}
(the default) for a statistic of the form \eqn{\sum_{i,j} x_{i,j}
  y_{i,j}}{sum[i,j] x[i,j]*y[i,j]}, where \eqn{y_{i,j}}{y[i,j]} is
the value of dyad \eqn{(i,j)} and \eqn{x_{i,j}}{x[i,j]} is the
term's covariate associated with it; and \code{"nonzero"} with the
edge considered to be present if its value is not 0. See
\code{\link{ergmTerm}} for more information.}
}
\description{
This term adds one network statistic equal to the number of
self-loops in the network.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "loop", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
