% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{loopcov-ergmTerm}
\alias{loopcov-ergmTerm}
\alias{InitErgmTerm.loopcov}
\alias{InitWtErgmTerm.loopcov}
\title{Main effect of a covariate on the presence of self-loops}
\usage{
# binary: loopcov(attr)

# valued: loopcov(attr, form="sum")
}
\arguments{
\item{attr}{a vertex attribute specification (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{form}{how to aggregate tie values in a valued ERGM: \code{"sum"}
(the default) for a statistic of the form \eqn{\sum_{i,j} x_{i,j}
  y_{i,j}}{sum[i,j] x[i,j]*y[i,j]}, where \eqn{y_{i,j}}{y[i,j]} is
the value of dyad \eqn{(i,j)} and \eqn{x_{i,j}}{x[i,j]} is the
term's covariate associated with it; and \code{"nonzero"} with the
edge considered to be present if its value is not 0. See
\code{\link{ergmTerm}} for more information.}
}
\description{
This term adds a single network statistic for each quantitative attribute or matrix column to the model equaling
\code{attr(i)} for self-loops \eqn{(i,i)} in the
network. For categorical attributes, see \code{loopfactor}.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "loopcov", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{quantitative nodal attribute}
\concept{self-loop}
\concept{undirected}
