% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{meandeg-ergmTerm}
\alias{meandeg-ergmTerm}
\alias{InitErgmTerm.meandeg}
\title{Mean vertex degree}
\usage{
# binary: meandeg
}
\description{
This term adds one network statistic to the model equal to the
average degree of a node. Note that this term is a constant multiple of
both \code{edges} and \code{density} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "meandeg", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
