% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molecule}
\alias{molecule}
\title{Synthetic network with 20 nodes and 28 edges}
\usage{
data(molecule)
}
\description{
This is a synthetic network of 20 nodes that is used as an example within
the \code{\link[=ergm]{ergm()}} documentation. It has an interesting elongated shape
\itemize{
\item reminencent of a chemical molecule.  It is stored as a
\code{\link[network:network]{network}} object.
}
}
\seealso{
florentine, sampson, network, plot.network, ergm
}
\keyword{datasets}
