% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{nearsimmelian-ergmTerm}
\alias{nearsimmelian-ergmTerm}
\alias{InitErgmTerm.nearsimmelian}
\title{Near simmelian triads}
\usage{
# binary: nearsimmelian
}
\description{
This term adds one statistic to the model equal to the number of near
Simmelian triads, as defined by Krackhardt and Handcock (2007). This is a
sub-graph of size three which is exactly one tie short of being complete.
}
\note{
This term can only be used with directed networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "nearsimmelian", "subsection")}
}
\concept{directed}
\concept{triad-related}
