% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.node.attr.R
\name{node-attr}
\alias{node-attr}
\alias{attrname}
\alias{on}
\alias{by}
\title{Specifying nodal attributes and their levels}
\description{
This document describes both the ways in which to specify nodal
attribute or functions and which levels for categorical factors to
include, and the helper functions for use in \code{InitErgmTerm}
implementations.
}
\details{
Term nodal attribute arguments, typically called \code{attrname}, \code{by},
\code{on}, etc. are interpreted as follows: \describe{

\item{a single character string}{Extract the vertex attribute.}

\item{a character vector of length > 1}{Extract the vertex
attributes and pastes them together, separated by dots.}

\item{a function}{The function is called on the LHS network,
expected to return a vector of appropriate length. (Shorter vectors
will be recycled as needed.)}

\item{a formula}{The expression on the RHS of the formula is
evaluated in an environment of the vertex attributes of the
network, expected to return a vector of appropriate
length. (Shorter vectors will be recycled as needed.) Within this
expression, the network itself accessible as either \code{.} or
\code{.nw}. For example,
\code{nodecov(~abs(Grade-mean(Grade))/network.size(.))} would return the
absolute difference of each actor's "Grade" attribute from its
network-wide mean, divided by the network size.}

}

For categorical attributes, to select which levels are of interest
and their ordering, use the argument \code{levels}.  It is interpreted
as follows: \describe{

\item{a numeric or logical vector}{Used for indexing of the default
set of levels (typically, unique values of the attribute) in
default older (typically lexicographic), i.e.,
\code{sort(unique(attr))[levels]}. Negative values exclude. To specify
numeric or logical levels literally, wrap in \code{\link[=I]{I()}}.}

\item{a character vector}{Use as is.}

\item{a function}{The function is called on the list of unique
values of the attribute, the values of the attribute themselves,
and the network itself, depending on its arity. Its return value is
interpreted as above.}

\item{a formula}{The expression on the RHS of the formula is
evaluated in an environment in which the network itself is
accessible as \code{.nw}, the list of unique values of the attribute as
\code{.} or as \code{.levels}, and the attribute vector itself as
\code{.attr}. Its return value is interpreted as above.}

}

Note that \code{levels} often has a default that is sensible for the
term in question.
}
