% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.diversity.R
\name{nodefactordistinct-ergmTerm}
\alias{nodefactordistinct-ergmTerm}
\alias{InitErgmTerm.nodefactordistinct}
\title{Number of distinct neighbor types}
\usage{
# binary: nodefactordistinct(attr, levels=TRUE)
}
\arguments{
\item{attr}{a vertex attribute specification (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{levels}{this optional argument controls which levels of the attribute
attributes and Levels (\code{\link[=nodal_attributes]{?nodal_attributes}}) for details.)}
}
\description{
This term adds a single network statistic to the
model, counting, for each node, the number of distinct values of
the attribute found among its neighbors.
}
\details{
This is a network analogue of the statistic introduced by
\insertCite{HoBl23m;textual}{ergm}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "nodefactordistinct", "subsection")}
}
\concept{categorical nodal attribute}
\concept{directed}
\concept{undirected}
