% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.blockop.R
\name{NodematchFilter-ergmTerm}
\alias{NodematchFilter-ergmTerm}
\alias{InitErgmTerm.NodematchFilter}
\title{Filtering on nodematch}
\usage{
# binary: NodematchFilter(formula, attrname)
}
\arguments{
\item{formula}{formula to be evaluated}

\item{attrname}{a character vector giving one or more names of attributes in the network's vertex attribute list.}
}
\description{
Evaluates the terms specified in \code{formula} on a network
constructed by taking \eqn{y} and removing any edges for which
\code{attrname(i)!=attrname(j)} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{operator}
