% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{nodemix-ergmTerm}
\alias{nodemix-ergmTerm}
\alias{InitErgmTerm.nodemix}
\alias{InitWtErgmTerm.nodemix}
\title{Nodal attribute mixing}
\usage{
# binary: nodemix(attr, base=NULL, b1levels=NULL, b2levels=NULL, levels=NULL, levels2=-1)

# valued: nodemix(attr, base=NULL, b1levels=NULL, b2levels=NULL, levels=NULL, levels2=-1, form="sum")
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{base}{deprecated}

\item{b1levels, b2levels, levels, level2}{control what statistics are included in the model and the order in which they appear. \code{levels2} apply to all networks; \code{levels} applies to unipartite networks; \code{b1levels} and \code{b2levels} apply to bipartite networks (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details)}

\item{form}{character how to aggregate tie values in a valued ERGM}
}
\description{
Nodal attribute mixing
}
\details{
By default, this term adds one network statistic to
the model for each possible pairing of attribute values. The
statistic equals the number of edges in the network in which the
nodes have that pairing of values. (When multiple attributes are specified, a
statistic is added for each combination of attribute values for
those attributes.) In other words, this term produces one statistic for
every entry in the mixing matrix for the attribute(s). By default, the ordering of
the attribute values is lexicographic: alphabetical (for nominal categories) or
numerical (for ordered categories).
}
\note{
The argument \code{base} is retained for backwards compatibility and may be
removed in a future version. When both \code{base} and \code{levels} are passed,
\code{levels} overrides \code{base}.

The argument \code{base} is retained for backwards compatibility and may be
removed in a future version. When both \code{base} and \code{levels2} are passed,
\code{levels2} overrides \code{base}.
}
\concept{categorical nodal attribute}
\concept{directed}
\concept{dyad-independent}
\concept{frequently-used}
\concept{undirected}
