% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.diversity.R
\name{nodeocovrange-ergmTerm}
\alias{nodeocovrange-ergmTerm}
\alias{InitErgmTerm.nodeocovrange}
\title{Range of covariate values for out-neighbors of a node}
\usage{
# binary: nodeocovrange(attr)
}
\arguments{
\item{attr}{a vertex attribute specification (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}
}
\description{
This term adds a single network statistic equalling
the sum over the nodes of the range over of its neighbors'
values.
}
\details{
This is a network analogue of the statistic introduced by
\insertCite{HoBl23m;textual}{ergm}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "nodeocovrange", "subsection")}
}
\concept{directed}
\concept{quantitative nodal attribute}
