% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{nodeofactor-ergmTerm}
\alias{nodeofactor-ergmTerm}
\alias{InitErgmTerm.nodeofactor}
\alias{InitWtErgmTerm.nodeofactor}
\title{Factor attribute effect for out-edges}
\usage{
# binary: nodeofactor(attr, base=1, levels=-1)

# valued: nodeofactor(attr, base=1, levels=-1, form="sum")
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{base}{deprecated}

\item{levels}{this optional argument controls which levels of the attribute
should be included and which should be excluded. (See Specifying Vertex
attributes and Levels (\code{?nodal_attributes}) for details.) For example, if
the "fruit" attribute has levels "orange", "apple", "banana", and "pear",
then to include just two levels, one for "apple" and one for "pear", use
any of \code{nodeofactor("fruit", levels=-(2:3))}, \code{nodeofactor("fruit", levels=c(1,4))}, and
\code{nodeofactor("fruit", levels=c("apple", "pear"))}. Note: if you are using numeric
values to specify the levels of a character variable, the levels will
correspond to the alphabetically sorted character levels.}

\item{form}{character how to aggregate tie values in a valued ERGM}
}
\description{
Factor attribute effect for out-edges
}
\details{
This term adds multiple network
statistics to the model, one for each of (a subset of) the unique
values of the \code{attr} attribute (or each combination of the
attributes given). Each of these statistics gives the number of
times a node with that attribute or those attributes appears as the
node of origin of a directed tie.
}
\note{
The argument \code{base} is retained for backwards compatibility and may be
removed in a future version. When both \code{base} and \code{levels} are passed,
\code{levels} overrides \code{base}.

To include all attribute values is usually not a good idea, because
the sum of all such statistics equals the number of edges and hence a linear
dependency would arise in any model also including \code{edges}. The default,
\code{levels=-1}, is therefore to omit the first (in lexicographic order)
attribute level. To include all levels, pass either \code{levels=TRUE}
(i.e., keep all levels) or \code{levels=NULL} (i.e., do not filter levels).

This term can only be used with directed networks.
}
\concept{categorical nodal attribute}
\concept{directed}
\concept{dyad-independent}
