% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R,
%   R/InitWtErgmTerm.operator.R
\name{NoLoops-ergmTerm}
\alias{NoLoops-ergmTerm}
\alias{InitErgmTerm.NoLoops}
\alias{InitWtErgmTerm.NoLoops}
\title{Remove self-loops before evaluating}
\usage{
# binary: NoLoops(formula)

# valued: NoLoops(formula)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}
}
\description{
Evaluates the given \code{formula} on a network constructed removing all self-loops.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "NoLoops", "subsection")}
}
\concept{operator}
\concept{self-loop}
