% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{nonzero-ergmTerm}
\alias{nonzero-ergmTerm}
\alias{InitWtErgmTerm.edges}
\title{Edges}
\usage{
# valued: nonzero
# valued: edges
}
\description{
Edges
}
\details{
This term adds one network statistic equal
to the number of edges (i.e. nonzero values) in the network. For undirected networks, \code{edges}
is equal to \code{kstar(1)} ; for directed networks, \code{edges} is equal to
both \code{ostar(1)} and \code{istar(1)} .
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
