% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{observed-ergmConstraint}
\alias{observed-ergmConstraint}
\alias{InitErgmConstraint.observed}
\title{Preserve the observed dyads of the given network}
\usage{
# observed
}
\description{
Preserve the observed dyads of the given network.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "observed", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
