% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{observed-ergmConstraint}
\alias{observed-ergmConstraint}
\alias{InitErgmConstraint.observed}
\title{Preserve the observed dyads of the given network}
\usage{
# observed
}
\description{
Preserve the observed dyads of the given network
}
\details{
Preserve the observed dyads of the given network.
}
\concept{dyad-independent}
