% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.degfactorial.R
\name{odegm1lfactorial-ergmTerm}
\alias{odegm1lfactorial-ergmTerm}
\alias{InitErgmTerm.odegm1lfactorial}
\title{Sum of log-factorials of outdegrees of non-out-isolates minus 1}
\usage{
# binary: odegm1lfactorial
}
\description{
This term adds one network statistic to the model,
equalling \eqn{\sum_{i=1}^n \mathbb{I}(od_i>0) \log(od_i - 1)!}.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "odegm1lfactorial", "subsection")}
}
\concept{directed}
