% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{odegree-ergmTerm}
\alias{odegree-ergmTerm}
\alias{InitErgmTerm.odegree}
\title{Out-degree}
\usage{
# binary: odegree(d, by=NULL, homophily=FALSE, levels=NULL)
}
\arguments{
\item{d}{a vector of distinct integers}
}
\description{
Out-degree
}
\details{
This term adds one network statistic to
the model for each element in \code{d} ; the \eqn{i} th such statistic equals
the number of nodes in the network of out-degree \code{d[i]} , i.e. the
number of nodes with exactly \code{d[i]} out-edges.
This term can only be used with directed networks; for undirected networks
see \code{degree} .
}
\concept{categorical nodal attribute}
\concept{directed}
\concept{frequently-used}
