% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{odegree1.5-ergmTerm}
\alias{odegree1.5-ergmTerm}
\alias{InitErgmTerm.odegree1.5}
\title{Out-degree to the 3/2 power}
\usage{
# binary: odegree1.5
}
\description{
This term adds one network statistic to the model equaling the sum
over the actors of each actor's outdegree taken to the 3/2 power
(or, equivalently, multiplied by its square root). This term is
analogous to the term of \insertCite{SnVa10i;textual}{ergm}, equation (12). This
term can only be used with directed networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "odegree1.5", "subsection")}
}
\concept{directed}
