% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{odegreedist-ergmConstraint}
\alias{odegreedist-ergmConstraint}
\alias{InitErgmConstraint.odegreedist}
\title{Preserve the outdegree distribution}
\usage{
# odegreedist
}
\description{
Preserve the outdegree distribution of the given network.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmConstraint", "odegreedist", "subsection")}
}
\concept{directed}
