% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{odegreepopularity-ergmTerm}
\alias{odegreepopularity-ergmTerm}
\alias{InitErgmTerm.odegreepopularity}
\title{Out-degree popularity (deprecated)}
\usage{
# binary: odegreepopularity
}
\description{
Out-degree popularity (deprecated)
}
\details{
see \code{odegree1.5} .
}
\concept{deprecated}
\concept{directed}
