% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{odegrees-ergmConstraint}
\alias{odegrees-ergmConstraint}
\alias{InitErgmConstraint.odegrees}
\title{Preserve outdegree for directed networks}
\usage{
# odegrees
}
\description{
Preserve outdegree for directed networks
}
\details{
For directed networks, preserve the outdegree of each vertex of the given
network, while allowing indegree to vary
}
