% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{Offset-ergmTerm}
\alias{Offset-ergmTerm}
\alias{InitErgmTerm.Offset}
\title{Terms with fixed coefficients}
\usage{
# binary: Offset(formula, coef, which)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}

\item{coef}{coefficients to the formula}

\item{which}{used to specify which of the parameters in the formula are fixed. It can be a logical vector (recycled as needed), a numeric vector of indices of parameters to be fixed, or a character vector of parameter names.}
}
\description{
This operator is analogous to the \code{offset()} wrapper, but the
coefficients are specified within the term and the curved ERGM
mechanism is used internally.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "Offset", "subsection")}
}
\concept{operator}
