% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.utils.R
\name{old.formula.utils}
\alias{old.formula.utils}
\alias{ergm.update.formula}
\alias{remove.offset.formula}
\title{Obsolete formula and model utilities. They will be removed in a
future release.}
\usage{
ergm.update.formula(object, new, ..., from.new = FALSE)

remove.offset.formula(object, response = NULL)
}
\arguments{
\item{object}{an \code{ergm} formula.}

\item{new, ..., from.new}{See \code{\link[=nonsimp_update.formula]{nonsimp_update.formula()}}.}

\item{response}{Name of the edge attribute whose value is to be
modeled in the valued ERGM framework. Defaults to \code{NULL} for
simple presence or absence, modeled via a binary ERGM.}
}
\description{
\code{ergm.update.formula} (DEPRECATED: use
\code{\link[statnet.common]{nonsimp_update.formula}} instead) is a
reimplementation of \code{\link{update.formula}} that does not
simplify.  Note that the resulting formula's environment is set as
follows. If \code{from.new==FALSE}, it is set to that of
object. Otherwise, a new sub-environment of object, containing, in
addition, variables in new listed in from.new (if a character
vector) or all of new (if TRUE).

\code{remove.offset.formula} deletes all \code{\link{offset}} terms
in an ERGM formula. (DEPRECATED: Use \code{statnet.common::filter_rhs.formula(object, function(x) (if(is.call(x)) x[[1]] else x)!="offset")}.)

[[1]: R:[1
}
