% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{opentriad-ergmTerm}
\alias{opentriad-ergmTerm}
\alias{InitErgmTerm.opentriad}
\title{Open triads}
\usage{
# binary: opentriad
}
\description{
This term
adds one statistic to the model equal to the number of 2-stars minus
three times the number of triangles in the network. It is currently
only implemented for undirected networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "opentriad", "subsection")}
}
\concept{triad-related}
\concept{undirected}
