% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{pack.Clist_as_num}
\alias{pack.Clist_as_num}
\title{Serialize model information in a \code{Clist} into a numeric vector}
\usage{
pack.Clist_as_num(Clist)
}
\arguments{
\item{Clist}{output of \code{ergm.Cprepare}}
}
\value{
a numeric vector concatenating the following:
\itemize{
\item number of terms in the model;
\item length of and encoded string of term names;
\item length of and encoded string of library names; and
\item vector of intputs to the model.
}

This is intended to be decoded by \code{unpack_*Model_as_double()} C routines.
}
\description{
This function takes the output of \code{ergm.Cprepare} that pertains to
the model (not the network) and serializes it by encoding term
names and libraries as numeric vectors and concatenating them with
inputs and other information.
}
