% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{pack.str_as_num}
\alias{pack.str_as_num}
\title{Convert a string into a "standard" format for passing as a numeric variable.}
\usage{
pack.str_as_num(s)
}
\arguments{
\item{s}{a character vector of length 1}
}
\value{
a numeric vector concatenating the following:
\itemize{
\item the number of characters in the input string and
\item the characters in the input string encoded using ASCII encoding.
}

This is intended to be decoded by \code{unpack_str_as_double()} C routines.
}
\description{
This function takes a character vector of length 1 and returns a numerical vector encoding it.
}
