% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{passthrough.curved.ergm_model}
\alias{passthrough.curved.ergm_model}
\title{Wrap a submodel's curved specification (if present) for output from an \code{InitErgmTerm} or \code{InitWtErgmTerm}.}
\usage{
passthrough.curved.ergm_model(m, namewrap = identity)
}
\arguments{
\item{m}{an \code{ergm_model} object}

\item{namewrap}{an optional function taking a character string and
returning a character string, gold on the model's curved
parameter names to wrap them.}
}
\value{
a list with elements \code{map}, \code{gradient}, and \code{params}
suitable for concatenating with an \code{InitErgmTerm} or
\code{InitWtErgmTerm} output list.
}
\description{
Given a \code{ergm} model and (optionally) a function with which to wrap
parameter names, wrap the calls to its \code{ergm.eta()} and
\code{ergm.etagrad()} into \code{map()} and \code{gradient()} functions, similarly
with the \code{params} element.
}
