% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ergm.R
\name{plot.ergm}
\alias{plot.ergm}
\title{Deprecated alias for \code{\link[=mcmc.diagnostics]{mcmc.diagnostics()}}.}
\usage{
\method{plot}{ergm}(x, ...)
}
\arguments{
\item{x}{an \code{\link[=ergm]{ergm()}} object.}

\item{...}{additional arguments, passed to \code{\link[=mcmc.diagnostics]{mcmc.diagnostics()}}.}
}
\description{
Originally, a plotting method for latent positions, before latent
space modelling was split out into the \code{latentnet} package, it now
prints the MCMC diagnostics, but it may be repurposed in the
future.
}
\seealso{
\code{\link[=mcmc.diagnostics]{mcmc.diagnostics()}}
}
