% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ergm.R
\name{predict.ergm}
\alias{predict.ergm}
\title{ERGM-based conditional tie probabilities}
\usage{
\method{predict}{ergm}(object, net, ...)
}
\arguments{
\item{object}{fitted ERGM model object}

\item{net}{network object}

\item{...}{other arguments passed to/from other methods}
}
\value{
N x N matrix of conditional tie probabilities, where N is the number of nodes
in \code{net}.
}
\description{
This function calculates \strong{conditional} tie probabilities for all dyads in a
given network based on an ERGM model.
}
