% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R,
%   R/InitWtErgmTerm.operator.R
\name{Prod-ergmTerm}
\alias{Prod-ergmTerm}
\alias{InitErgmTerm.Prod}
\alias{InitWtErgmTerm.Prod}
\title{A product (or an arbitrary power combination) of one or more formulas}
\usage{
# binary: Prod(formulas, label)

# valued: Prod(formulas, label)
}
\arguments{
\item{formulas}{formulas to be evaluated}

\item{label}{used to specify the names of the elements of the resulting term product vector. If \code{label} is of length 1, it will be recycled with indices appended.}
}
\description{
A product (or an arbitrary power combination) of one or more formulas
}
\details{
This operator evaluates a list of formulas whose corresponnding RHS
statistics will be multiplied elementwise. They are required to be nonnegative.

If a formula has an LHS, it is interpreted as follows:
\itemize{
\item a numeric scalar: Network statistics of this formula will be exponentiated by this.
\item a numeric vector: Corresponding network statistics of this formula will be exponentiated by this.
\item a numeric matrix: Vector of network statistics will be exponentiated by this using the same pattern as matrix multiplication.
\item a character string: One of several predefined linear combinations. Currently supported presets are as follows:
\itemize{
\item \code{"prod"}: Network statistics of this formula will be multiplied together; equivalent to \code{matrix(1,1,p)} , where \code{p} is the length of the network statistic vector.
\item \code{"geomean"}: Network statistics of this formula will be geometrically averaged; equivalent to \code{matrix(1/p,1,p)} , where \code{p} is the length of the network statistic vector.
}
}

Note that each formula must either produce the same number of
statistics or be mapped through a matrix to produce the same
number of statistics.

A single formula is also permissible. This can be useful if one
wishes to, say, multiply together the statistics returned by a formula.

Curved models are supported, subject to some limitations. In particular, the first model's etamap will be used, overwriting the others. If \code{label} is not of length 1, it should have an \code{attr} -style attribute \code{"curved"} specifying the names for the curved parameters.
}
\note{
The current implementation piggybacks on the \code{Log} , \code{Exp} , and \code{Sum} operators, essentially \code{Exp(~Sum(~Log(formula), label))} . This may result in loss of precision, particularly for extremely large or small statistics. The implementation may change in the future.
}
\concept{operator}
