% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.projection.R
\name{Project-ergmTerm}
\alias{Project-ergmTerm}
\alias{InitErgmTerm.Project}
\alias{InitErgmTerm.Proj1}
\alias{Proj1-ergmTerm}
\alias{InitErgmTerm.Proj2}
\alias{Proj2-ergmTerm}
\title{Evaluation on a projection of a bipartite network}
\usage{
# binary: Project(formula, mode)

# binary: Proj1(formula)

# binary: Proj2(formula)
}
\arguments{
\item{formula}{a one-sided \code{\link[=ergm]{ergm()}}-style formula with the terms to be evaluated}

\item{mode}{the mode onto which to project: 1 or 2}
}
\description{
This operator on a bipartite network evaluates the
formula on the undirected, valued network constructed by
projecting it onto its specified mode. \code{Proj1(formula)} and
\code{Proj2(formula)} are aliases for \code{Project(formula, 1)} and
\code{Project(formula, 2)}, respectively.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "Project", "subsection")}
}
\concept{bipartite}
\concept{operator}
