% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{randomtoggle-ergmProposal}
\alias{randomtoggle-ergmProposal}
\alias{InitErgmProposal.randomtoggle}
\title{Propose a randomly selected dyad to toggle}
\description{
Propose a randomly selected dyad to toggle
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="randomtoggle"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="randomtoggle"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="randomtoggle"))}}
}
