% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_test.ergm.R
\name{rank_test.ergm}
\alias{rank_test.ergm}
\title{A lack-of-fit test for ERGMs}
\usage{
rank_test.ergm(x, plot = FALSE)
}
\arguments{
\item{x}{an \code{\link[=ergm]{ergm()}} object.}

\item{plot}{if \code{TRUE}, plot the empirical distribution.}
}
\value{
The sample quantile of the observed network's probability
among the predicted.
}
\description{
A simple test reporting the sample quantile of the observed
network's probability in the distribution under the MLE. This is a
conservative p-value for the null hypothesis of the observed
network being a draw from the distribution of interest.
}
