% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{simmelian-ergmTerm}
\alias{simmelian-ergmTerm}
\alias{InitErgmTerm.simmelian}
\title{Simmelian triads}
\usage{
# binary: simmelian
}
\description{
This term adds one
statistic to the model equal to the number of Simmelian triads, as defined
by Krackhardt and Handcock (2007). This is a complete sub-graph of size
three.
}
\note{
This term can only be used with directed networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{triad-related}
