% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{simmelianties-ergmTerm}
\alias{simmelianties-ergmTerm}
\alias{InitErgmTerm.simmelianties}
\title{Ties in simmelian triads}
\usage{
# binary: simmelianties
}
\description{
This term adds
one statistic to the model equal to the number of ties in the network that
are associated with Simmelian triads, as defined by \insertCite{KrHa07h;textual}{ergm}.
Each Simmelian has six ties in it but, because Simmelians can
overlap in terms of nodes (and associated ties), the total number of ties in
these Simmelians is less than six times the number of Simmelians. Hence this
is a measure of the clustering of Simmelians (given the number of
Simmelians).
}
\note{
This term can only be used with directed networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "simmelianties", "subsection")}
}
\concept{directed}
\concept{triad-related}
