% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{smallerthan-ergmTerm}
\alias{smallerthan-ergmTerm}
\alias{InitWtErgmTerm.smallerthan}
\title{Number of dyads with values strictly smaller than a threshold}
\usage{
# valued: smallerthan(threshold=0)
}
\arguments{
\item{threshold}{vector of numerical values}
}
\description{
Adds the number of statistics equal to the length of \code{threshold}
equaling to the number of dyads whose values are exceeded by the
corresponding element of \code{threshold} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "smallerthan", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
