% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.hints.R
\name{sparse-ergmConstraint}
\alias{sparse-ergmConstraint}
\alias{InitErgmConstraint.sparse}
\title{Sparse network}
\usage{
# sparse
}
\description{
Sparse network
}
\details{
The network is sparse. This typically results in a Tie-Non-Tie (TNT) proposal regime.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{dyad-independent}
