% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{SPDyad-ergmProposal}
\alias{SPDyad-ergmProposal}
\alias{InitErgmProposal.SPDyad}
\title{TODO}
\description{
TODO
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="SPDyad"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="SPDyad"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="SPDyad"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "SPDyad", "subsection")}
}
\keyword{internal}
