\name{statnet-internal}
\title{Internal statnet Objects}
% terms
\alias{ergm.checkargs}
\alias{ergm.checkdirected}
%
\alias{InitErgm.absdiff}
\alias{InitErgm.bounded.degree}
\alias{InitErgm.bounded.idegree}
\alias{InitErgm.bounded.istar}
\alias{InitErgm.bounded.kstar}
\alias{InitErgm.bounded.odegree}
\alias{InitErgm.bounded.ostar}
\alias{InitErgm.bounded.triangle}
\alias{InitErgm.sociality}
\alias{InitErgm.ctriad}
\alias{InitErgm.cycle}
\alias{InitErgm.density}
\alias{InitErgm.degree}
\alias{InitErgm.dsp}
\alias{InitErgm.duration}
\alias{InitErgm.dyadcov}
\alias{InitErgm.dyadcov}
\alias{InitErgm.edgecov}
\alias{InitErgm.edges}
\alias{InitErgm.esp}
\alias{InitErgm.gwdegree}
\alias{InitErgm.gwdsp}
\alias{InitErgm.gwesp}
\alias{InitErgm.hamming}
\alias{InitErgm.idegree}
\alias{InitErgm.istar}
\alias{InitErgm.isolates}
\alias{InitErgm.kstar}
\alias{InitErgm.localtriangle}
\alias{InitErgm.mutual}
\alias{InitErgm.nodematch}
\alias{InitErgm.nodemain}
\alias{InitErgm.nodemix}
\alias{InitErgm.receiver}
\alias{InitErgm.receivercov}
\alias{InitErgm.sender}
\alias{InitErgm.sendercov}
\alias{InitErgm.nodefactor}
\alias{InitErgm.odegree}
\alias{InitErgm.ostar}
\alias{InitErgm.smalldiff}
\alias{InitErgm.triangle}
\alias{InitErgm.tripercent}
\alias{InitErgm.ttriad}
\alias{InitErgm.spatial}
\alias{InitErgm.match}
\alias{InitErgm.nodecov}
\alias{InitErgm.geodegree}
\alias{InitErgm.balance}
%
\alias{degreedistfactor}
\alias{drawcircle}
\alias{drawpie}
\alias{dspartnerdist}
\alias{ergm.checkargs}
\alias{ergm.checkdegeneracy}
\alias{ergm.checkdirected}
\alias{ergm.curved}
\alias{ergm.curved.statsmatrix}
\alias{ergm.curved.update}
\alias{ergm.degenerate}
\alias{ergm.estimate}
\alias{ergm.eta}
\alias{ergm.etagrad}
\alias{ergm.etagradmult}
\alias{ergm.etamap}
\alias{ergm.geodesicmatrix}
\alias{ergm.geodesicmatrix.edgelist}
\alias{ergm.geodesicmatrix.old}
\alias{ergm.geodist}
\alias{ergm.geodistdist}
\alias{ergm.geodistn}
\alias{ergm.getMCMCsample}
\alias{ergm.getnetwork}
\alias{ergm.independencemodel}
\alias{ergm.initialfit}
\alias{ergm.logisticdeviance}
\alias{ergm.logitreg}
\alias{ergm.mainfitloop}
\alias{ergm.marquardt}
\alias{ergm.marquardt2}
\alias{ergm.nodegeodesics}
\alias{ergm.pairgeodesic}
\alias{ergm.raftery.diag}
\alias{ergm.revisetheta0}
\alias{ergm-terms}
\alias{ergm.t.summary}
\alias{espartnerdist}
\alias{get.node.attr}
\alias{is.ergm}
\alias{is.latent}
\alias{is.latent.cluster}
\alias{is.matrixnetwork}
\alias{llik.fun}
\alias{llik.fun2}
\alias{llik.fun3}
\alias{llik.grad}
\alias{llik.grad2}
\alias{llik.grad3}
\alias{llik.hessian}
\alias{llik.hessian2}
\alias{llik.info3}
\alias{llik.mcmcvar3}
\alias{mcmc.diagnostics.default}
\alias{midarrow}
\alias{plot.mcmc.ergm}
\alias{print.raftery.diag.ergm}
\alias{robust.inverse}
\alias{rspartnerdist}
\alias{set.mfrow}
\alias{summary.formula}
\alias{traceplot.ergm}
\alias{twopathdist}
%
\alias{ergm.sufftoprob}
\alias{ergm.Cprepare}
\alias{ergm.getmodel}
\alias{ergm.getterms}
\alias{ergm.gettermnames}
\alias{ergm.mple}
\alias{ergm.statseval}
\alias{fulldistdist}
\alias{fullgcount}
%\alias{get.node.attr}
%\alias{geodist}
%\alias{has.loops}
%\alias{is.directed}
%\alias{is.hyper}
%\alias{lpz.dist}
%\alias{mahal}
%\alias{mlpY}
%\alias{mlpY.grad}
%\alias{mlpY0}
%\alias{mlpY0.grad}
%\alias{mlpYmdsZ}
%\alias{mlpYmdsZ.grad}
%\alias{mlpYnull}
%\alias{prepmahal}
%\alias{symmetrize}
%
\alias{ergm.MCMCse}
%\alias{ergm.latentplot}
%\alias{mlpY.grad.plot}
%\alias{mlpY.plot}
%\alias{mlpYmdsZ.grad.plot}
%\alias{mlpYmdsZ.plot}
\alias{ergm.MCMCacf}
\alias{ergm.pen.glm}
\alias{logistftest}
%
\alias{ergm.mainloop}
\alias{ergm.robmon}
%
% These need to be documented!
%
\alias{print.network.series}
\alias{summary.statsmatrix.ergm}
\alias{summary.network.series}
\alias{ergm.boundDeg}
\alias{is.invertible}
\alias{ostar2deg}
\alias{sufftoprob}
\alias{degreedist}
\alias{mixingmatrix}

\description{
  Internal ergm functions.
}
%\usage{
%anova.ergm(object, \dots)
%}
\details{
  Most of these are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
}
%\arguments{
% \item{\dots}{further arguments passed to or used by methods.}
%}
\seealso{ergm, statnet-package}
\keyword{internal}
