% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{StdNormal-ergmReference}
\alias{StdNormal-ergmReference}
\alias{InitErgmReference.StdNormal}
\title{Standard Normal reference}
\usage{
# StdNormal
}
\description{
Specifies each dyad's baseline distribution to be the normal distribution
with mean 0 and variance 1.
}
\seealso{
\code{\link{ergmReference}} for index of reference distributions currently visible to the package.
}
\concept{continuous}
