% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{sum-ergmTerm}
\alias{sum-ergmTerm}
\alias{InitWtErgmTerm.sum}
\title{Sum of dyad values (optionally taken to a power)}
\usage{
# valued: sum(pow=1)
}
\arguments{
\item{pow}{power of dyad values. Defaults to 1.}
}
\description{
This term adds one statistic equal to the sum of
dyad values taken to the power \code{pow}.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{undirected}
