% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.statistics.network.R
\name{summary_statistics}
\alias{summary_statistics}
\alias{summary}
\alias{summary.formula}
\alias{summary_statistics.ergm}
\alias{summary_statistics.network.list}
\alias{summary_statistics.network}
\alias{summary_statistics.matrix}
\alias{summary_statistics.default}
\title{Calculation of network or graph statistics}
\usage{
summary_statistics(object, ..., basis = NULL)

\method{summary}{formula}(object, \dots)

\method{summary_statistics}{ergm}(object, ..., basis = NULL)

\method{summary_statistics}{network.list}(object, response = NULL, ...,
  basis = NULL)

\method{summary_statistics}{network}(object, response = NULL, ...,
  basis = NULL)

\method{summary_statistics}{matrix}(object, response = NULL, ...,
  basis = NULL)

\method{summary_statistics}{default}(object, response = NULL, ...,
  basis = NULL)
}
\arguments{
\item{object}{Either an \code{\link{formula}} object (see above) or an
\code{\link{ergm}} model object.  In the latter case,
\code{summary_statistics} is called for the \code{object$formula} object.
In the former case, \code{object} is of the form \code{y ~ <model terms>},
where \code{y} is a \code{\link[network]{network}} object or a matrix that
can be coerced to a \code{\link[network]{network}} object.  For the details
on the possible \code{<model terms>}, see \code{\link{ergm-terms}}.  To
create a \code{\link[network]{network}} object in , use the \code{network()}
function, then add nodal attributes to it using the \code{\%v\%} operator if
necessary.}

\item{\dots}{further arguments passed to or used by methods.}

\item{basis}{An optional \code{\link[network]{network}} object relative to
which the global statistics should be calculated.}

\item{response}{Name of the edge attribute whose value is to be modeled.
Defaults to \code{NULL} for simple presence or absence, modeled via binary
ERGM terms. Passing anything but \code{NULL} uses valued ERGM terms.}
}
\value{
A vector of statistics measured on the network.
}
\description{
Used to calculate the specified statistics for an observed network if its
argument is a formula for an \code{\link{ergm}}.  See
\code{\link{ergm-terms}} for more information on the statistics that may be
specified.
}
\details{
If \code{object} is of class \code{\link{formula}}, then
\code{\link[base]{summary}} may be used in lieu of \code{summary_statistics}
because \code{summary.formula} calls the \code{summary_statistics} function.
%If neither of those are given as the
%first argument then a \code{ergm.Cprepare} object is expected. This last
%option is meant for internal use.
The function actually cumulates the change statistics when
removing edges from the observed network one by one until the empty network
results.  Since each model term has a prespecified value (zero by default)
for the corresponding statistic(s) on an empty network, these change
statistics give the absolute statistics on the original network.

\code{summary.formula} for networks understands the \code{\link{lasttoggle}}
"API".

%More information can be found by looking at the documentation of
%\code{\link{ergm}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{ergm}: an \code{\link{ergm}} fit method, extracting its model from the fit.

\item \code{network.list}: a method for a \code{\link{network.list}} on the LHS of the formula.

\item \code{network}: a method for a \code{\link{network}} on the LHS of the formula.

\item \code{matrix}: a method for a \code{\link{matrix}} on the LHS of the formula.

\item \code{default}: a fallback method.
}}

\examples{

#
# Lets look at the Florentine marriage data
#
data(florentine)
#
# test the summary_statistics function
#
summary(flomarriage ~ edges + kstar(2))
m <- as.matrix(flomarriage)
summary(m ~ edges)  # twice as large as it should be
summary(m ~ edges, directed=FALSE) # Now it's correct

}
\seealso{
\code{\link[=ergm]{ergm()}}, \code{\link[=network]{network()}}, \link{ergm-terms}
}
\keyword{models}
