% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{symmetrize.network}
\alias{symmetrize.network}
\title{Return a symmetrized version of a binary network}
\usage{
symmetrize.network(x, rule = c("weak", "strong", "upper", "lower"))
}
\arguments{
\item{x}{a \code{\link{network}} object.}

\item{rule}{a string specifying how the network is to be
symmetrized; see \code{\link[sna:symmetrize]{sna::symmetrize()}} for details.}
}
\description{
Return a symmetrized version of a binary network
}
