% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{TNT-ergmProposal}
\alias{TNT-ergmProposal}
\alias{InitErgmProposal.TNT}
\title{Default MH algorithm}
\description{
Stratifies the population of dyads
edge status: those having ties and those having no ties (hence T/NT).
This is useful for improving performance in sparse networks,
because it gives at least 50\\% chance of proposing a toggle of an existing edge.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="TNT"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="TNT"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="TNT"))}}
}
