% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{transitive-ergmTerm}
\alias{transitive-ergmTerm}
\alias{InitErgmTerm.transitive}
\title{Transitive triads}
\usage{
# binary: transitive
}
\description{
This term adds one statistic to the model, equal to the number of triads in
the network that are transitive. The transitive triads are those of type
\verb{120D} , \verb{030T} , \verb{120U} , or \code{300} in the categorization
of Davis and Leinhardt (1972). For details on the 16 possible triad types,
see \code{?triad.classify} in the \CRANpkg{sna} package.
Note the distinction from the \code{ttriple} term. This term can only be
used with directed networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "transitive", "subsection")}
}
\concept{directed}
\concept{triad-related}
