% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{transitiveweights-ergmTerm}
\alias{transitiveweights-ergmTerm}
\alias{InitWtErgmTerm.transitiveweights}
\title{Transitive weights}
\usage{
# valued: transitiveweights(twopath="min", combine="max", affect="min")
}
\arguments{
\item{twopath}{the minimum
of the constituent dyads ( \code{"min"} ) or their geometric mean
( \code{"geomean"} )}

\item{combine}{the maximum of the
2-path strengths ( \code{"max"} ) or their sum ( \code{"sum"} )}

\item{affect}{the minimum of the focus dyad and the
combined strength of the two paths ( \code{"min"} ) or their
geometric mean ( \code{"geomean"} )}
}
\description{
This statistic implements the transitive weights
statistic defined by Krivitsky (2012), Equation 13. For each of these options,
the first (and the default) is more stable but also more
conservative, while the second is more sensitive but more likely
to induce a multimodal distribution of networks.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "transitiveweights", "subsection")}
}
\concept{directed}
\concept{nonnegative}
\concept{triad-related}
\concept{undirected}
