% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.hints.R
\name{triadic-ergmHint}
\alias{triadic-ergmHint}
\alias{InitErgmConstraint.triadic}
\alias{triadic-ergmConstraint}
\title{Network with strong clustering (triad-closure) effects}
\usage{
# triadic(triFocus = 0.25)
}
\arguments{
\item{triFocus}{A number between 0 and 1, indicating how often triad-focused proposals should be made relative to the standard proposals.}
}
\description{
The network has a high clustering coefficient. This typically results in alternating between the Tie-Non-Tie (TNT) proposal and a triad-focused proposal.
}
\seealso{
\code{\link{ergmHint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmHint", "triadic", "subsection")}
}
\concept{dyad-dependent}
