% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{twopath-ergmTerm}
\alias{twopath-ergmTerm}
\alias{InitErgmTerm.twopath}
\title{2-Paths}
\usage{
# binary: twopath
}
\description{
This term adds one statistic to the model, equal to the number of 2-paths in
the network. For a directed network this is defined as a pair of edges
\eqn{(i{\rightarrow}j), (j{\rightarrow}k)}{(i,j), (j,k)} , where \eqn{i} and
\eqn{j} must be distinct. That is, it is a directed path of length 2 from
\eqn{i} to \eqn{k} via \eqn{j} . For directed networks a 2-path is also a
mixed 2-star but the interpretation is usually different; see \code{m2star} .
For undirected networks a twopath is defined as a pair of edges
\eqn{\{i,j\}, \{j,k\}} . That is, it is an undirected path of length 2 from
\eqn{i} to \eqn{k} via \eqn{j} , also known as a 2-star.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{undirected}
