% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{Unif-ergmReference}
\alias{Unif-ergmReference}
\alias{InitErgmReference.Unif}
\title{Continuous Uniform reference}
\usage{
# Unif(a,b)
}
\arguments{
\item{a, b}{minimum and maximum to the baseline discrete uniform distribution, both inclusive. Both values must be finite.}
}
\description{
Specifies each dyad's baseline distribution to be continuous uniform
between \code{a} and \code{b}: \eqn{h(y)=1} , with the support being \verb{[a, b]}.
}
\seealso{
\code{\link{ergmReference}} for index of reference distributions currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmReference", "Unif", "subsection")}
}
\concept{continuous}
