% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_ergm.R
\name{vif_ergm}
\alias{vif_ergm}
\alias{vif_ergm.ergm_model}
\alias{vif_ergm.ergm}
\alias{print.vif_ergm}
\title{Generalized Variance-Inflation Factors for ERGMs}
\usage{
vif_ergm(object, ...)

\method{vif_ergm}{ergm_model}(object, v, drop_terms = NULL, drop_coef = NULL)

\method{vif_ergm}{ergm}(object, sources = "model", ...)

\method{print}{vif_ergm}(x, ...)
}
\arguments{
\item{object}{a fitted ERGM model object as returned by \code{\link[=ergm]{ergm()}} or ERGM
model specificaton as returned by \code{\link[=ergm_model]{ergm_model()}}. See Methods section
below.}

\item{...}{Other arguments passed to/from other methods.}

\item{v}{Variance-covariance matrix}

\item{drop_terms}{character vector of term labels}

\item{drop_coef}{character vector of coefficient names}

\item{sources}{one of \code{"all"}, \code{"model"}, or \code{"estimation"}. Whether to
return variance-covariance matrix from the ERGM model, the estimation
process, or both combined. This argument is passed to \code{\link[=vcov.ergm]{vcov.ergm()}}}

\item{x}{Object of class "vif_ergm"}
}
\value{
An object of class \code{vif_ergm}, a list of two data frames:
\itemize{
\item \code{vif_coef} with classical VIFs calculated per coefficient containing the
following columns:
\itemize{
\item \code{coef} -- ERGM parameter name
\item \code{vif} -- VIF
\item \code{df} -- Degrees of freedom, always 1
}
\item \code{vif_term} with GVIFs calculated per term containing the following columns:
\itemize{
\item \code{term} -- ERG model term as in the model formula
\item \code{vif} -- GVIF
\item \code{df} -- Degrees of freedom
\item \code{vif_avg} -- The value of \eqn{GVIF^(1 / (Df * 2))}, see Fox & Monette
(1992, sec. 3)
}
}
}
\description{
Variance Inflation-Factor (VIF) and its generalization (GVIF) due to Fox &
Monette (1992) adopted to the ERGM context (see also Duxbury 2018).
}
\section{Methods (by class)}{
\itemize{
\item \code{ergm_model}: Based on \code{ergm_model} object and VCov matrix

\item \code{ergm}: Method for \code{\link[=ergm]{ergm()}} models
}}

\references{
Duxbury, S. W. (2018). Diagnosing multicollinearity in
exponential random graph models. \emph{Sociological Methods & Research}
\doi{10.1177/0049124118782543}.

Fox, J., Monette, G. (1992). Generalized Collinearity Diagnostics. \emph{Journal
of the American Statistical Association} 87(417):178--183
\doi{10.2307/2290467}.
}
