% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.median.R
\name{wtd.median}
\alias{wtd.median}
\title{Weighted Median}
\usage{
wtd.median(x, na.rm = FALSE, weight = FALSE)
}
\arguments{
\item{x}{Vector of data, same length as \code{weight}}

\item{na.rm}{Logical: Should NAs be stripped before computation proceeds?}

\item{weight}{Vector of weights}
}
\value{
Returns an empirical .5 quantile from a weighted sample.
}
\description{
Compute weighted median.
}
\details{
Uses a simple algorithm based on sorting.
}
\keyword{robust}
