% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{absdiff}
\alias{absdiff}
\alias{InitErgmTerm.absdiff}
\title{Absolute difference}
\usage{
absdiff(attr, pow) # binary

absdiff(attr, pow, form="sum") # valued
}
\arguments{
\item{attr}{name of vertex attribute}

\item{pow}{power, default to 1}
}
\description{
Absolute difference
}
\details{
The attr argument specifies a quantitative attribute (see Specifying
Vertex attributes and Levels (? nodal_attributes) for details). This term
adds one network statistic to the model equaling the sum of
abs(attr\link{i}-attr\link{j})^pow for all edges (i,j) in the network.

Note that ergm versions 3.9.4 and earlier used different arguments for this
term. See the above section on versioning for invoking the old behavior.
}
\concept{binary}
\concept{directed}
\concept{dyad-independent}
\concept{nodal-quantitative}
\concept{undirected}
