% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{atleast-ergmTerm}
\alias{atleast-ergmTerm}
\alias{InitWtErgmTerm.atleast}
\title{Number of dyads with values greater than or equal to a threshold}
\usage{
# valued: atleast(threshold=0)
}
\arguments{
\item{threshold}{vector of numerical values}
}
\description{
Adds the number of
statistics equal to the length of \code{threshold}
equaling to the number of dyads whose values equal or exceed the
corresponding element of \code{threshold} .
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
