% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{b1degrees-ergmConstraint}
\alias{b1degrees-ergmConstraint}
\alias{InitErgmConstraint.b1degrees}
\title{Preserve the actor degree for bipartite networks}
\usage{
# b1degrees
}
\description{
For bipartite networks, preserve the degree for the first mode of each vertex of the given
network, while allowing the degree for the second mode to vary.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{bipartite}
