% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{b2degrees-ergmConstraint}
\alias{b2degrees-ergmConstraint}
\alias{InitErgmConstraint.b2degrees}
\title{Preserve the receiver degree for bipartite networks}
\usage{
# b2degrees
}
\description{
For bipartite networks, preserve the degree for the second mode of each vertex of the given
network, while allowing the degree for the first mode to vary.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{bipartite}
