% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{Bernoulli-ergmReference}
\alias{Bernoulli-ergmReference}
\alias{InitErgmReference.Bernoulli}
\title{Bernoulli reference}
\usage{
# Bernoulli
}
\description{
Specifies each
dyad's baseline distribution to be Bernoulli with probability of
the tie being \eqn{0.5} . This is the only reference measure used
in binary mode.
}
\seealso{
\code{\link{ergmReference}} for index of reference distributions currently visible to the package.
}
\concept{binary}
\concept{discrete}
\concept{finite}
\concept{nonnegative}
