% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_model.R
\name{call.ErgmTerm}
\alias{call.ErgmTerm}
\title{Locate and call an ERGM term initialization function.}
\usage{
call.ErgmTerm(term, env, nw, ..., term.options = list())
}
\arguments{
\item{term}{A term from an \code{\link[=ergm]{ergm()}} formula: typically a \code{\link{name}} or a
\code{\link{call}}.}

\item{env}{Environment in which it is to be evaluated.}

\item{nw}{A \code{\link[network:network]{network}} object.}

\item{...}{Additional term options.}

\item{term.options}{A list of optional settings such as calculation
tuning options to be passed to the \code{InitErgmTerm} functions.}
}
\value{
The list returned by the the \code{InitErgmTerm} or
\code{InitWtErgmTerm} function, with package name autodetected if
neede.
}
\description{
A helper function that searches attached and loaded packages for a
term with a specifies name, calls it with the specified arguments,
and returns the result.
}
\keyword{internal}
