% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_model.utils.R
\name{coef.sublength.model}
\alias{coef.sublength.model}
\alias{coef.length.model}
\alias{eta.sublength.model}
\alias{eta.length.model}
\title{Extract Number of parameters in ergm Model}
\usage{
\method{coef}{sublength.model}(object, offset = NA, ...)

\method{coef}{length.model}(object, offset = NA, ...)

eta.sublength.model(object, offset = NA, ...)

eta.length.model(object, offset = NA, ...)
}
\arguments{
\item{object}{an \code{\link{ergm_model}} object}

\item{offset}{If \code{NA} (the default), all model terms are counted;
if \code{TRUE}, only offset terms are counted; and if
\code{FALSE}, offset terms are skipped.}

\item{\dots}{other arguments.}
}
\value{
\code{coef.length.model} returns the length of the
parameter vector of the model.
}
\description{
\code{coef.sublength.model} returns a vector
containing the number of model parameters corresponding to each
model term.

\code{coef.length.model} returns the total number of parameters of
an \code{\link{ergm_model}}.

\code{eta.sublength.model} returns a vector containing
the number of canonical parameters (also the number of sufficient
statistics) corresponding to each model term.

\code{eta.length.model} returns the total number of
canonical parameters.
}
\note{
These are \emph{not} methods at this time. This may change in the
future.
}
\keyword{models}
