% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohab}
\alias{cohab}
\alias{cohab_MixMat}
\alias{cohab_PopWts}
\alias{cohab_TargetStats}
\title{Target statistics and model fit to a hypothetical network population based on
egocentric data on cohabitation}
\source{
Group, Stat N. (2011). \emph{Another Relevant Article about this
Dataset}.
}
\usage{
data(cohab)
}
\description{
This dataset consists of three objects,
a mixing matrix, a vector of targets for sufficient statistics, and a
data.frame of model-derived weights for a network dataset in which only
egocentric reports about partnerships are observed:
}
\details{
\describe{
\item{cohab_MixMat}{Mixing matrix for 50K pop (explain)}
\item{cohab_TargetStats}{Vector of target statistics for 50K pop (explain)}
\item{cohab_PopWts}{Data frame including model-derived weights (explain)}
}

(Insert explanation here regarding how these objects were created.)
}
\references{
Krivitsky, P.N., Hunter, D.R., Morris, M., and Klumb, C. (2011).
\emph{ergm 4.0: New Features and Improvements}.

Group, Stat N. (2011). \emph{Another Relevant Article about this Dataset}.
}
\seealso{
ergm
}
\keyword{datasets}
