% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{CondOutDegreeDist-ergmProposal}
\alias{CondOutDegreeDist-ergmProposal}
\alias{InitErgmProposal.CondOutDegreeDist}
\title{MHp for odegreedist constraints}
\description{
MHp for \eqn{constraints= ~odegreedist}. For directed networks, similar to
\code{InitErgmProposal.CondDegreeDist}, except for outdegree case
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="CondOutDegreeDist"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="CondOutDegreeDist"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="CondOutDegreeDist"))}}
}
