%  File man/deoffset.etamap.Rd in package ergm, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2017 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.model.utils.R
\name{deoffset.etamap}
\alias{deoffset.etamap}
\title{\code{ergm.model}'s \code{etamap} with all offset terms removed and remapped}
\usage{
deoffset.etamap(etamap)
}
\arguments{
\item{etamap}{the \code{etamap} element of the \code{ergm.model}.}
}
\value{
a copy of the input as if all offset terms were dropped
from the model
}
\description{
\code{ergm.model}'s \code{etamap} with all offset terms removed and remapped
}
