% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_model.utils.R
\name{deoffset.etamap}
\alias{deoffset.etamap}
\title{\code{ergm_model}'s \code{etamap} with all offset terms removed and remapped}
\usage{
deoffset.etamap(etamap)
}
\arguments{
\item{etamap}{the \code{etamap} element of the \code{ergm_model}.}
}
\value{
a copy of the input as if all offset terms were dropped
from the model
}
\description{
\code{ergm_model}'s \code{etamap} with all offset terms removed and remapped
}
