% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli}
\alias{ecoli}
\alias{ecoli1}
\alias{ecoli2}
\title{Two versions of an E. Coli network dataset}
\source{
The data set is based on the RegulonDB network of \insertCite{SaSu01r;textual}{ergm}
and was modified by \insertCite{ShSh02n;textual}{ergm}.
}
\usage{
data(ecoli)
}
\description{
This network data set comprises two versions of a biological network in
which the nodes are operons in \emph{Escherichia Coli} and a directed edge
from one node to another indicates that the first encodes the transcription
factor that regulates the second.
}
\details{
The network object \code{ecoli1} is directed, with 423 nodes and 519 arcs.
The object \code{ecoli2} is an undirected version of the same network, in
which all arcs are treated as edges and the five isolated nodes (which
exhibit only self-regulation in \code{ecoli1}) are removed, leaving 418
nodes.
}
\section{Licenses and Citation}{
 When publishing results obtained using this
data set, the original authors \insertCite{SaSu01r,ShSh02n}{ergm}
should be cited, along with this \code{R} package.
}

\references{
\insertAllCited{}

\%Saul and Filkov (2007)

\%Hummel et al (2010)
}
\keyword{datasets}
