% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\name{edges-ergmTerm}
\alias{edges-ergmTerm}
\alias{InitErgmTerm.edges}
\title{Edges}
\usage{
# binary: edges
}
\description{
Number of edges in the network.
}
\details{
This term adds one network statistic equal to the number of
edges (i.e. nonzero values) in the network. For undirected networks, \code{edges}
is equal to \code{kstar(1)}; for directed networks, edges is equal to both
\code{ostar(1)} and \code{istar(1)}.
}
\seealso{
Other dyad-independent: 
\code{\link{absdiff-ergmTerm}}

Other directed: 
\code{\link{absdiff-ergmTerm}}

Other undirected: 
\code{\link{absdiff-ergmTerm}}
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
