% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-disambiguation.R
\name{edges}
\alias{edges}
\title{\code{edges} (disambiguation)}
\description{
\code{edges} may refer to:
\itemize{
\item \link[=edges-term]{An ERGM statistic} (\code{help("edges-term")})
\item \link[=edges-constraint]{An ERGM sample space constraint} (\code{help("edges-constraint")})
}
}
