% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.R
\name{equalto-ergmTerm}
\alias{equalto-ergmTerm}
\alias{InitWtErgmTerm.equalto}
\title{Number of dyads with values equal to a specific value (within tolerance)}
\usage{
# valued: equalto(value=0, tolerance=0)
}
\arguments{
\item{value}{numerical threshold}

\item{tolerance}{numerical threshold}
}
\description{
Adds one statistic equal to the number of dyads whose values
are within \code{tolerance} of \code{value} , i.e., between
\code{value-tolerance} and \code{value+tolerance} , inclusive.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
