% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-defunct.R
\name{ergm-defunct}
\alias{ergm-defunct}
\alias{robust.inverse}
\alias{plot.network.ergm}
\alias{ergm.getterms}
\alias{plot.mcmc.list.ergm}
\alias{plot.ergm}
\alias{summary.statistics}
\alias{ergm.checkargs}
\alias{ergm.checkbipartite}
\alias{ergm.checkdirected}
\alias{summary.gof}
\alias{ergm.getMCMCsample}
\alias{ergm.MHP.table}
\alias{MHproposal}
\alias{MHproposal.character}
\alias{MHproposal.ergm}
\alias{MHproposal.formula}
\alias{ergm.init.methods}
\alias{ergm.ConstraintImplications}
\alias{ergm.mcmcslave}
\alias{ergm.update.formula}
\alias{remove.offset.formula}
\alias{network.update}
\alias{ergm.getmodel}
\alias{ergm.getglobalstats}
\alias{as.edgelist.compressed}
\alias{as.network.uncompressed}
\alias{standardize.network}
\alias{newnw.extract}
\alias{san.ergm}
\alias{is.inCH}
\title{Functions that have been removed from this package}
\usage{
sociality(object, ...)

robust.inverse(...)

\method{plot}{network.ergm}(...)

ergm.getterms(...)

\method{plot}{mcmc.list.ergm}(...)

\method{plot}{ergm}(...)

\method{summary}{statistics}(...)

ergm.checkargs(...)

ergm.checkbipartite(...)

ergm.checkdirected(...)

\method{summary}{gof}(...)

ergm.getMCMCsample(...)

ergm.MHP.table(...)

MHproposal(...)

MHproposal.character(...)

MHproposal.ergm(...)

MHproposal.formula(...)

ergm.init.methods(...)

ergm.ConstraintImplications(...)

ergm.mcmcslave(...)

ergm.update.formula(...)

remove.offset.formula(...)

network.update(...)

ergm.getmodel(...)

ergm.getglobalstats(...)

\method{as.edgelist}{compressed}(...)

\method{as.network}{uncompressed}(...)

standardize.network(...)

newnw.extract(...)

\method{san}{ergm}(...)

is.inCH(...)
}
\arguments{
\item{...}{Arguments to defunct functions.}
}
\description{
Functions that have been removed after a period of deprecation.
}
\keyword{internal}
