\name{ergm-deprecated}
\alias{ergm-deprecated}
\alias{degreedistfactor}
\alias{robust.inverse}
\alias{enformulate.curved}
\alias{enformulate.curved.ergm}
\alias{enformulate.curved.formula}
\alias{robust.inverse}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Functions that will no longer be supported in future releases of the package
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons

\itemize{
\item{enformulate.curved}
\item{enformulate.curved.ergm}
\item{enformulate.curved.formula}
\item{robust.inverse}
}
}
