% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_ergm-deprecated.R, R/InitErgmTerm.R,
%   R/ergm.check.R, R/ergm.getglobalstats.R, R/ergm.getterms.R, R/ergm_model.R,
%   R/formula.utils.R, R/gof.ergm.R, R/plot.ergm.R
\name{ergm-deprecated}
\alias{ergm-deprecated}
\alias{InitErgmTerm.degreepopularity}
\alias{InitErgmTerm.idegreepopularity}
\alias{InitErgmTerm.odegreepopularity}
\alias{ergm.checkargs}
\alias{ergm.checkbipartite}
\alias{ergm.checkdirected}
\alias{ergm.getglobalstats}
\alias{ergm.getterms}
\alias{ergm.getmodel}
\alias{ergm.update.formula}
\alias{remove.offset.formula}
\alias{summary.gof}
\alias{plot.ergm}
\title{Functions that will no longer be supported in future releases of the package}
\usage{
InitErgmTerm.degreepopularity(nw, arglist, ...)

InitErgmTerm.idegreepopularity(nw, arglist, ...)

InitErgmTerm.odegreepopularity(nw, arglist, ...)

ergm.checkargs(fname, arglist, varnames = NULL, vartypes = NULL,
  defaultvalues = list(), required = NULL)

ergm.checkbipartite(fname, nw.bipartiteflag, requirement, extramessage = "")

ergm.checkdirected(fname, nw.directedflag, requirement, extramessage = "")

ergm.getglobalstats(nw, m, response = NULL)

ergm.getterms(formula)

ergm.getmodel(object, ...)

ergm.update.formula(object, new, ..., from.new = FALSE)

remove.offset.formula(object, response = NULL)

\method{summary}{gof}(object, ...)

\method{plot}{ergm}(x, ...)
}
\arguments{
\item{nw, arglist, ..., fname, varnames, vartypes, defaultvalues, required, nw.bipartiteflag, requirement, extramessage, nw.directedflag, m, response, formula, object, new, from.new, x}{Arguments to deprecated functions.}
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons

\code{\link[=summary.gof]{summary.gof()}} is a deprecated alias for \code{\link[=print.gof]{print.gof()}}.

\code{plot.ergm}: deprecated alias for \code{\link[=mcmc.diagnostics]{mcmc.diagnostics()}}.
}
\section{Functions}{
\itemize{
\item \code{InitErgmTerm.degreepopularity}: Use \code{\link{degree1.5}} instead.

\item \code{InitErgmTerm.idegreepopularity}: Use \code{\link{idegree1.5}} instead.

\item \code{InitErgmTerm.odegreepopularity}: Use \code{\link{odegree1.5}} instead.

\item \code{ergm.checkargs}: Use \code{\link[=check.ErgmTerm]{check.ErgmTerm()}} instead.

\item \code{ergm.checkbipartite}: Use \code{\link[=check.ErgmTerm]{check.ErgmTerm()}} instead.

\item \code{ergm.checkdirected}: Use \code{\link[=check.ErgmTerm]{check.ErgmTerm()}} instead.

\item \code{ergm.getglobalstats}: Use \code{\link[=summary.ergm_model]{summary.ergm_model()}} instead.

\item \code{ergm.getterms}: Use \code{\link[statnet.common:list_rhs.formula]{statnet.common::list_rhs.formula()}} and \code{\link[statnet.common:eval_lhs.formula]{statnet.common::eval_lhs.formula()}} instead.

\item \code{ergm.getmodel}: Use \code{ergm_model} instead.

\item \code{ergm.update.formula}: Use \code{\link[statnet.common]{nonsimp_update.formula}} instead.

\item \code{remove.offset.formula}: Use \code{\link[statnet.common:filter_rhs.formula]{statnet.common::filter_rhs.formula()}} such as \code{statnet.common::filter_rhs.formula(object, function(x) (if(is.call(x)) x[[1]] else x)!="offset")} instead.

[[1]: R:[1
}}

