% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-deprecated.R, R/InitErgmTerm.R,
%   R/ergm.utility.R, R/ergm_model.R, R/formula.utils.R, R/network.update.R,
%   R/summary.ergm_model.R
\name{ergm-deprecated}
\alias{ergm-deprecated}
\alias{InitErgmTerm.degreepopularity}
\alias{InitErgmTerm.idegreepopularity}
\alias{InitErgmTerm.odegreepopularity}
\alias{newnw.extract}
\alias{standardize.network}
\alias{ergm.getmodel}
\alias{ergm.update.formula}
\alias{remove.offset.formula}
\alias{offset.info.formula}
\alias{network.update}
\alias{ergm.getglobalstats}
\title{Functions that will no longer be supported in future releases of the package}
\usage{
InitErgmTerm.degreepopularity(nw, arglist, ...)

InitErgmTerm.idegreepopularity(nw, arglist, ...)

InitErgmTerm.odegreepopularity(nw, arglist, ...)

newnw.extract(oldnw, z = NULL, output = "network", response = NULL)

standardize.network(nw, preserve.eattr = TRUE)

ergm.getmodel(object, ...)

ergm.update.formula(object, new, ..., from.new = FALSE)

remove.offset.formula(object, response = NULL)

offset.info.formula(object, response = NULL, ...)

network.update(nw, newmatrix, matrix.type = NULL, output = "network",
  ignore.nattr = c("bipartite", "directed", "hyper", "loops", "mnext",
  "multiple", "n"), ignore.vattr = c())

ergm.getglobalstats(nw, m, response = NULL)
}
\arguments{
\item{nw, arglist, ..., fname, varnames, vartypes, defaultvalues, required, nw.bipartiteflag, requirement, extramessage, nw.directedflag, m, response, formula, object, new, from.new, x, model, proposal, eta0, control, verbose, update.nws, Clist, prev.run, burnin, samplesize, interval, maxedges, attrname, label, coord, jitter, thresh, usearrows, mode, displayisolates, interactive, xlab, ylab, xlim, ylim, pad, label.pad, displaylabels, boxed.labels, label.pos, label.bg, vertex.sides, vertex.rot, arrowhead.cex, label.cex, loop.cex, vertex.cex, edge.col, label.col, vertex.col, label.border, vertex.border, edge.lty, label.lty, vertex.lty, edge.lwd, label.lwd, edge.len, edge.curve, edge.steps, loop.steps, object.scale, uselen, usecurve, suppress.axes, vertices.last, layout.par, cex.main, cex.sub, seed, latent.control, colornames, latent, preserve.eattr}{Arguments to deprecated functions.}
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons
}
\section{Functions}{
\itemize{
\item \code{InitErgmTerm.degreepopularity}: Use \code{\link{degree1.5}} instead.

\item \code{InitErgmTerm.idegreepopularity}: Use \code{\link{idegree1.5}} instead.

\item \code{InitErgmTerm.odegreepopularity}: Use \code{\link{odegree1.5}} instead.

\item \code{newnw.extract}: Use the \code{\link{pending_update_network}} "API".

\item \code{ergm.getmodel}: Use \code{ergm_model} instead.

\item \code{ergm.update.formula}: Use \code{\link[statnet.common]{nonsimp_update.formula}} instead.

\item \code{remove.offset.formula}: Use \code{\link[statnet.common:filter_rhs.formula]{statnet.common::filter_rhs.formula()}} such as \code{statnet.common::filter_rhs.formula(object, function(x) (if(is.call(x)) x[[1]] else x)!="offset")} instead.

\item \code{offset.info.formula}: \code{offset.info.formula} returns the offset
vectors associated with a formula.

\item \code{network.update}: Use \code{\link[=update.network]{update.network()}} instead.

\item \code{ergm.getglobalstats}: Use \code{\link[=summary.ergm_model]{summary.ergm_model()}} instead.
}}

\keyword{internal}
\keyword{misc}
