% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-deprecated.R, R/ergmlhs.R
\name{ergm-deprecated}
\alias{ergm-deprecated}
\alias{coef.ergm}
\alias{$.ergm}
\alias{convert_ergmlhs}
\title{Functions that will no longer be supported in future releases of the package}
\usage{
\method{coef}{ergm}(object, ...)

\method{$}{ergm}(x, name)

convert_ergmlhs(lhs)
}
\arguments{
\item{..., object, x}{Arguments to deprecated functions.}

\item{name}{See \link{Extract}.}
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons
}
\section{Functions}{
\itemize{
\item \code{coef(ergm)}: extracts the \code{ergm} parameters; may be
removed in favour of the default method once the number of \code{ergm}
objects with \verb{$coef} elements in the wild is sufficiently low.

\item \code{$}: accesses elements of \code{ergm} objects;
needed for backwards compatibility when components get renamed.

\item \code{convert_ergmlhs()}: \code{convert_ergmlhs} converts old-style settings to new-style settings.

}}
\keyword{internal}
\keyword{misc}
