% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-deprecated.R, R/InitErgmTerm.R,
%   R/ergm.san.R, R/formula.utils.R
\name{ergm-deprecated}
\alias{ergm-deprecated}
\alias{coef.ergm}
\alias{$.ergm}
\alias{InitErgmTerm.degreepopularity}
\alias{InitErgmTerm.hammingmix}
\alias{hammingmix}
\alias{InitErgmTerm.idegreepopularity}
\alias{InitErgmTerm.odegreepopularity}
\alias{san.ergm}
\alias{offset.info.formula}
\title{Functions that will no longer be supported in future releases of the package}
\usage{
\method{coef}{ergm}(object, ...)

\method{$}{ergm}(x, name)

InitErgmTerm.degreepopularity(nw, arglist, ...)

InitErgmTerm.hammingmix(nw, arglist, ..., version = packageVersion("ergm"))

InitErgmTerm.idegreepopularity(nw, arglist, ...)

InitErgmTerm.odegreepopularity(nw, arglist, ...)

\method{san}{ergm}(
  object,
  formula = object$formula,
  constraints = object$constraints,
  target.stats = object$target.stats,
  nsim = NULL,
  basis = NULL,
  output = c("network", "edgelist", "ergm_state"),
  only.last = TRUE,
  control = object$control$SAN,
  verbose = FALSE,
  offset.coef = NULL,
  ...
)

offset.info.formula(object, ...)
}
\arguments{
\item{name}{See \link{Extract}.}

\item{nw, arglist, ..., fname, varnames, vartypes, defaultvalues, required, nw.bipartiteflag, requirement, extramessage, nw.directedflag, m, response, formula, object, new, from.new, x, model, proposal, eta0, control, verbose, update.nws, Clist, prev.run, burnin, samplesize, interval, maxedges, attrname, label, coord, jitter, thresh, usearrows, mode, displayisolates, interactive, xlab, ylab, xlim, ylim, pad, label.pad, displaylabels, boxed.labels, label.pos, label.bg, vertex.sides, vertex.rot, arrowhead.cex, label.cex, loop.cex, vertex.cex, edge.col, label.col, vertex.col, label.border, vertex.border, edge.lty, label.lty, vertex.lty, edge.lwd, label.lwd, edge.len, edge.curve, edge.steps, loop.steps, object.scale, uselen, usecurve, suppress.axes, vertices.last, layout.par, cex.main, cex.sub, seed, latent.control, colornames, latent, preserve.eattr}{Arguments to deprecated functions.}
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons
}
\section{Functions}{
\itemize{
\item \code{coef.ergm}: extracts the \code{ergm} parameters; may be
removed in favour of the default method once the number of \code{ergm}
objects with \verb{$coef} elements in the wild is sufficiently low.

\item \code{$.ergm}: accesses elements of \code{ergm} objects;
needed for backwards compatibility when components get renamed.

\item \code{InitErgmTerm.degreepopularity}: Use \code{\link{degree1.5}} instead.

\item \code{InitErgmTerm.idegreepopularity}: Use \code{\link{idegree1.5}} instead.

\item \code{InitErgmTerm.odegreepopularity}: Use \code{\link{odegree1.5}} instead.

\item \code{san.ergm}: The developers are not aware of a use case for this function. Please contact them if you would like to prevent its removal.

\item \code{offset.info.formula}: \code{offset.info.formula} returns the offset
vectors associated with a formula.
}}

\keyword{internal}
\keyword{misc}
