%  File man/ergm_deprecated.Rd in package ergm, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2017 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{ergm-deprecated}
\alias{robust.inverse}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Functions that will no longer be supported in future releases of the package
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons

\itemize{
\item{robust.inverse}
}
}
