% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-deprecated.R, R/InitErgmTerm.R,
%   R/ergm.Cprepare.R, R/ergm.check.R, R/ergm.design.R, R/ergm.getMCMCsample.R,
%   R/ergm.getterms.R, R/ergm_model.R, R/ergm_proposal.R,
%   R/formula.utils.R, R/gof.ergm.R, R/mcmc.diagnostics.ergm.R, R/nparam.R,
%   R/plot.ergm.R, R/plot.network.ergm.R, R/summary.ergm_model.R,
%   R/summary.statistics.network.R
\name{ergm-deprecated}
\alias{ergm-deprecated}
\alias{colMeans.mcmc.list}
\alias{sweep.mcmc.list}
\alias{get.miss.dyads}
\alias{InitErgmTerm.degreepopularity}
\alias{InitErgmTerm.idegreepopularity}
\alias{InitErgmTerm.odegreepopularity}
\alias{ergm.Cprepare.el}
\alias{ergm.checkargs}
\alias{ergm.checkbipartite}
\alias{ergm.checkdirected}
\alias{ergm.Cprepare.miss}
\alias{ergm.getMCMCsample}
\alias{ergm.mcmcslave}
\alias{ergm.getterms}
\alias{ergm.getmodel}
\alias{ergm.MHP.table}
\alias{MHproposal}
\alias{MHproposal.character}
\alias{MHproposal.ergm}
\alias{MHproposal.formula}
\alias{ergm.init.methods}
\alias{ergm.ConstraintImplications}
\alias{ergm.update.formula}
\alias{remove.offset.formula}
\alias{offset.info.formula}
\alias{print.gofobject}
\alias{summary.gof}
\alias{summary.gofobject}
\alias{plot.gofobject}
\alias{plot.mcmc.list.ergm}
\alias{coef.length.model}
\alias{plot.ergm}
\alias{plot.network.ergm}
\alias{ergm.getglobalstats}
\alias{summary.statistics}
\alias{summary.statistics.formula}
\alias{summary.statistics.network}
\title{Functions that will no longer be supported in future releases of the package}
\usage{
colMeans.mcmc.list(...)

sweep.mcmc.list(...)

get.miss.dyads(constraints, constraints.obs)

InitErgmTerm.degreepopularity(nw, arglist, ...)

InitErgmTerm.idegreepopularity(nw, arglist, ...)

InitErgmTerm.odegreepopularity(nw, arglist, ...)

ergm.Cprepare.el(x, attrname = NULL, prototype = NULL)

ergm.checkargs(fname, arglist, varnames = NULL, vartypes = NULL,
  defaultvalues = list(), required = NULL)

ergm.checkbipartite(fname, nw.bipartiteflag, requirement,
  extramessage = "")

ergm.checkdirected(fname, nw.directedflag, requirement,
  extramessage = "")

ergm.Cprepare.miss(nw)

ergm.getMCMCsample(nw, model, proposal, eta0, control, verbose = FALSE,
  response = NULL, update.nws = TRUE, ...)

ergm.mcmcslave(Clist, proposal, eta0, control, verbose, ...,
  prev.run = NULL, burnin = NULL, samplesize = NULL,
  interval = NULL, maxedges = NULL)

ergm.getterms(formula)

ergm.getmodel(object, ...)

ergm.MHP.table(...)

MHproposal(...)

MHproposal.character(...)

MHproposal.ergm(...)

MHproposal.formula(...)

ergm.init.methods(...)

ergm.ConstraintImplications(...)

ergm.update.formula(object, new, ..., from.new = FALSE)

remove.offset.formula(object, response = NULL)

offset.info.formula(object, response = NULL, ...)

\method{print}{gofobject}(x, ...)

\method{summary}{gof}(object, ...)

\method{summary}{gofobject}(object, ...)

\method{plot}{gofobject}(x, ...)

\method{plot}{mcmc.list.ergm}(...)

\method{coef}{length.model}(...)

\method{plot}{ergm}(x, ...)

\method{plot}{network.ergm}(x, attrname = NULL,
  label = network.vertex.names(x), coord = NULL, jitter = TRUE,
  thresh = 0, usearrows = TRUE, mode = "fruchtermanreingold",
  displayisolates = TRUE, interactive = FALSE, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, pad = 0.2,
  label.pad = 0.5, displaylabels = FALSE, boxed.labels = TRUE,
  label.pos = 0, label.bg = "white", vertex.sides = 8,
  vertex.rot = 0, arrowhead.cex = 1, label.cex = 1, loop.cex = 1,
  vertex.cex = 1, edge.col = 1, label.col = 1, vertex.col = 2,
  label.border = 1, vertex.border = 1, edge.lty = 1,
  label.lty = NULL, vertex.lty = 1, edge.lwd = 0,
  label.lwd = par("lwd"), edge.len = 0.5, edge.curve = 0.1,
  edge.steps = 50, loop.steps = 20, object.scale = 0.01,
  uselen = FALSE, usecurve = FALSE, suppress.axes = TRUE,
  vertices.last = TRUE, new = TRUE, layout.par = NULL,
  cex.main = par("cex.main"), cex.sub = par("cex.sub"), seed = NULL,
  latent.control = list(maxit = 500, trace = 0, dyadsample = 10000,
  penalty.sigma = c(5, 0.5), nsubsample = 200), colornames = "rainbow",
  verbose = FALSE, latent = FALSE, ...)

ergm.getglobalstats(nw, m, response = NULL)

summary.statistics(object, ...)

\method{summary}{statistics.formula}(object, ...)

\method{summary}{statistics.network}(object, ...)
}
\arguments{
\item{nw, arglist, ..., fname, varnames, vartypes, defaultvalues, required, nw.bipartiteflag, requirement, extramessage, nw.directedflag, m, response, formula, object, new, from.new, x, model, proposal, eta0, control, verbose, update.nws, Clist, prev.run, burnin, samplesize, interval, maxedges, attrname, label, coord, jitter, thresh, usearrows, mode, displayisolates, interactive, xlab, ylab, xlim, ylim, pad, label.pad, displaylabels, boxed.labels, label.pos, label.bg, vertex.sides, vertex.rot, arrowhead.cex, label.cex, loop.cex, vertex.cex, edge.col, label.col, vertex.col, label.border, vertex.border, edge.lty, label.lty, vertex.lty, edge.lwd, label.lwd, edge.len, edge.curve, edge.steps, loop.steps, object.scale, uselen, usecurve, suppress.axes, vertices.last, layout.par, cex.main, cex.sub, seed, latent.control, colornames, latent, constraints, constraints.obs, prototype}{Arguments to deprecated functions.}
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons

\code{ergm.Cprepare.el} has been replaced by \code{\link[=to_ergm_Cdouble.matrix]{to_ergm_Cdouble.matrix()}}

\code{ergm.Cprepare.miss} has been deprecated in favor of \code{\link{to_ergm_Cdouble}(is.na(nw))} or the \code{\link{rlebdm}}-based representations.

\code{\link[=print.gofobject]{print.gofobject()}} is a deprecated alias for \code{\link[=print.gof]{print.gof()}}.

\code{\link[=summary.gof]{summary.gof()}} is a deprecated alias for \code{\link[=print.gof]{print.gof()}}.

\code{\link[=summary.gofobject]{summary.gofobject()}} is a deprecated alias for \code{\link[=print.gof]{print.gof()}}.

\code{\link[=plot.gofobject]{plot.gofobject()}} is a deprecated alias for \code{\link[=plot.gof]{plot.gof()}}.

\code{plot.mcmc.list.ergm} is the obsolete name for \code{\link[=ergm_plot.mcmc.list]{ergm_plot.mcmc.list()}}.

\code{\link[=coef.length.model]{coef.length.model()}} has been replaced by the generic \code{\link[=nparam]{nparam()}}.

\code{plot.ergm}: deprecated alias for \code{\link[=mcmc.diagnostics]{mcmc.diagnostics()}}.

Use \code{\link[latentnet:plot.ergmm]{latentnet::plot.ergmm()}} instead.

\code{\link[=summary.statistics]{summary.statistics()}} is a deprecated name of \code{\link[=summary_formula]{summary_formula()}}.
}
\section{Functions}{
\itemize{
\item \code{InitErgmTerm.degreepopularity}: Use \code{\link{degree1.5}} instead.

\item \code{InitErgmTerm.idegreepopularity}: Use \code{\link{idegree1.5}} instead.

\item \code{InitErgmTerm.odegreepopularity}: Use \code{\link{odegree1.5}} instead.

\item \code{ergm.checkargs}: Use \code{\link[=check.ErgmTerm]{check.ErgmTerm()}} instead.

\item \code{ergm.checkbipartite}: Use \code{\link[=check.ErgmTerm]{check.ErgmTerm()}} instead.

\item \code{ergm.checkdirected}: Use \code{\link[=check.ErgmTerm]{check.ErgmTerm()}} instead.

\item \code{ergm.getMCMCsample}: Use \code{\link[=ergm_MCMC_sample]{ergm_MCMC_sample()}} instead.

\item \code{ergm.mcmcslave}: Use \code{\link[=ergm_MCMC_slave]{ergm_MCMC_slave()}} instead.

\item \code{ergm.getterms}: Use \code{\link[statnet.common:list_rhs.formula]{statnet.common::list_rhs.formula()}} and \code{\link[statnet.common:eval_lhs.formula]{statnet.common::eval_lhs.formula()}} instead.

\item \code{ergm.getmodel}: Use \code{ergm_model} instead.

\item \code{ergm.MHP.table}: Deprecated name for \code{\link[=ergm_proposal_table]{ergm_proposal_table()}}.

\item \code{MHproposal}: Deprecated name of \code{\link[=ergm_proposal]{ergm_proposal()}}.

\item \code{MHproposal.character}: Deprecated name of \code{\link[=ergm_proposal]{ergm_proposal()}}.

\item \code{MHproposal.ergm}: Deprecated name of \code{\link[=ergm_proposal]{ergm_proposal()}}.

\item \code{MHproposal.formula}: Deprecated name of \code{\link[=ergm_proposal]{ergm_proposal()}}.

\item \code{ergm.init.methods}: Deprecated: specify in the \code{InitErgmReference.*} implementation.

\item \code{ergm.ConstraintImplications}: Deprecated: specify in the \code{InitErgmConstraint.*} implementation.

\item \code{ergm.update.formula}: Use \code{\link[statnet.common]{nonsimp_update.formula}} instead.

\item \code{remove.offset.formula}: Use \code{\link[statnet.common:filter_rhs.formula]{statnet.common::filter_rhs.formula()}} such as \code{statnet.common::filter_rhs.formula(object, function(x) (if(is.call(x)) x[[1]] else x)!="offset")} instead.

\item \code{offset.info.formula}: \code{offset.info.formula} returns the offset
vectors associated with a formula.

\item \code{ergm.getglobalstats}: Use \code{\link[=summary.ergm_model]{summary.ergm_model()}} instead.
}}

\keyword{misc}
