% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-deprecated.R, R/InitErgmTerm.R,
%   R/formula.utils.R
\name{ergm-deprecated}
\alias{ergm-deprecated}
\alias{coef.ergm}
\alias{$.ergm}
\alias{control.ergm.godfather}
\alias{InitErgmTerm.degreepopularity}
\alias{InitErgmTerm.hammingmix}
\alias{hammingmix}
\alias{InitErgmTerm.idegreepopularity}
\alias{InitErgmTerm.odegreepopularity}
\alias{offset.info.formula}
\title{Functions that will no longer be supported in future releases of the package}
\usage{
\method{coef}{ergm}(object, ...)

\method{$}{ergm}(x, name)

control.ergm.godfather(term.options = NULL)

InitErgmTerm.degreepopularity(nw, arglist, ...)

InitErgmTerm.hammingmix(nw, arglist, ..., version = packageVersion("ergm"))

InitErgmTerm.idegreepopularity(nw, arglist, ...)

InitErgmTerm.odegreepopularity(nw, arglist, ...)

offset.info.formula(object, ...)
}
\arguments{
\item{name}{See \link{Extract}.}

\item{nw, arglist, ..., formula, object, x}{Arguments to deprecated functions.}
}
\description{
Functions that have been superceed, were never documented, or will be removed from the package for other reasons
}
\section{Functions}{
\itemize{
\item \code{coef(ergm)}: extracts the \code{ergm} parameters; may be
removed in favour of the default method once the number of \code{ergm}
objects with \verb{$coef} elements in the wild is sufficiently low.

\item \code{$}: accesses elements of \code{ergm} objects;
needed for backwards compatibility when components get renamed.

\item \code{control.ergm.godfather()}: constructs a control list for \code{\link[=ergm.godfather]{ergm.godfather()}}; not used at this time.

\item \code{InitErgmTerm.degreepopularity()}: Use \code{\link[=degree1.5-ergmTerm]{degree1.5}} instead.

\item \code{InitErgmTerm.idegreepopularity()}: Use \code{\link[=idegree1.5-ergmTerm]{idegree1.5}} instead.

\item \code{InitErgmTerm.odegreepopularity()}: Use \code{\link[=odegree1.5-ergmTerm]{odegree1.5}} instead.

\item \code{offset.info.formula()}: \code{offset.info.formula} returns the offset
vectors associated with a formula.

}}
\keyword{internal}
\keyword{misc}
