\name{ergm-internal}
\title{Internal ergm Objects}

\alias{ConstraintImplications}

\alias{InitConstraint.bd}
\alias{InitConstraint.degreedist}
\alias{InitConstraint.degrees}
\alias{InitConstraint.edges}
\alias{InitConstraint.hamming}
\alias{InitConstraint.indegreedist}
\alias{InitConstraint.nodedegrees}
\alias{InitConstraint.observed}
\alias{InitConstraint.outdegreedist}
\alias{InitConstraint.ranks}

\alias{InitErgm.altistar}
\alias{InitErgm.altostar}
\alias{InitErgm.b1kappa}
\alias{InitErgm.b1share}
\alias{InitErgm.b2kappa}
\alias{InitErgm.b2share}
\alias{InitErgm.bichange}
\alias{InitErgm.biduration}
\alias{InitErgm.biendure}
\alias{InitErgm.bounded.degree}
\alias{InitErgm.bounded.idegree}
\alias{InitErgm.bounded.istar}
\alias{InitErgm.bounded.kstar}
\alias{InitErgm.bounded.odegree}
\alias{InitErgm.bounded.ostar}
\alias{InitErgm.bounded.triangle}
\alias{InitErgm.degreep}
\alias{InitErgm.factor}
\alias{InitErgm.geodegree}
\alias{InitErgm.gwb1706}
\alias{InitErgm.gwb1share}
\alias{InitErgm.gwb2706}
\alias{InitErgm.gwb2share}
\alias{InitErgm.gwdegree706}
\alias{InitErgm.gwdegreealpha}
\alias{InitErgm.gwdegreelambda}
\alias{InitErgm.hamming.weighted}
\alias{InitErgm.hammingmix.constant}
\alias{InitErgm.heideriandynamic}
\alias{InitErgm.hiertriad}
\alias{InitErgm.hiertriaddegree}
\alias{InitErgm.icvar}
\alias{InitErgm.idc}
\alias{InitErgm.intransitivedynamic}
\alias{InitErgm.intransitivity}
\alias{InitErgm.kappa}
\alias{InitErgm.monopolymixmat}
\alias{InitErgm.simmeliandynamic}
\alias{InitErgm.spatial}
\alias{InitErgm.transitivedynamic}
\alias{InitErgm.transitivity}

\alias{InitErgmTerm.adegcor}
\alias{InitErgmTerm.b1degree.edgecov}
\alias{InitErgmTerm.b1mindegree}
\alias{InitErgmTerm.b1mindegree.edgecov}
\alias{InitErgmTerm.b2degree.edgecov}
\alias{InitErgmTerm.b2mindegree}
\alias{InitErgmTerm.b2mindegree.edgecov}
\alias{InitErgmTerm.coincidence}
\alias{InitErgmTerm.concurrentties2}
\alias{InitErgmTerm.conddegmple}
\alias{InitErgmTerm.ctriad}
\alias{InitErgmTerm.cyclicalties}
\alias{InitErgmTerm.cyclicalties2}
\alias{InitErgmTerm.homoproportion}
\alias{InitErgmTerm.indegreepopularity}
\alias{InitErgmTerm.match}
\alias{InitErgmTerm.nodemain}
\alias{InitErgmTerm.outdegreepopularity}
\alias{InitErgmTerm.pdegcor}
\alias{InitErgmTerm.rdegcor}
\alias{InitErgmTerm.transitiveties}
\alias{InitErgmTerm.transitiveties2}
\alias{InitErgmTerm.triangles}
\alias{InitErgmTerm.ttriad}

\alias{InitMHP.CondDegree}
\alias{InitMHP.CondDegreeDist}
\alias{InitMHP.CondInDegreeDist}
\alias{InitMHP.CondOutDegreeDist}
\alias{InitMHP.ConstantEdges}
\alias{InitMHP.HammingConstantEdges}
\alias{InitMHP.HammingTNT}
\alias{InitMHP.TNT}
\alias{InitMHP.TNT10}
\alias{InitMHP.nobetweengroupties}
\alias{InitMHP.randomtoggle}
\alias{InitMHP.randomtoggleNonObserved}

\alias{InitWtErgmTerm.CMP}
\alias{InitWtErgmTerm.absdiff}
\alias{InitWtErgmTerm.absdiffcat}
\alias{InitWtErgmTerm.atleast}
\alias{InitWtErgmTerm.cyclicalties}
\alias{InitWtErgmTerm.cyclicalweights}
\alias{InitWtErgmTerm.deference}
\alias{InitWtErgmTerm.edgecov}
\alias{InitWtErgmTerm.greaterthan}
\alias{InitWtErgmTerm.inconsistency}
\alias{InitWtErgmTerm.ininterval}
\alias{InitWtErgmTerm.mutual}
\alias{InitWtErgmTerm.nodecov}
\alias{InitWtErgmTerm.nodecovar}
\alias{InitWtErgmTerm.nodefactor}
\alias{InitWtErgmTerm.nodeicov}
\alias{InitWtErgmTerm.nodeicovar}
\alias{InitWtErgmTerm.nodeifactor}
\alias{InitWtErgmTerm.nodeisqrtcovar}
\alias{InitWtErgmTerm.nodemain}
\alias{InitWtErgmTerm.nodeocov}
\alias{InitWtErgmTerm.nodeocovar}
\alias{InitWtErgmTerm.nodeofactor}
\alias{InitWtErgmTerm.nodeosqrtcovar}
\alias{InitWtErgmTerm.nodesqrtcovar}
\alias{InitWtErgmTerm.nonconformity}
\alias{InitWtErgmTerm.nonzero}
\alias{InitWtErgmTerm.sum}
\alias{InitWtErgmTerm.transitiveties}
\alias{InitWtErgmTerm.transitiveweights}

\alias{InitWtMHP.DescRank}
\alias{InitWtMHP.DescRankEquivalent}
\alias{InitWtMHP.Poisson}
\alias{InitWtMHP.PoissonNonObserved}
\alias{InitWtMHP.StdNormal}
\alias{InitWtMHP.StdNormalRank}
\alias{InitWtMHP.ZIPoisson}

\alias{MHproposal}
\alias{MHproposal.MHproposal}
\alias{MHproposal.NULL}
\alias{MHproposal.character}
\alias{MHproposal.ergm}
\alias{MHproposal.formula}

\alias{approx.hotelling.diff.test}
\alias{append.rhs.formula}
\alias{as.edgelist}
\alias{as.edgelist.compressed}
\alias{as.network.uncompressed}
\alias{autoboot.ergm}
\alias{autoboot.se.ergm}
\alias{central.network}
\alias{check.ErgmTerm}
\alias{coef.length.model}
\alias{coef.sublength.model}
\alias{colMeans.mcmc.list}
\alias{compress.data.frame}
\alias{control.ergm.toplevel}
\alias{control.simulate.ergm.toplevel}
\alias{copy.named}
\alias{degreedist}
\alias{degreedistfactor}
\alias{delete.isolates}
\alias{drawcircle}
\alias{drawpie}
\alias{dspartnerdist}
\alias{ergm.Cprepare}
\alias{ergm.Cprepare.el}
\alias{ergm.Cprepare.miss}
\alias{ergm.MCMCse}
\alias{ergm.MCMCse.lognormal}
\alias{ergm.MCMCse.old}
\alias{ergm.MCMLE}
\alias{ergm.MPIcluster.started}
\alias{ergm.bounddeg}
\alias{ergm.bridge.0.llk}
\alias{ergm.bridge.hammingstart.llk}
\alias{ergm.bridge.preproc}
\alias{ergm.checkargs}
\alias{ergm.checkbipartite}
\alias{ergm.checkconstraints.model}
\alias{ergm.checkdegeneracy}
\alias{ergm.checkdirected}
\alias{ergm.checkextreme.model}
\alias{ergm.coefficient.degeneracy}
\alias{ergm.compute.degeneracy}
\alias{ergm.curved.statsmatrix}
\alias{ergm.degeneracy}
\alias{ergm.design}
\alias{ergm.el.lasttoggle}
\alias{ergm.estimate}
\alias{ergm.eta}
\alias{ergm.etagrad}
\alias{ergm.etagradmult}
\alias{ergm.etamap}
\alias{ergm.geodesicmatrix}
\alias{ergm.geodesicmatrix.edgelist}
\alias{ergm.geodistdist}
\alias{ergm.geodistn}
\alias{ergm.getCluster}
\alias{ergm.getMCMCsample}
\alias{ergm.getglobalstats}
\alias{ergm.getmodel}
\alias{ergm.getnetwork}
\alias{ergm.getterms}
\alias{ergm.independencemodel}
\alias{ergm.initialfit}
\alias{ergm.logisticdeviance}
\alias{ergm.logitreg}
\alias{ergm.mahalanobis}
\alias{ergm.mapl}
\alias{ergm.maple}
\alias{ergm.mcmcslave}
\alias{ergm.mple}
\alias{ergm.nodegeodesics}
\alias{ergm.pairgeodesic}
\alias{ergm.pen.glm}
\alias{ergm.phase12}
\alias{ergm.pl}
\alias{ergm.plinfo}
\alias{ergm.reviseinit}
\alias{ergm.rhs.formula}
\alias{ergm.robmon}
\alias{ergm.sample.eta2theta}
\alias{ergm.sample.tomcmc}
\alias{ergm.stepping}
\alias{ergm.stocapprox}
\alias{ergm.stopCluster}
\alias{ergm.stopCluster.MPIcluster}
\alias{ergm.stopCluster.default}
\alias{ergm.sufftoprob}
\alias{ergm.t.summary}
\alias{ergm.update.formula}
\alias{espartnerdist}
\alias{get.InitErgm.fname}
\alias{get.node.attr}
\alias{init.methods}
\alias{is.ergm}
\alias{is.invertible}
\alias{largest.components}
\alias{license.statnet}
\alias{llik.fun}
\alias{llik.fun.EF}
\alias{llik.fun.median}
\alias{llik.fun.obs}
\alias{llik.fun.obs.robust}
\alias{llik.fun2}
\alias{llik.fun3}
\alias{llik.grad}
\alias{llik.grad.obs}
\alias{llik.grad2}
\alias{llik.grad3}
\alias{llik.hessian}
\alias{llik.hessian.naive}
\alias{llik.hessian.obs}
\alias{llik.hessian2}
\alias{llik.info3}
\alias{llik.mcmcvar3}
\alias{logistftest}
\alias{logp.rank.ergm}
\alias{mcmc.diagnostics.default}
\alias{midarrow}
\alias{model.matrix.pen.glm}
\alias{model.transform.formula}
\alias{mvmodel}
\alias{mvmodel.default}
\alias{mvmodel.ergm}
\alias{mvmodel.formula}
\alias{myLibLoc}
\alias{newnw.extract}
\alias{nologLik.message}
\alias{nvattr.copy.network}
\alias{ostar2deg}
\alias{paste.and}
\alias{plot.mcmc.list.ergm}
\alias{robust.inverse}
\alias{rspartnerdist}
\alias{set.offset.formula}
\alias{sociality.default}
\alias{sociality.ergm}
\alias{sociality.formula}
\alias{sociality.network}
\alias{sort.data.frame}
\alias{statnet.edit}
\alias{statnetShowDoc}
\alias{statnetbrowseURL}
\alias{summary.statsmatrix.ergm}
\alias{sweep.mcmc.list}
\alias{term.list.formula}
\alias{twopathdist}
\alias{updatemodel.ErgmTerm}
\alias{vector.namesmatch}
\alias{xyplot.mcmc.list.ergm}

\alias{print.control.list}
\alias{set.control.class}
\alias{check.control.class}

\description{
Internal ergm functions.}

\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).}

\seealso{ergm, statnet-package}
\keyword{internal}
