\name{ergm-internal}
\title{Internal ergm Objects}
% terms
\alias{ergm.checkargs}
\alias{ergm.checkdirected}
\alias{ergm.checkdegeneracy}
\alias{ergm.checkbipartite}
%
\alias{ConstraintImplications}
\alias{InitConstraint.bd}
\alias{InitConstraint.degreedist}
\alias{InitConstraint.degreesHexad}
\alias{InitConstraint.degreesTetrad}
\alias{InitConstraint.degrees}
\alias{InitConstraint.edges}
\alias{InitConstraint.hamming}
\alias{InitConstraint.indegreedist}
\alias{InitConstraint.outdegreedist}
\alias{InitConstraint.nodedegrees}
\alias{InitConstraint.observed}
%

\alias{InitErgmTerm.absdiff}
\alias{InitErgmTerm.absdiffcat}
\alias{InitErgmTerm.adegcor}
\alias{InitErgmTerm.altkstar}
\alias{InitErgmTerm.asymmetric}
\alias{InitErgmTerm.b1concurrent}
\alias{InitErgmTerm.b1degree}   
\alias{InitErgmTerm.b1factor}
\alias{InitErgmTerm.b1star}
\alias{InitErgmTerm.b1starmix}
\alias{InitErgmTerm.b1twostar}
\alias{InitErgmTerm.b2concurrent}
\alias{InitErgmTerm.b2degree}
\alias{InitErgmTerm.b2factor}
\alias{InitErgmTerm.b2star}
\alias{InitErgmTerm.b2starmix}
\alias{InitErgmTerm.b2twostar}
\alias{InitErgmTerm.balance}
\alias{InitErgmTerm.concurrent}
\alias{InitErgmTerm.ctriad}
\alias{InitErgmTerm.ctriple}
\alias{InitErgmTerm.cycle}
\alias{InitErgmTerm.degcor}
\alias{InitErgmTerm.degcrossprod}
\alias{InitErgmTerm.degree}
\alias{InitErgmTerm.density}
\alias{InitErgmTerm.dsp}
\alias{InitErgmTerm.dyadcov}
\alias{InitErgmTerm.edgecov}
\alias{InitErgmTerm.edges}
\alias{InitErgmTerm.esp}
\alias{InitErgmTerm.gwb1degree}
\alias{InitErgmTerm.gwb2degree}
\alias{InitErgmTerm.gwdegree}
\alias{InitErgmTerm.gwdsp}
\alias{InitErgmTerm.gwesp}
\alias{InitErgmTerm.gwidegree}
\alias{InitErgmTerm.gwnsp}
\alias{InitErgmTerm.gwodegree}
\alias{InitErgmTerm.hamming}
\alias{InitErgmTerm.hammingmix}
\alias{InitErgmTerm.idegree}
\alias{InitErgmTerm.indegreepopularity}
\alias{InitErgmTerm.intransitive}
\alias{InitErgmTerm.isolates}
\alias{InitErgmTerm.istar}
\alias{InitErgmTerm.kstar}
\alias{InitErgmTerm.localtriangle}
\alias{InitErgmTerm.m2star}
\alias{InitErgmTerm.match}
\alias{InitErgmTerm.meandeg}
\alias{InitErgmTerm.mutual}
\alias{InitErgmTerm.nearsimmelian}
\alias{InitErgmTerm.nodecov}
\alias{InitErgmTerm.nodefactor}
\alias{InitErgmTerm.nodeicov}
\alias{InitErgmTerm.nodeifactor}
\alias{InitErgmTerm.nodemain}
\alias{InitErgmTerm.nodematch}
\alias{InitErgmTerm.nodemix}
\alias{InitErgmTerm.nodeocov}
\alias{InitErgmTerm.nodeofactor}
\alias{InitErgmTerm.nsp}
\alias{InitErgmTerm.odegree}
\alias{InitErgmTerm.ostar}
\alias{InitErgmTerm.outdegreepopularity}
\alias{InitErgmTerm.rdegcor}
\alias{InitErgmTerm.receiver}
\alias{InitErgmTerm.sender}
\alias{InitErgmTerm.simmelian}
\alias{InitErgmTerm.simmelianties}
\alias{InitErgmTerm.smalldiff}
\alias{InitErgmTerm.sociality}
\alias{InitErgmTerm.threepath}
\alias{InitErgmTerm.transitive}
\alias{InitErgmTerm.transitiveties}
\alias{InitErgmTerm.triadcensus}
\alias{InitErgmTerm.triangle}
\alias{InitErgmTerm.triangles}
\alias{InitErgmTerm.tripercent}
\alias{InitErgmTerm.ttriad}
\alias{InitErgmTerm.ttriple}
\alias{InitErgmTerm.twopath}

%
\alias{check.ErgmTerm}
\alias{assignvariables}
\alias{check.ErgmTerm.summarystats}
\alias{ergm.getglobalstats}
\alias{get.InitErgm.fname}
\alias{updatemodel.ErgmTerm}
\alias{copy.named}
\alias{ergm.update.formula}
\alias{term.list.formula}
\alias{model.matrix.pen.glm}
\alias{zerowarnings}
\alias{extremewarnings}

%
\alias{InitMHP.CondDegree}
\alias{InitMHP.CondDegreeDist}
\alias{InitMHP.CondInDegreeDist}
\alias{InitMHP.CondOutDegreeDist}
\alias{InitMHP.ConstantEdges}
\alias{InitMHP.dissolution}
\alias{InitMHP.formation}
\alias{InitMHP.formationTNT}
\alias{InitMHP.HammingConstantEdges}
\alias{InitMHP.HammingTNT}
\alias{InitMHP.nobetweengroupties}
\alias{InitMHP.randomtoggle}
\alias{InitMHP.randomtoggleNonObserved}
\alias{InitMHP.TNT}
\alias{InitMHP.TNT10}
\alias{InitMHP.CondDegreeHexad}
\alias{InitMHP.CondDegreeTetrad}
\alias{MHproposal}
\alias{MHproposal.character}
\alias{MHproposal.ergm}
\alias{MHproposal.formula}
\alias{MHproposal.MHproposal}
\alias{MHproposal.NULL}
\alias{MHproposals}
%
\alias{degreedistfactor}
\alias{drawcircle}
\alias{drawpie}
\alias{dspartnerdist}
\alias{ergm.compute.degeneracy}
\alias{ergm.curved}
\alias{ergm.curved.statsmatrix}
\alias{ergm.curved.update}
\alias{ergm.degenerate}
\alias{ergm.design}
\alias{ergm.estimate}
\alias{ergm.estimate.only}
\alias{ergm.eta}
\alias{ergm.etagrad}
\alias{ergm.etagradmult}
\alias{ergm.etamap}
\alias{ergm.geodesicmatrix}
\alias{ergm.geodesicmatrix.edgelist}
\alias{ergm.geodesicmatrix.old}
\alias{ergm.geodist}
\alias{ergm.geodistdist}
\alias{ergm.geodistn}
\alias{ergm.getMCMCsample}
\alias{ergm.mcmcslave}
\alias{ergm.getnetwork}
\alias{ergm.independencemodel}
\alias{ergm.initialfit}
\alias{ergm.logisticdeviance}
\alias{ergm.logitreg}
\alias{ergm.mainfitloop}
\alias{ergm.marquardt}
\alias{ergm.marquardt2}
\alias{ergm.pl}
\alias{ergm.nodegeodesics}
\alias{ergm.pairgeodesic}
\alias{ergm.phase12}
\alias{ergm.raftery.diag}
\alias{ergm.revisetheta0}
\alias{ergm.rhs.formula} 
\alias{ergm.stepping}
\alias{ergm.stocapprox}
\alias{ergm.t.summary}
\alias{espartnerdist}
\alias{get.node.attr}
\alias{is.ergm}
\alias{is.matrixnetwork}
\alias{llik.fun}
\alias{llik.fun.miss}
\alias{llik.fun2}
\alias{llik.fun3}
\alias{llik.grad}
\alias{llik.grad.miss}
\alias{llik.grad2}
\alias{llik.grad3}
\alias{llik.hessian}
\alias{llik.hessian.miss} 
\alias{llik.hessian2}
\alias{llik.info3}
\alias{llik.mcmcvar3}
\alias{llik.fun.EF}
\alias{llik.fun.median}
\alias{llik.fun.miss.robust}
\alias{llik.hessian.naive}
\alias{mcmc.diagnostics.default}
\alias{midarrow}
\alias{newnw.extract}
\alias{plot.mcmc.ergm}
\alias{print.raftery.diag.ergm}
\alias{robust.inverse}
\alias{rspartnerdist}
\alias{set.mfrow}
\alias{traceplot.ergm}
\alias{twopathdist}
%
\alias{ergm.sufftoprob}
\alias{ergm.Cprepare}
\alias{ergm.getmodel}
\alias{ergm.getterms}
\alias{ergm.gettermnames}
\alias{ergm.mple}
\alias{ergm.statseval}
\alias{fulldistdist}
\alias{fullgcount}
\alias{statnet.edit}
%\alias{get.node.attr}
%\alias{geodist}
%\alias{has.loops}
%\alias{is.directed}
%\alias{is.hyper}
%\alias{lpz.dist}
%\alias{mahal}
%\alias{mlpY}
%\alias{mlpY.grad}
%\alias{mlpY0}
%\alias{mlpY0.grad}
%\alias{mlpYmdsZ}
%\alias{mlpYmdsZ.grad}
%\alias{mlpYnull}
%\alias{prepmahal}
%\alias{symmetrize}
%
\alias{ergm.MCMCse}
\alias{ergm.MCMCse.lognormal}
%\alias{ergm.latentplot}
%\alias{mlpY.grad.plot}
%\alias{mlpY.plot}
%\alias{mlpYmdsZ.grad.plot}
%\alias{mlpYmdsZ.plot}
\alias{ergm.MCMCacf}
\alias{ergm.pen.glm}
\alias{logistftest}
%
\alias{ergm.mainloop}
\alias{ergm.robmon}
%

\alias{as.edgelist.compressed}
\alias{as.network.uncompressed}

%
% These need to be documented!
%

\alias{as.directed}
\alias{as.edgelist}
\alias{ergm.degeneracy} % At least this needs to be documented in summary.ergm and/or mcmc.diagnostics.ergm
\alias{print.network.series}
\alias{summary.statsmatrix.ergm}
\alias{summary.network.series}
\alias{license.statnet}
\alias{license.statnet1}
\alias{statnetShowDoc}
\alias{statnetbrowseURL}
\alias{ergm.bounddeg}
\alias{is.invertible}
\alias{ostar2deg}
\alias{sufftoprob}
\alias{degreedist}
\alias{mixingmatrix}
\alias{print.mixingmatrix}
\alias{print.summary.ergm} 
\alias{Summary.ergm.future}

\description{
  Internal ergm functions.
}
%\usage{
%anova.ergm(object, \dots)
%}
\details{
  Most of these are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
%\arguments{
% \item{\dots}{further arguments passed to or used by methods.}
%}
\seealso{ergm, statnet-package}
\keyword{internal}
