%  File ergm/man/ergm-internal.Rd
%  Part of the statnet package, http://statnet.org
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) in
%    http://statnet.org/attribution
%
%  Copyright 2012 the statnet development team
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{ergm-internal}
\title{Internal ergm Objects}

\alias{ConstraintImplications}

\alias{InitConstraint.atleast}
\alias{InitConstraint.atmost}
\alias{InitConstraint.bd}
\alias{InitConstraint.degreedist}
\alias{InitConstraint.degrees}
\alias{InitConstraint.degreesHexad}
\alias{InitConstraint.degreesTetrad}
\alias{InitConstraint.degreessimple}
\alias{InitConstraint.edges}
\alias{InitConstraint.hamming}
\alias{InitConstraint.indegreedist}
\alias{InitConstraint.nodedegrees}
\alias{InitConstraint.observed}
\alias{InitConstraint.outdegreedist}
\alias{InitConstraint.ranks}

\alias{InitErgm.D.dyad}
\alias{InitErgm.D.edge}
\alias{InitErgm.D.off}
\alias{InitErgm.altistar}
\alias{InitErgm.altostar}
\alias{InitErgm.b1kappa}
\alias{InitErgm.b1share}
\alias{InitErgm.b2kappa}
\alias{InitErgm.b2share}
\alias{InitErgm.bichange}
\alias{InitErgm.biduration}
\alias{InitErgm.biendure}
\alias{InitErgm.bounded.degree}
\alias{InitErgm.bounded.idegree}
\alias{InitErgm.bounded.istar}
\alias{InitErgm.bounded.kstar}
\alias{InitErgm.bounded.odegree}
\alias{InitErgm.bounded.ostar}
\alias{InitErgm.bounded.triangle}
\alias{InitErgm.degreep}
\alias{InitErgm.dissolve}
\alias{InitErgm.duration}
\alias{InitErgm.factor}
\alias{InitErgm.formation}
\alias{InitErgm.geodegree}
\alias{InitErgm.gwb1706}
\alias{InitErgm.gwb1share}
\alias{InitErgm.gwb2706}
\alias{InitErgm.gwb2share}
\alias{InitErgm.gwdegree706}
\alias{InitErgm.gwdegreealpha}
\alias{InitErgm.gwdegreelambda}
\alias{InitErgm.hamming.weighted}
\alias{InitErgm.hammingmix.constant}
\alias{InitErgm.heideriandynamic}
\alias{InitErgm.hiertriad}
\alias{InitErgm.hiertriaddegree}
\alias{InitErgm.icvar}
\alias{InitErgm.idc}
\alias{InitErgm.intransitivedynamic}
\alias{InitErgm.intransitivity}
\alias{InitErgm.kappa}
\alias{InitErgm.monopolymixmat}
\alias{InitErgm.simmeliandynamic}
\alias{InitErgm.spatial}
\alias{InitErgm.transitivedynamic}
\alias{InitErgm.transitivity}

\alias{InitErgmTerm.absdiff}
\alias{InitErgmTerm.absdiffcat}
\alias{InitErgmTerm.adegcor}
\alias{InitErgmTerm.altkstar}
\alias{InitErgmTerm.asymmetric}
\alias{InitErgmTerm.b1concurrent}
\alias{InitErgmTerm.b1degree}
\alias{InitErgmTerm.b1degree.edgecov}
\alias{InitErgmTerm.b1factor}
\alias{InitErgmTerm.b1mindegree}
\alias{InitErgmTerm.b1mindegree.edgecov}
\alias{InitErgmTerm.b1star}
\alias{InitErgmTerm.b1starmix}
\alias{InitErgmTerm.b1twostar}
\alias{InitErgmTerm.b2concurrent}
\alias{InitErgmTerm.b2degree}
\alias{InitErgmTerm.b2degree.edgecov}
\alias{InitErgmTerm.b2factor}
\alias{InitErgmTerm.b2mindegree}
\alias{InitErgmTerm.b2mindegree.edgecov}
\alias{InitErgmTerm.b2star}
\alias{InitErgmTerm.b2starmix}
\alias{InitErgmTerm.b2twostar}
\alias{InitErgmTerm.balance}
\alias{InitErgmTerm.coincidence}
\alias{InitErgmTerm.concurrent}
\alias{InitErgmTerm.concurrentties}
\alias{InitErgmTerm.concurrentties2}
\alias{InitErgmTerm.conddegmple}
\alias{InitErgmTerm.ctriad}
\alias{InitErgmTerm.ctriple}
\alias{InitErgmTerm.cycle}
\alias{InitErgmTerm.cyclicalties}
\alias{InitErgmTerm.cyclicalties2}
\alias{InitErgmTerm.degcor}
\alias{InitErgmTerm.degcrossprod}
\alias{InitErgmTerm.degree}
\alias{InitErgmTerm.density}
\alias{InitErgmTerm.dsp}
\alias{InitErgmTerm.dyadcov}
\alias{InitErgmTerm.edge.ages}
\alias{InitErgmTerm.edgecov}
\alias{InitErgmTerm.edgecov.ages}
\alias{InitErgmTerm.edges}
\alias{InitErgmTerm.edges.ageinterval}
\alias{InitErgmTerm.esp}
\alias{InitErgmTerm.gwb1degree}
\alias{InitErgmTerm.gwb2degree}
\alias{InitErgmTerm.gwdegree}
\alias{InitErgmTerm.gwdsp}
\alias{InitErgmTerm.gwesp}
\alias{InitErgmTerm.gwidegree}
\alias{InitErgmTerm.gwnsp}
\alias{InitErgmTerm.gwodegree}
\alias{InitErgmTerm.hamming}
\alias{InitErgmTerm.hammingmix}
\alias{InitErgmTerm.homoproportion}
\alias{InitErgmTerm.idegree}
\alias{InitErgmTerm.indegreepopularity}
\alias{InitErgmTerm.intransitive}
\alias{InitErgmTerm.isolates}
\alias{InitErgmTerm.istar}
\alias{InitErgmTerm.kstar}
\alias{InitErgmTerm.localtriangle}
\alias{InitErgmTerm.m2star}
\alias{InitErgmTerm.match}
\alias{InitErgmTerm.meandeg}
\alias{InitErgmTerm.mutual}
\alias{InitErgmTerm.nearsimmelian}
\alias{InitErgmTerm.nodecov}
\alias{InitErgmTerm.nodefactor}
\alias{InitErgmTerm.nodeicov}
\alias{InitErgmTerm.nodeifactor}
\alias{InitErgmTerm.nodemain}
\alias{InitErgmTerm.nodematch}
\alias{InitErgmTerm.nodemix}
\alias{InitErgmTerm.nodeocov}
\alias{InitErgmTerm.nodeofactor}
\alias{InitErgmTerm.nsp}
\alias{InitErgmTerm.odegree}
\alias{InitErgmTerm.opentriad}
\alias{InitErgmTerm.ostar}
\alias{InitErgmTerm.outdegreepopularity}
\alias{InitErgmTerm.pdegcor}
\alias{InitErgmTerm.rdegcor}
\alias{InitErgmTerm.receiver}
\alias{InitErgmTerm.sender}
\alias{InitErgmTerm.simmelian}
\alias{InitErgmTerm.simmelianties}
\alias{InitErgmTerm.smalldiff}
\alias{InitErgmTerm.sociality}
\alias{InitErgmTerm.threepath}
\alias{InitErgmTerm.transitive}
\alias{InitErgmTerm.transitiveties}
\alias{InitErgmTerm.transitiveties2}
\alias{InitErgmTerm.triadcensus}
\alias{InitErgmTerm.triangle}
\alias{InitErgmTerm.triangles}
\alias{InitErgmTerm.tripercent}
\alias{InitErgmTerm.ttriad}
\alias{InitErgmTerm.ttriple}
\alias{InitErgmTerm.twopath}

\alias{InitMHP.CondDegree}
\alias{InitMHP.CondDegreeDist}
\alias{InitMHP.CondDegreeHexad}
\alias{InitMHP.CondDegreeSimple}
\alias{InitMHP.CondDegreeTetrad}
\alias{InitMHP.CondInDegreeDist}
\alias{InitMHP.CondOutDegreeDist}
\alias{InitMHP.ConstantEdges}
\alias{InitMHP.HammingConstantEdges}
\alias{InitMHP.HammingTNT}
\alias{InitMHP.TNT}
\alias{InitMHP.TNT10}
\alias{InitMHP.dissolution}
\alias{InitMHP.dissolutionMLE}
\alias{InitMHP.dissolutionNonObservedMLE}
\alias{InitMHP.formation}
\alias{InitMHP.formationMLE}
\alias{InitMHP.formationMLETNT}
\alias{InitMHP.formationNonObservedMLE}
\alias{InitMHP.formationTNT}
\alias{InitMHP.nobetweengroupties}
\alias{InitMHP.randomtoggle}
\alias{InitMHP.randomtoggleNonObserved}

\alias{MHproposal}
\alias{MHproposal.MHproposal}
\alias{MHproposal.NULL}
\alias{MHproposal.character}
\alias{MHproposal.ergm}
\alias{MHproposal.formula}
\alias{MHproposals}

\alias{approx.hotelling.diff.test}
\alias{append.rhs.formula}
\alias{as.edgelist}
\alias{as.edgelist.compressed}
\alias{as.network.uncompressed}
\alias{assignvariables}
\alias{central.network}
\alias{check.ErgmTerm}
\alias{coef.length.model}
\alias{coef.sublength.model}
\alias{colMeans.mcmc.list}
\alias{compress.data.frame}
\alias{control.ergm.toplevel}
\alias{control.godfather}
\alias{control.logLik.stergm}
\alias{control.simulate.ergm.toplevel}
\alias{copy.named}
\alias{degreedist}
\alias{degreedistfactor}
\alias{delete.isolates}
\alias{drawcircle}
\alias{drawpie}
\alias{dspartnerdist}
\alias{ergm.Cprepare}
\alias{ergm.Cprepare.el}
\alias{ergm.Cprepare.miss}
\alias{ergm.MCMCse}
\alias{ergm.MCMCse.lognormal}
\alias{ergm.MCMCse.old}
\alias{ergm.MCMLE}
\alias{ergm.MPIcluster.started}
\alias{ergm.bounddeg}
\alias{ergm.bridge.0.llk}
\alias{ergm.bridge.hammingstart.llk}
\alias{ergm.bridge.preproc}
\alias{ergm.checkargs}
\alias{ergm.checkbipartite}
\alias{ergm.checkconstraints.model}
\alias{ergm.checkdegeneracy}
\alias{ergm.checkdirected}
\alias{ergm.checkextreme.model}
\alias{ergm.coefficient.degeneracy}
\alias{ergm.compute.degeneracy}
\alias{ergm.curved.statsmatrix}
\alias{ergm.degeneracy}
\alias{ergm.design}
\alias{ergm.el.lasttoggle}
\alias{ergm.estimate}
\alias{ergm.eta}
\alias{ergm.etagrad}
\alias{ergm.etagradmult}
\alias{ergm.etamap}
\alias{ergm.geodesicmatrix}
\alias{ergm.geodesicmatrix.edgelist}
\alias{ergm.geodistdist}
\alias{ergm.geodistn}
\alias{ergm.getCluster}
\alias{ergm.getMCMCsample}
\alias{ergm.getglobalstats}
\alias{ergm.getmodel}
\alias{ergm.getnetwork}
\alias{ergm.getterms}
\alias{ergm.godfather}
\alias{ergm.independencemodel}
\alias{ergm.initialfit}
\alias{ergm.logisticdeviance}
\alias{ergm.logitreg}
\alias{ergm.mahalanobis}
\alias{ergm.mapl}
\alias{ergm.maple}
\alias{ergm.mcmcslave}
\alias{ergm.mple}
\alias{ergm.nodegeodesics}
\alias{ergm.pairgeodesic}
\alias{ergm.pen.glm}
\alias{ergm.phase12}
\alias{ergm.pl}
\alias{ergm.plinfo}
\alias{ergm.reviseinit}
\alias{ergm.rhs.formula}
\alias{ergm.robmon}
\alias{ergm.sample.eta2theta}
\alias{ergm.sample.tomcmc}
\alias{ergm.stepping}
\alias{ergm.stocapprox}
\alias{ergm.stopCluster}
\alias{ergm.stopCluster.MPIcluster}
\alias{ergm.stopCluster.default}
\alias{ergm.sufftoprob}
\alias{ergm.t.summary}
\alias{ergm.update.formula}
\alias{espartnerdist}
\alias{get.InitErgm.fname}
\alias{get.node.attr}
\alias{init.methods}
\alias{is.ergm}
\alias{is.invertible}
\alias{largest.components}
\alias{license.statnet}
\alias{llik.fun}
\alias{llik.fun.EF}
\alias{llik.fun.median}
\alias{llik.fun.obs}
\alias{llik.fun.obs.robust}
\alias{llik.fun2}
\alias{llik.fun3}
\alias{llik.grad}
\alias{llik.grad.obs}
\alias{llik.grad2}
\alias{llik.grad3}
\alias{llik.hessian}
\alias{llik.hessian.naive}
\alias{llik.hessian.obs}
\alias{llik.hessian2}
\alias{llik.info3}
\alias{llik.mcmcvar3}
\alias{logLik.stergm.CMLE}
\alias{logistftest}
\alias{mcmc.diagnostics.default}
\alias{midarrow}
\alias{model.matrix.pen.glm}
\alias{model.transform.formula}
\alias{mvmodel}
\alias{mvmodel.default}
\alias{mvmodel.ergm}
\alias{mvmodel.formula}
\alias{myLibLoc}
\alias{newnw.extract}
\alias{nologLik.message}
\alias{nvattr.copy.network}
\alias{ostar2deg}
\alias{paste.and}
\alias{plot.mcmc.list.ergm}
\alias{robust.inverse}
\alias{rspartnerdist}
\alias{set.offset.formula}
\alias{sociality.default}
\alias{sociality.ergm}
\alias{sociality.formula}
\alias{sociality.network}
\alias{sort.data.frame}
\alias{statnet.edit}
\alias{statnetShowDoc}
\alias{statnetbrowseURL}
\alias{stergm.CMLE}
\alias{stergm.EGMME}
\alias{stergm.EGMME.SA}
\alias{stergm.EGMME.SA.Phase2.C}
\alias{stergm.EGMME.initialfit}
\alias{stergm.RM}
\alias{stergm.getMCMCsample}
\alias{summary.statsmatrix.ergm}
\alias{sweep.mcmc.list}
\alias{term.list.formula}
\alias{twopathdist}
\alias{updatemodel.ErgmTerm}
\alias{vector.namesmatch}
\alias{xyplot.mcmc.list.ergm}

\description{
Internal ergm functions.}

\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).}

\seealso{ergm, statnet-package}
\keyword{internal}
