\name{ergm-internal}
\title{Internal ergm Objects}
% terms
\alias{ergm.checkargs}
\alias{ergm.checkdirected}
\alias{ergm.checkdegeneracy}
\alias{ergm.checkbipartite}
%
\alias{ConstraintImplications}
\alias{InitConstraint.bd}
\alias{InitConstraint.degreedist}
\alias{InitConstraint.degrees}                  
\alias{InitConstraint.edges}
\alias{InitConstraint.hamming}
\alias{InitConstraint.indegreedist}
\alias{InitConstraint.outdegreedist}
\alias{InitConstraint.nodedegrees}
%
\alias{check.ErgmTerm}
\alias{assignvariables}
\alias{InitErgmTerm.absdiff}
\alias{InitErgmTerm.absdiffcat}
\alias{InitErgmTerm.altkstar}
\alias{InitErgmTerm.asymmetric}
\alias{InitErgmTerm.b1degree}   
\alias{InitErgmTerm.b2degree}
\alias{InitErgmTerm.b1star}
\alias{InitErgmTerm.cycle}
\alias{InitErgmTerm.edgecov}
\alias{InitErgmTerm.isolates}
\alias{InitErgmTerm.match}
\alias{InitErgmTerm.mutual}
\alias{InitErgmTerm.nodefactor}
\alias{InitErgmTerm.nodeifactor}
\alias{InitErgmTerm.nodematch}
\alias{InitErgmTerm.nodemix}
\alias{InitErgmTerm.nodeofactor}
\alias{check.ErgmTerm.summarystats}
\alias{ergm.getglobalstats}
\alias{get.InitErgm.fname}
\alias{updatemodel.ErgmTerm}
\alias{zerowarnings}
\alias{extremewarnings}


%
\alias{InitErgm.absdiff}
\alias{InitErgm.absdiffcat}
\alias{InitErgm.b1concurrent}
\alias{InitErgm.b1factor}
\alias{InitErgm.b1degree}
\alias{InitErgm.altkstar}
\alias{InitErgm.b1star}
\alias{InitErgm.asymmetric}
\alias{InitErgm.balance}
\alias{InitErgm.bounded.degree}
\alias{InitErgm.bounded.idegree}
\alias{InitErgm.bounded.istar}
\alias{InitErgm.bounded.kstar}
\alias{InitErgm.bounded.odegree}
\alias{InitErgm.bounded.ostar}
\alias{InitErgm.bounded.triangle}
\alias{InitErgm.concurrent}
\alias{InitErgm.ctriple}
\alias{InitErgm.ctriad}
\alias{InitErgm.cycle}
\alias{InitErgm.degree}
\alias{InitErgm.density}
\alias{InitErgm.dsp}
\alias{InitErgm.dyadcov}
\alias{InitErgm.b2concurrent}
\alias{InitErgm.b2degree}
\alias{InitErgm.edgecov}
\alias{InitErgm.edges}
\alias{InitErgm.esp}
\alias{InitErgm.b2star}
\alias{InitErgm.b2factor}
\alias{InitErgm.gwb1degree}
\alias{InitErgm.gwdegree}
\alias{InitErgm.gwdsp}
\alias{InitErgm.gwb2degree}
\alias{InitErgm.gwesp}
\alias{InitErgm.gwidegree}
\alias{InitErgm.gwodegree}
\alias{InitErgm.hamming}
\alias{InitErgm.hammingmix.constant}
\alias{InitErgm.hammingmix}
\alias{InitErgm.idegree}
\alias{InitErgm.intransitive}
\alias{InitErgm.isolates}
\alias{InitErgm.istar}
\alias{InitErgm.kstar}
\alias{InitErgm.localtriangle}
\alias{InitErgm.m2star}
\alias{InitErgm.meandeg}
\alias{InitErgm.mutual}
\alias{InitErgm.nearsimmelian}
\alias{InitErgm.nodecov}
\alias{InitErgm.nodemain}
\alias{InitErgm.nodefactor}
\alias{InitErgm.nodeicov}
\alias{InitErgm.nodeifactor}
\alias{InitErgm.nodematch}
\alias{InitErgm.match}
\alias{InitErgm.nodemix}
\alias{InitErgm.mix}
\alias{InitErgm.nodeocov}
\alias{InitErgm.nodeofactor}
\alias{InitErgm.odegree}
\alias{InitErgm.ostar}
\alias{InitErgm.receiver}
\alias{InitErgm.sender}
\alias{InitErgm.simmelian}
\alias{InitErgm.simmelianties}
\alias{InitErgm.smalldiff}
\alias{InitErgm.sociality}
\alias{InitErgm.transitive}
\alias{InitErgm.triadcensus}
\alias{InitErgm.triadcensus.directedonly}
\alias{InitErgm.triangle}
\alias{InitErgm.triangles}
\alias{InitErgm.tripercent}
\alias{InitErgm.ttriple}
\alias{InitErgm.ttriad}
\alias{InitErgm.twopath}
%
\alias{InitMHP.CondDegree}
\alias{InitMHP.CondDegreeDist}
\alias{InitMHP.CondInDegreeDist}
\alias{InitMHP.CondOutDegreeDist}
\alias{InitMHP.ConstantEdges}
\alias{InitMHP.dissolution}
\alias{InitMHP.formation}
\alias{InitMHP.formationTNT}
\alias{InitMHP.HammingConstantEdges}
\alias{InitMHP.HammingTNT}
\alias{InitMHP.nobetweengroupties}
\alias{InitMHP.randomtoggle}
\alias{InitMHP.randomtoggleNonObserved}
\alias{InitMHP.TNT}
\alias{MHproposal}
\alias{MHproposal.character}
\alias{MHproposal.ergm}
\alias{MHproposal.formula}
\alias{MHproposal.MHproposal}
\alias{MHproposal.NULL}
\alias{MHproposals}
%
\alias{degreedistfactor}
\alias{drawcircle}
\alias{drawpie}
\alias{dspartnerdist}
\alias{ergm.compute.degeneracy}
\alias{ergm.curved}
\alias{ergm.curved.statsmatrix}
\alias{ergm.curved.update}
\alias{ergm.degenerate}
\alias{ergm.estimate}
\alias{ergm.estimate.only}
\alias{ergm.eta}
\alias{ergm.etagrad}
\alias{ergm.etagradmult}
\alias{ergm.etamap}
\alias{ergm.geodesicmatrix}
\alias{ergm.geodesicmatrix.edgelist}
\alias{ergm.geodesicmatrix.old}
\alias{ergm.geodist}
\alias{ergm.geodistdist}
\alias{ergm.geodistn}
\alias{ergm.getMCMCsample}
\alias{ergm.mcmcslave}
\alias{ergm.getnetwork}
\alias{ergm.independencemodel}
\alias{ergm.initialfit}
\alias{ergm.logisticdeviance}
\alias{ergm.logitreg}
\alias{ergm.mainfitloop}
\alias{ergm.marquardt}
\alias{ergm.marquardt2}
\alias{ergm.pl}
\alias{ergm.nodegeodesics}
\alias{ergm.pairgeodesic}
\alias{ergm.phase12}
\alias{ergm.raftery.diag}
\alias{ergm.revisetheta0}
\alias{ergm.rhs.formula} 
\alias{ergm.stocapprox}
\alias{ergm.t.summary}
\alias{espartnerdist}
\alias{get.node.attr}
\alias{is.ergm}
\alias{is.matrixnetwork}
\alias{llik.fun}
\alias{llik.fun.miss}
\alias{llik.fun2}
\alias{llik.fun3}
\alias{llik.grad}
\alias{llik.grad.miss}
\alias{llik.grad2}
\alias{llik.grad3}
\alias{llik.hessian}
\alias{llik.hessian.miss} 
\alias{llik.hessian2}
\alias{llik.info3}
\alias{llik.mcmcvar3}
\alias{mcmc.diagnostics.default}
\alias{midarrow}
\alias{newnw.extract}
\alias{plot.mcmc.ergm}
\alias{print.raftery.diag.ergm}
\alias{robust.inverse}
\alias{rspartnerdist}
\alias{set.mfrow}
\alias{traceplot.ergm}
\alias{twopathdist}
%
\alias{ergm.sufftoprob}
\alias{ergm.Cprepare}
\alias{ergm.getmodel}
\alias{ergm.getterms}
\alias{ergm.gettermnames}
\alias{ergm.mple}
\alias{ergm.statseval}
\alias{fulldistdist}
\alias{fullgcount}
\alias{statnet.edit}
%\alias{get.node.attr}
%\alias{geodist}
%\alias{has.loops}
%\alias{is.directed}
%\alias{is.hyper}
%\alias{lpz.dist}
%\alias{mahal}
%\alias{mlpY}
%\alias{mlpY.grad}
%\alias{mlpY0}
%\alias{mlpY0.grad}
%\alias{mlpYmdsZ}
%\alias{mlpYmdsZ.grad}
%\alias{mlpYnull}
%\alias{prepmahal}
%\alias{symmetrize}
%
\alias{ergm.MCMCse}
%\alias{ergm.latentplot}
%\alias{mlpY.grad.plot}
%\alias{mlpY.plot}
%\alias{mlpYmdsZ.grad.plot}
%\alias{mlpYmdsZ.plot}
\alias{ergm.MCMCacf}
\alias{ergm.pen.glm}
\alias{logistftest}
%
\alias{ergm.mainloop}
\alias{ergm.robmon}
%
\alias{InitMHP.TNT10}

%
% These need to be documented!
%

\alias{as.directed}
\alias{as.edgelist}
\alias{ergm.degeneracy} % At least this needs to be documented in summary.ergm and/or mcmc.diagnostics.ergm
\alias{print.network.series}
\alias{summary.statsmatrix.ergm}
\alias{summary.network.series}
\alias{license.statnet}
\alias{license.statnet1}
\alias{statnetShowDoc}
\alias{statnetbrowseURL}
\alias{ergm.bounddeg}
\alias{is.invertible}
\alias{ostar2deg}
\alias{sufftoprob}
\alias{degreedist}
\alias{mixingmatrix}
\alias{print.mixingmatrix}
\alias{print.summary.ergm} 
\alias{Summary.ergm.future}

\description{
  Internal ergm functions.
}
%\usage{
%anova.ergm(object, \dots)
%}
\details{
  Most of these are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
%\arguments{
% \item{\dots}{further arguments passed to or used by methods.}
%}
\seealso{ergm, statnet-package}
\keyword{internal}
