%  File man/ergm-internal.Rd in package ergm, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2017 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{ergm-internal}
\title{Internal ergm Objects}



%\alias{ergm.MCMC.packagenames}
%\alias{which.package.InitFunction}

%\alias{approx.hotelling.diff.test}


%\alias{control.ergm.toplevel}
%\alias{control.simulate.ergm.toplevel}

%\alias{ergm.MCMCse}
%\alias{ergm.MCMCse.lognormal}



%\alias{ergm.bridge.preproc}

%\alias{ergm.coefficient.degeneracy}
%\alias{ergm.compute.degeneracy}
%\alias{ergm.curved.statsmatrix}

%\alias{ergm.estimate}
\alias{ergm.logisticdeviance}
\alias{ergm.logitreg}
%\alias{ergm.mapl}
%\alias{ergm.maple}
%\alias{ergm.pen.glm}
%\alias{ergm.phase12}
%\alias{ergm.reviseinit}
%\alias{ergm.robmon}
%\alias{ergm.sample.eta2theta}
%\alias{ergm.sample.tomcmc}
%\alias{ergm.stepping}
%\alias{ergm.stocapprox}
%\alias{ergm.sufftoprob}
%\alias{ergm.t.summary}
%\alias{get.InitErgm.fname}
%\alias{init.methods}
%\alias{llik.fun}
%\alias{llik.fun.EF}
%\alias{llik.fun.median}
%\alias{llik.fun.obs}
%\alias{llik.fun.obs.robust}
%\alias{llik.fun2}
%\alias{llik.fun3}
%\alias{llik.grad}
%\alias{llik.grad.obs}
%\alias{llik.grad2}
%\alias{llik.grad3}
%\alias{llik.hessian}
%\alias{llik.hessian.naive}
%\alias{llik.hessian.obs}
%\alias{llik.hessian2}
%\alias{llik.info3}
%\alias{llik.mcmcvar3}
\alias{logistftest}

%\alias{midarrow}
%\alias{model.matrix.pen.glm}

%\alias{mk.conlist}
%\alias{myLibLoc}
%\alias{nologLik.message}
\alias{summary.statsmatrix.ergm}


%\alias{updatemodel.ErgmTerm}

%\alias{prune.conlist}
\alias{geweke.diag.ar}

%\alias{mk.edge.to.pos.lasttoggle.f}
%\alias{as.edgelist.matrix}

\description{
Internal \code{\link[=ergm-package]{ergm}} functions.}

\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).}

\seealso{ergm-package}
\keyword{internal}
