% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.design.R
\name{ergm.design}
\alias{ergm.design}
\title{Obtain the set of informative dyads based on the network structure.}
\usage{
ergm.design(nw, ...)
}
\arguments{
\item{nw}{a \code{\link{network}} object.}

\item{...}{term options.}
}
\value{
\code{ergm.design} returns a \code{\link{rlebdm}} of
informative (non-missing, non fixed) dyads.
}
\description{
Note that this function is not recommended for general use, since
it only supports only one way of specifying observational
structure---through \code{NA} edges. It is likely to be deprecated in
the future.
}
