% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R
\docType{data}
\name{ergm_GWDECAY}
\alias{ergm_GWDECAY}
\title{Curved settings for geometric weights for the \verb{gw*} terms}
\format{
An object of class \code{list} of length 3.
}
\usage{
ergm_GWDECAY
}
\description{
This is a list containing \code{map} and \code{gradient} for the weights described by \insertCite{Hu07c;textual}{ergm}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
