% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.getnetwork.R, R/ergm.getterms.R,
%   R/ergm_model.R, R/formula.utils.R
\name{ergm.getnetwork}
\alias{ergm.getnetwork}
\alias{ergm.getterms}
\alias{ergm_model}
\alias{ergm_model.formula}
\alias{ergm.getmodel}
\alias{offset.info.formula}
\title{Internal Functions for Querying, Validating and Extracting from ERGM
Formulas}
\usage{
ergm.getnetwork(formula, loopswarning = TRUE)

ergm.getterms(formula)

ergm_model(object, ...)

\method{ergm_model}{formula}(formula, nw, response = NULL, silent = FALSE,
  role = "static", ..., term.options = list(), extra.aux = list())

ergm.getmodel(object, ...)

offset.info.formula(object, response = NULL)
}
\arguments{
\item{formula}{a formula of the form \code{network ~ model.term(s)}}

\item{loopswarning}{whether warnings about loops should be printed (T or
F);default=TRUE}

\item{object}{object from which the model is to be constructed.}

\item{\dots}{additional parameters for model formulation}

\item{nw}{the network of interest}

\item{response}{charcter, name of edge attribute containing edge weights}

\item{silent}{logical, whether to print the warning messages from the
initialization of each model term; default=FALSE}

\item{role}{A hint about how the model will be used. Used primarily for
dynamic network models.}
}
\value{
\code{ergm.getnetwork} returns: \itemize{ \item the network
from the formula IF (i) the formula was correctly structured and
(ii) the network is found within the formula's enviornment }

\code{ergm.getterms} returns: \itemize{ \item the terms
object associated with the formula and returned by the native R
function \code{\link{terms}}. see
\code{\link[stats]{terms.object}} for details about the
components }

\code{ergm_model} returns an  \code{ergm_model} object as a list
containing:
\item{ formula}{the formula inputted to
\code{\link{ergm_model}}}
\item{coef.names}{a vector of coefficient names}
\item{offset}{a logical vector of whether each term was "offset", i.e.
fixed}
\item{terms}{a list of terms and 'term components' initialized by the
appropriate \code{InitErgmTerm.X} function.}
\item{network.stats0}{NULL always??}
\item{etamap}{the theta -> eta mapping as a list returned from
<ergm.etamap>}
}
\description{
The \code{ergm.getnetwork} function ensures that the
network in a given formula is valid; if so, the network is
returned; if not, execution is halted with warnings

DEPRECATED. The \code{ergm.getterms} function returns the terms of
a given formula and ensures that the formula is indeed a formula
with the necessary \code{~} operator

These are all functions that are generally not called directly by
users, but may be employed by other depending packages.  The
\code{ergm_model} function parses the given formula, and
initiliazes each ergm term via the \code{InitErgmTerm} functions to
create a \code{ergm_model} object.

\code{ergm.getmodel} is a deprecated name for the \code{ergm_model.formula} method.

\code{offset.info.formula} returns the offset
vectors associated with a formula.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Main method for constructing a model object from an \code{\link[=ergm]{ergm()}} formula of the form \code{network ~ model.term(s)} or \code{~ model.term(s)} with the network passed separately.
}}

\note{
These functions are not meant to be called by the end-user but may be useful to extension developers. This API is not to be considered fixed and may change between versions.
}
