% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_state.R, R/summary.ergm_model.R
\docType{data}
\name{ergm_state}
\alias{ergm_state}
\alias{ergm_state.edgelist}
\alias{ergm_state.matrix}
\alias{ergm_state.network}
\alias{is.ergm_state}
\alias{as.edgelist.ergm_state}
\alias{as.matrix.ergm_state}
\alias{as.network.ergm_state}
\alias{network.edgecount.ergm_state}
\alias{network.dyadcount.ergm_state}
\alias{network.size.ergm_state}
\alias{network.naedgecount.ergm_state}
\alias{\%ergmlhs\%.ergm_state}
\alias{\%ergmlhs\%<-.ergm_state}
\alias{as.rlebdm.ergm_state}
\alias{as.ergm_model.ergm_state}
\alias{is.curved.ergm_state}
\alias{param_names.ergm_state}
\alias{nparam.ergm_state}
\alias{update.ergm_state}
\alias{ergm_state.ergm_state}
\alias{ERGM_STATE_R_CHANGED}
\alias{ERGM_STATE_C_CHANGED}
\alias{ERGM_STATE_RECONCILED}
\alias{summary.ergm_state}
\title{A Representation of ERGM state}
\format{An object of class \code{integer} of length 1.}
\usage{
ergm_state(x, ...)

\method{ergm_state}{edgelist}(x, nw0, model = NULL, proposal = NULL, stats = NULL, ...)

\method{ergm_state}{matrix}(x, nw0, model = NULL, proposal = NULL, stats = NULL, ...)

\method{ergm_state}{network}(
  x,
  response = NULL,
  model = NULL,
  proposal = NULL,
  stats = NULL,
  ...
)

is.ergm_state(x)

\method{as.edgelist}{ergm_state}(x, ...)

\method{as.matrix}{ergm_state}(x, matrix.type = NULL, ...)

\method{as.network}{ergm_state}(x, ..., populate = TRUE)

\method{network.edgecount}{ergm_state}(x, na.omit = TRUE, ...)

\method{network.dyadcount}{ergm_state}(x, na.omit = TRUE, ...)

\method{network.size}{ergm_state}(x, ...)

\method{network.naedgecount}{ergm_state}(x, ...)

\method{\%ergmlhs\%}{ergm_state}(lhs, setting)

\method{\%ergmlhs\%}{ergm_state}(lhs, setting) <- value

\method{as.rlebdm}{ergm_state}(x, ...)

\method{as.ergm_model}{ergm_state}(x, ...)

\method{is.curved}{ergm_state}(object, ...)

\method{param_names}{ergm_state}(object, ...)

\method{nparam}{ergm_state}(object, ...)

\method{update}{ergm_state}(
  object,
  el = NULL,
  nw0 = NULL,
  response = NULL,
  model = NULL,
  proposal = NULL,
  stats = NULL,
  ...
)

\method{ergm_state}{ergm_state}(
  x,
  response = NULL,
  model = NULL,
  proposal = NULL,
  stats = NULL,
  ...
)

ERGM_STATE_R_CHANGED

ERGM_STATE_C_CHANGED

ERGM_STATE_RECONCILED

\method{summary}{ergm_state}(object, ...)
}
\arguments{
\item{nw0}{a \code{\link{network}} object, whose edges are absent or ignored.}

\item{model}{an \code{\link{ergm_model}} object.}

\item{response}{a character vector representing the response attribute. If
\code{NULL}, binary network is assumed.}

\item{na.omit}{Whether missing edges should be counted. Note that
missing edge information is not stored.}

\item{nw}{a \code{\link{network}} object.}
}
\value{
At this time, an \code{ergm_state} object is (subject to
change) a list containing the following elements:
\describe{

\item{el}{a \code{\link{tibble}} \code{\link{edgelist}} representing the edge state of the network}

\item{nw0}{a \code{\link{network}} object with all edges removed.}

\item{model}{an \code{\link{ergm_model}} object.}

\item{proposal}{an \code{\link{ergm_proposal}} object.}

\item{ext.state}{a list of length equalling to the number of terms in the model.}

\item{ext.flag}{one of \code{ERGM_STATE_R_CHANGED}, \code{ERGM_STATE_C_CHANGED}, and \code{ERGM_STATE_R_RECONCILED}.}

\item{stats}{a numeric vector of network statistics or some other
statistics used to resume.}}
}
\description{
\code{ergm_state} is an semi-internal class for passing
around results of MCMC sampling, particularly when the result is
used to start another MCMC sampler. It is deliberately loosely
specified, and its structure and even name are subject to change.
}
\section{Methods (by class)}{
\itemize{
\item \code{edgelist}: a method for constructing an ergm_state from an \code{\link{edgelist}} object and an empty \code{\link{network}}.

\item \code{matrix}: a method for constructing an ergm_state from a matrix object and an empty \code{\link{network}}.

\item \code{network}: a method for constructing an ergm_state from a \code{\link{network}} object.

\item \code{ergm_state}: Note that this method fails when
\code{na.omit=FALSE}, since missing edges are not stored.

\item \code{ergm_state}: Note that this method fails with
its default argument, since missing edges are not stored.

\item \code{ergm_state}: A stub that returns 0.

\item \code{ergm_state}: a method for updating an \code{ergm_state} and reconciling extended state.

\item \code{ergm_state}: a method for constructing an \code{ergm_state}.

\item \code{ergm_state}: a very low-level function that calculates summary statistics associated with an \code{\link{ergm_state}} object.
}}

\keyword{datasets}
\keyword{internal}
