% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{fixedas-ergmConstraint}
\alias{fixedas-ergmConstraint}
\alias{InitErgmConstraint.fixedas}
\title{Preserve and preclude edges}
\usage{
# fixedas(present, absent)
}
\arguments{
\item{present, absent}{edgelist or network}
}
\description{
Preserve the edges in 'present' and preclude the edges in 'absent'.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.
}
\concept{directed}
\concept{dyad-independent}
\concept{undirected}
