% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofN.R
\name{gofN}
\alias{gofN}
\alias{plot.gofN}
\title{Linear model diagnostics for multinetwork linear models}
\usage{
gofN(object, GOF, subset = TRUE, control = control.gof.ergm(), ...,
  obs.twostage = FALSE)

\method{plot}{gofN}(x, against = NULL, which = 1:2, col = 1,
  pch = 1, cex = 1, ..., ask = dev.interactive())
}
\arguments{
\item{object}{an \code{\link{ergm}} object.}

\item{GOF}{a one-sided \code{\link{ergm}} formula specifying network
statistics whose goodness of fit to test.}

\item{subset}{argument for the \code{\link[=ergm-terms]{N}} term.}

\item{control}{See \code{\link[=control.gof.ergm]{control.gof.ergm()}}.}

\item{\dots}{additional arguments to \code{\link[=simulate.ergm]{simulate.ergm()}} and
\code{\link[=summary.ergm_model]{summary.ergm_model()}}.}

\item{obs.twostage}{Either \code{FALSE} or an integer. If not \code{FALSE},
the constrained sample variance is estimated by simulating
conditional on the observed networks. However, a more accurate
estimate can be obtained via a two-stage process: \enumerate{

\item Sample networks from the model without the observational
constraint.

\item Conditional on each of those networks, sample with the
observational constraint, estimating the variance within each
sample and then averaging over the first-stage sample.

}

Then, \code{obs.twostage} specifies the number of unconstrained networks
to simulate from, which should divide the \code{\link[=control.gof.ergm]{control.gof.ergm()}}'s
\code{nsim} argument evenly.}

\item{against}{vector of values, network attribute, or a formula whose RHS gives an expression in terms of network attributes to plot against; if \code{NULL} (default), plots against fitted values.}

\item{which}{which to plot (\code{1} for residuals plot, \code{2} for \eqn{\sqrt{|R_i|}} scale plot).}

\item{col, pch, cex}{vector of values (wrapped in \code{\link[=I]{I()}}), network attribute, or a formula whose RHS gives an expression in terms of network attributes to plot against.}
}
\description{
Linear model diagnostics for multinetwork linear models
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: A plotting method, making residuals and scale-location plots.
}}

