% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofN.R
\name{gofN}
\alias{gofN}
\title{Linear model diagnostics for multinetwork linear models}
\usage{
gofN(object, GOF, subset = TRUE, control = control.gof.ergm(), ...)
}
\arguments{
\item{object}{an \code{\link{ergm}} object.}

\item{GOF}{a one-sided \code{\link{ergm}} formula specifying network statistics whose goodness of fit to test.}

\item{subset}{argument for the \code{\link[=ergm-terms]{N}} term.}

\item{\dots}{additional arguments to \code{simulate.ergm()} and \code{summary.ergm_model()}.}

\item{See}{\code{\link[=control.gof.ergm]{control.gof.ergm()}}.}
}
\description{
Linear model diagnostics for multinetwork linear models
}
