% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofN.R
\name{gofN}
\alias{gofN}
\alias{plot.gofN}
\alias{summary.gofN}
\title{Linear model diagnostics for multinetwork linear models}
\usage{
gofN(object, GOF = NULL, subset = TRUE,
  control = control.gofN.ergm(), ...)

\method{plot}{gofN}(x, against = NULL, which = 1:2, col = 1,
  pch = 1, cex = 1, ..., ask = length(which) > 1 &&
  dev.interactive(TRUE))

\method{summary}{gofN}(object, by = NULL, ...)
}
\arguments{
\item{object}{an \code{\link{ergm}} object.}

\item{GOF}{a one-sided \code{\link{ergm}} formula specifying network
statistics whose goodness of fit to test, or \code{\link{NULL}}; if \code{NULL},
uses the original model.}

\item{subset}{argument for the \code{\link[=ergm-terms]{N}} term.}

\item{control}{See \code{\link[=control.gofN.ergm]{control.gofN.ergm()}}.}

\item{\dots}{additional arguments to functions (\code{\link[=simulate.ergm]{simulate.ergm()}}
and \code{\link[=summary.ergm_model]{summary.ergm_model()}} for the constructor, \code{\link[=plot]{plot()}},
\code{\link[=qqnorm]{qqnorm()}}, and \code{\link[=qqline]{qqline()}} for the plotting method).}

\item{x}{a \code{\link{gofN}} object.}

\item{against}{vector of values, network attribute, or a formula whose RHS gives an expression in terms of network attributes to plot against; if \code{NULL} (default), plots against fitted values.}

\item{which}{which to plot (\code{1} for residuals plot, \code{2} for \eqn{\sqrt{|R_i|}}{sqrt(|R_i|)} scale plot, and \code{3} for normal quantile-quantile plot).}

\item{col, pch, cex}{vector of values (wrapped in \code{\link[=I]{I()}}), network attribute, or a formula whose RHS gives an expression in terms of network attributes to plot against.}

\item{ask}{whether the user should be prompted between the plots.}

\item{by}{a numeric or character vector, or a formula whose RHS gives an expression in terms of network attributes, used as a grouping variable for summarizing the values.}
}
\value{
An object of class \code{gofN}: a named list containing a list
for every statistic in the specified \code{GOF} formula with the
following elements vectors of length equal to the number of
subnetworks:

\item{observed}{For completely observed networks, their value of
the statistic. For partially observed networks, the expected value
of their imputations under the model.}

\item{fitted}{Expected value of the statistic under the model.}

\item{var}{Variance of the statistic under the model.}

\item{var.obs}{Conditional variance under imputation statistic.}

\item{pearson}{The Pearson residual computed from the above.}

In addition, the following \code{\link{attr}}-style attributes are included:

\item{nw}{The observed multinetwork object.}

\item{subset}{A logical vector giving the subset of networks that were used.}

\item{control}{Control parameters passed.}
}
\description{
Linear model diagnostics for multinetwork linear models
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: A plotting method, making residual and scale-location plots.

\item \code{summary}: A simple summary function.
}}

\examples{
data(samplk)
monks <- Networks(samplk1, samplk2, samplk3,samplk1, samplk2, samplk3,samplk1, samplk2, samplk3)
fit <- ergm(monks~N(~edges+nodematch("group")))
fit.gof <- gofN(fit) # GOF = original model
summary(fit.gof)
plot(fit.gof)
fit.gof <- gofN(fit, GOF=~triangles)
summary(fit.gof)
plot(fit.gof)

samplk1[1,]<-NA
samplk2[,2]<-NA
monks <- Networks(samplk1, samplk2, samplk3,samplk1, samplk2, samplk3,samplk1, samplk2, samplk3)
fit <- ergm(monks~N(~edges+nodematch("group")))
fit.gof <- gofN(fit) # GOF = original model
summary(fit.gof)
plot(fit.gof)
fit.gof <- gofN(fit, GOF=~triangles)
summary(fit.gof)
plot(fit.gof)

# Default is good enough in this case, but sometimes, we might want to set it higher. E.g.,
\dontrun{
fit.gof <- gofN(fit, GOF=~edges, control=control.gofN.ergm(nsim=400))
}

}
