% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valued.R
\name{is.valued}
\alias{is.valued}
\alias{is.valued.ergm_state}
\alias{is.valued.edgelist}
\alias{is.valued.ergm}
\alias{is.valued.network}
\title{Function to check whether an ERGM fit or some aspect of it is valued}
\usage{
is.valued(object, ...)

\method{is.valued}{ergm_state}(object, ...)

\method{is.valued}{edgelist}(object, ...)

\method{is.valued}{ergm}(object, ...)

\method{is.valued}{network}(object, ...)
}
\arguments{
\item{object}{the object to be tested.}

\item{...}{additional arguments for methods, currently unused.}
}
\description{
Function to check whether an ERGM fit or some aspect of it is valued
}
\section{Methods (by class)}{
\itemize{
\item \code{is.valued(ergm_state)}: a method for \code{\link{ergm_state}} objects.

\item \code{is.valued(edgelist)}: a method for \code{\link{edgelist}} objects.

\item \code{is.valued(ergm)}: a method for \code{\link{ergm}} objects.

\item \code{is.valued(network)}: a method for \code{\link{network}} objects that tests whether the network has been instrumented with a valued \code{\link{\%ergmlhs\%}} \code{"response"} specification, typically by \code{\link[=ergm_preprocess_response]{ergm_preprocess_response()}}. Note that it is \emph{not} a test for whether a network has edge attributes. This method is primarily for internal use.

}}
