% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.Cprepare.R
\name{lasttoggle}
\alias{lasttoggle}
\alias{last.toggle}
\alias{last-toggle}
\title{Storing last toggle information in a network}
\description{
An informal extension to \code{\link{network}} objects allowing
some limited temporal information to be stored.
WARNING: THIS DOCUMENTATION IS PROVIDED AS A COURTESY, AND THE API
DESCRIBED IS SUBJECT TO CHANGE WITHOUT NOTICE, DOWN TO COMPLETE
REMOVAL. NOT ALL FUNCTIONS THAT COULD SUPPORT IT DO. USE AT YOUR
OWN RISK.
}
\details{
While \code{\link[networkDynamic]{networkDynamic}} provides a flexible,
consistent method for storing dynamic networks, the \code{C} routines of
\code{\link[=ergm-package]{ergm}} and
\code{\link[tergm:tergm-package]{tergm}} required a simpler and more
lightweight representation.

Though this is an API intended for internal use, some functions,
like \code{\link[tergm]{stergm}} (for EGMME),
\code{\link[tergm:simulate.stergm]{simulate}}, and
\code{\link[=summary.formula]{summary}} can be passed networks with
this information using the following \code{\link{network}} (i.e.,
\code{\link{\%n\%}}) attributes:
\describe{

\item{\code{"time"}}{the
time stamp associated with the network}

\item{\code{"lasttoggle"}}{an integer vector analogous to the one returned by \code{\link[=to_ergm_Cdouble.network]{to_ergm_Cdouble.network()}} with an attribute: number of elements, a list of tails, a list of heads, and time the edge was last toggled.}
}

On the C side, it is represented by a hash table.

Again, this API is subject to change without notice.
}
\keyword{internal}
