\name{molecule}
\alias{molecule}
\title{Synthetic network with 20 nodes and 28 edges}
\description{
  This is a synthetic network of 20 nodes that is used as an example within 
  the \code{\link{ergm}} documentation. It has an interesting elongated
  shape - reminencent of a chemical molecule.
  It is stored as a \code{\link[network]{network}} object.
}
\usage{
 data(molecule)
}
\seealso{florentine, sampson, network, plot.network, ergm}
\keyword{datasets}
