%  File ergm/man/molecule.Rd
%  Part of the statnet package, http://statnet.org
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) in
%    http://statnet.org/attribution
%
%  Copyright 2012 the statnet development team
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{molecule}
\alias{molecule}
\title{Synthetic network with 20 nodes and 28 edges}
\description{
  This is a synthetic network of 20 nodes that is used as an example within 
  the \code{\link{ergm}} documentation. It has an interesting elongated
  shape - reminencent of a chemical molecule.
  It is stored as a \code{\link[network]{network}} object.
}
\usage{
 data(molecule)
}
\seealso{florentine, sampson, network, plot.network, ergm}
\keyword{datasets}
