%  File man/molecule.Rd in package ergm, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2014 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{molecule}
\alias{molecule}
\title{Synthetic network with 20 nodes and 28 edges}
\description{
  This is a synthetic network of 20 nodes that is used as an example within 
  the \code{\link{ergm}} documentation. It has an interesting elongated
  shape - reminencent of a chemical molecule.
  It is stored as a \code{\link[network]{network}} object.
}
\usage{
 data(molecule)
}
\seealso{florentine, sampson, network, plot.network, ergm}
\keyword{datasets}
