% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.multilayer.R
\name{network_view}
\alias{network_view}
\title{Extract a binary network "view".}
\usage{
network_view(x, attrname, test.f = as.logical)
}
\arguments{
\item{x}{a \code{\link{network}} object.}

\item{attrname}{a character vector of length 1 giving the name of
the edge attribute to test.}

\item{test.f}{an optional vector function that takes a vector of
selected edge attribute values and returns a logical vector
indicating whether an attribute should be kept; returning \code{NA}
sets the edge to missing.}
}
\value{
A \code{\link{network}} object with vertex and network attributes retained from \code{x}.
}
\description{
Returns a binary network with only those edges retained that fulfil
a specific criterion.
}
\examples{
data(florentine)
flo <- flomarriage
flo[,,add.edges=TRUE] <- as.matrix(flomarriage) | as.matrix(flobusiness)
flo[,, names.eval="m"] <- as.matrix(flomarriage)
flobusiness[3,5] <- NA
flo[,, names.eval="b"] <- as.matrix(flobusiness)
flo
(flob <- network_view(flo, "b"))
(flobusiness) # for comparison

stopifnot(all(identical(as.matrix(flobusiness),as.matrix(flob))))
}
