% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.node.attr.R
\name{node-attr-api}
\alias{node-attr-api}
\alias{ergm_get_vattr}
\alias{ergm_attr_levels}
\title{Helper functions for specifying nodal attribute levels

These functions are meant to be used in \code{InitErgmTerm} and other
implementations to provide the user with a way to extract nodal
attributes and select their levels in standardized and flexible
ways described under \code{\link{node-attr}}.}
\usage{
ergm_get_vattr(object, nw, bip = c("n", "b1", "b2"), ...)

ergm_attr_levels(object, attr, nw, levels = sort(unique(attr)), ...)
}
\arguments{
\item{object}{An argument specifying the nodal attribute to select
or which levels to include.}

\item{nw}{Network on the LHS of the formula.}

\item{bip}{Bipartedness mode: affects either length of attribute
vector returned or the length permited: \code{"n"} for full network, \code{"b1"}
for first mode of a bipartite network, and \code{"b2"} for the second.}

\item{...}{Additional argument to the functions of network or to
the formula's environment.}

\item{attr}{A vector of length equal to the number of nodes,
specifying the attribute vector.}

\item{levels}{Starting set of levels to use; defaults to the sorted
list of unique attributes.}

\item{type}{Permitted types of the attribute; no checking by
default.}
}
\value{
A vector of length equal to the number of nodes giving the
selected attribute function. It may also have an attribute
\code{"name"}, which controls the suggested name of the attribute
combination.

A vector of levels to use and their order.
}
\description{
\code{ergm_get_vattr} extracts and processes the specified
nodal attribute vector. It is strongly recommended that
\code{\link[=check.ErgmTerm]{check.ErgmTerm()}}'s corresponding
\code{vartype="function,formula,character"}.

\code{ergm_attr_levels} filters the levels of the
attribute.  It is strongly recommended that \code{\link[=check.ErgmTerm]{check.ErgmTerm()}}'s
corresponding
\code{vartype="function,formula,character,numeric,logical,NULL"}
}
