% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{nodematch-ergmTerm}
\alias{nodematch-ergmTerm}
\alias{InitErgmTerm.nodematch}
\alias{InitWtErgmTerm.nodematch}
\title{Uniform homophily and differential homophily}
\usage{
# binary: nodematch(attr, diff=FALSE, keep=NULL, levels=NULL)

# valued: nodematch(attr, diff=FALSE, keep=NULL, levels=NULL, form="sum")
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{keep}{This argument is retained for backwards compatibility and may be
removed in a future version. When both \code{keep} and \code{levels} are passed,
\code{levels} overrides \code{keep}.}

\item{levels}{The optional \code{levels} argument controls which levels of the attribute
should be included and which should be excluded. (See Specifying Vertex
attributes and Levels (\code{?nodal_attributes}) for details.) For example, if
the "fruit" attribute has levels "orange", "apple", "banana", and "pear",
then to include just two levels, one for "apple" and one for "pear", use
any of `}
}
\description{
Uniform homophily and differential homophily
}
\details{
When \code{diff=FALSE} , this term adds one network statistic
to the model, which counts the number of edges \eqn{(i,j)} for which
\code{attr(i)==attr(j)} . This is also called ''uniform homophily,'' because each group is assumed to have the same propensity for within-group ties. When multiple attribute names are given, the
statistic counts only ties for which all of the attributes
match. When \code{diff=TRUE} , \eqn{p} network statistics are added
to the model, where \eqn{p} is the number of unique values of the
\code{attr} attribute. The \eqn{k} th such statistic counts the
number of edges \eqn{(i,j)} for which \verb{attr(i) == } \code{	attr(j) == value(k)} , where \code{value(k)} is the \eqn{k} th
smallest unique value of the \code{attr} attribute. This is also called ''differential homophily,'' because each group is allowed to have a unique propensity for within-group ties. Note that a statistical test of uniform vs. differential homophily should be conducted using the ANOVA function.

By default, matches on all levels \eqn{k} are
counted.nodematch This works for both
\code{diff=TRUE} and \code{diff=FALSE} .
}
\concept{categorical nodal attribute}
\concept{directed}
\concept{dyad-independent}
\concept{frequently-used}
\concept{match}
\concept{undirected}
