% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.R, R/InitWtErgmTerm.R
\name{nodeocov-ergmTerm}
\alias{nodeocov-ergmTerm}
\alias{InitErgmTerm.nodeocov}
\alias{InitWtErgmTerm.nodeocov}
\title{Main effect of a covariate for out-edges}
\usage{
# binary: nodeocov(attr)

# valued: nodeocov(attr, form="sum")
}
\arguments{
\item{attr}{a vertex attribute specification (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}

\item{form}{character how to aggregate tie values in a valued ERGM}
}
\description{
This term adds a single network statistic for each quantitative attribute or matrix column to the model equaling the total
value of \code{attr(i)} for all edges \eqn{(i,j)} in the network. This
term may only be used with directed networks. For categorical attributes,
see \code{nodeofactor} .
}
\note{
\pkg{ergm} versions 3.9.4 and earlier used different arguments for this
term. See \code{\link{ergm-options}} for how to invoke the old behaviour.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "nodeocov", "subsection")}
}
\concept{directed}
\concept{dyad-independent}
\concept{quantitative nodal attribute}
