% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_names.R
\name{param_names}
\alias{param_names}
\alias{param_names.default}
\title{Names of the parameters associated with an object.}
\usage{
param_names(object, ...)

\method{param_names}{default}(object, ...)
}
\arguments{
\item{object}{An object for which parameter names are defined.}

\item{...}{Additional arguments to methods.}
}
\description{
This is a generic that returns a vector giving the names of the parameters associated with a model or a model fit.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: By default, the names of the \code{\link[=coef]{coef()}} vector is returned.
}}

