% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ergm.R
\name{predict.formula}
\alias{predict.formula}
\alias{predict.ergm}
\title{ERGM-based conditional tie probabilities}
\usage{
\method{predict}{formula}(object, theta, output = c("data.frame",
  "matrix"), ...)

\method{predict}{ergm}(object, ...)
}
\arguments{
\item{object}{a formula or a fitted ERGM model object}

\item{theta}{numeric vector of ERGM model parameter values}

\item{output}{character, type of object returned. Defaults to \code{"data.frame"}.
See section Value below.}

\item{...}{other arguments passed to/from other methods}
}
\value{
Type of object returned depends on the argument \code{output}. If
\code{output="data.frame"} the function will return a data frame with columns:
\itemize{
\item \code{tail},\code{head} -- indices of nodes identifying a dyad
\item \code{p} -- predicted conditional tie probability
If \code{output="matrix"} the function will return an "adjacency matrix" with the
predicted conditional tie probabilities.
}
}
\description{
Calculate predicted \strong{conditional} tie probabilities for all dyads in a
network based on an ERGM model. Currently there are two methods implemented:
\itemize{
\item \code{\link[=predict.formula]{predict.formula()}} requires an ERGM model formula with an existing network
object on the left hand side, and a vector of corresponding parameter values.
\item \code{\link[=predict.ergm]{predict.ergm()}} requires only a fitted model object
}
}
