% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlebdm.R
\name{rlebdm}
\alias{rlebdm}
\alias{as.rlebdm}
\alias{as.rlebdm.matrix}
\alias{as.rlebdm.edgelist}
\alias{as.matrix.rlebdm}
\alias{print.rlebdm}
\alias{!.rlebdm}
\alias{|.rlebdm}
\alias{&.rlebdm}
\alias{<.rlebdm}
\alias{>.rlebdm}
\alias{<=.rlebdm}
\alias{>=.rlebdm}
\alias{==.rlebdm}
\alias{!=.rlebdm}
\alias{as.edgelist.rlebdm}
\title{RLE-Compressed Boolean Dyad Matrix}
\usage{
rlebdm(x, n)

as.rlebdm(x, ...)

\method{as.rlebdm}{matrix}(x, ...)

\method{as.rlebdm}{edgelist}(x, ...)

\method{as.matrix}{rlebdm}(x, ...)

\method{print}{rlebdm}(x, compact = TRUE, ...)

\method{!}{rlebdm}(x)

\method{|}{rlebdm}(e1, e2)

\method{&}{rlebdm}(e1, e2)

\method{<}{rlebdm}(e1, e2)

\method{>}{rlebdm}(e1, e2)

\method{<=}{rlebdm}(e1, e2)

\method{>=}{rlebdm}(e1, e2)

\method{==}{rlebdm}(e1, e2)

\method{!=}{rlebdm}(e1, e2)

\method{as.edgelist}{rlebdm}(x, ...)
}
\arguments{
\item{x}{for [rlebdm()], an [rle()] object or a vector that is converted to one; it will be coerced to [logical()] before processing; for [as.rlebdm.matrix()], a matrix.}

\item{n}{the dimensions of the square matrix represented.}

\item{compact}{whether to print the matrix compactly (dots and stars) or to print it as a logical matrix.}

\item{e1, e2}{arguments to the binary operations.}
}
\description{
A simple class representing boolean (logical) square matrix
run-length encoded in a column-major order.
}
