% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.R
\name{SPDyad-ergmProposal}
\alias{SPDyad-ergmProposal}
\alias{InitErgmProposal.SPDyad}
\title{A proposal alternating between TNT and a triad-focused
proposal}
\description{
The specified proportion of the time, the proposal
proceeds along the lines of \insertCite{WaAt13a;textual}{ergm}, albeit
with different weighting. A dyad is selected uniformly at random
from among those dyads with at least one shared partnership or
transitivity of the specified type.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="SPDyad"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="SPDyad"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="SPDyad"))}}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "SPDyad", "subsection")}
}
\keyword{internal}
