% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.networks.R
\name{split.network}
\alias{split.network}
\title{A \code{\link[=split]{split()}} method for \code{\link[network:network]{network::network}} objects.}
\usage{
\method{split}{network}(x, f, drop = FALSE, sep = ".", lex.order = FALSE,
  ...)
}
\arguments{
\item{x}{a \code{\link[network:network]{network::network}} object.}

\item{f, drop, sep, lex.order}{see \code{\link[=split]{split()}}; note that \code{f} must have length equal to \code{network.size(x)}.}

\item{...}{additional arguments, currently unused.}
}
\value{
A \code{\link{network.list}} containing the networks. These networks
will inherit all vertex and edge attributes, as well as relevant
network attributes.
}
\description{
Split a network into subnetworks on a factor.
}
\seealso{
\code{\link[network:get.inducedSubgraph]{network::get.inducedSubgraph()}}
}
