% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ergm_model.R
\name{summary.ergm_model}
\alias{summary.ergm_model}
\title{Evaluate network summary statistics from an initialized ergm model}
\usage{
\method{summary}{ergm_model}(object, nw = NULL, ...)
}
\arguments{
\item{object}{an \code{\link{ergm_model}} object.}

\item{nw}{a \code{\link{network}} whose statistics are to be evaluated,
though an \code{\link{ergm_state}} object will also work. If \code{NULL},
returns empty network's statistics for that model.}

\item{...}{Further arguments to methods.}
}
\description{
Returns a vector of the model's statistics for a given network or
an empty network. This is a low-level function that should not be
used by end-users, but may be useful to developers.
}
\seealso{
\code{\link[=summary_formula]{summary_formula()}}
}
\keyword{internal}
