% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R
\name{symmetrize}
\alias{symmetrize}
\alias{symmetrize.default}
\alias{symmetrize.network}
\title{Return a symmetrized version of a binary network}
\usage{
symmetrize(x, rule = c("weak", "strong", "upper", "lower"), ...)

\method{symmetrize}{default}(x, rule = c("weak", "strong", "upper", "lower"),
  ...)

\method{symmetrize}{network}(x, rule = c("weak", "strong", "upper", "lower"))
}
\arguments{
\item{x}{an object representing a network.}

\item{rule}{a string specifying how the network is to be
symmetrized; see \code{\link[sna:symmetrize]{sna::symmetrize()}} for details.}
}
\description{
Return a symmetrized version of a binary network
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: The default method, passing the input on to \code{\link[sna:symmetrize]{sna::symmetrize()}}.

\item \code{network}: A method for \code{\link{network}} objects, which preserves network and vertex attributes.
}}

