% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.operator.R, R/ergm.Cprepare.R,
%   R/rlebdm.R, R/to_ergm_Cdouble.R
\name{to_ergm_Cdouble.character}
\alias{to_ergm_Cdouble.character}
\alias{to_ergm_Cdouble.ergm_model}
\alias{to_ergm_Cdouble.network}
\alias{to_ergm_Cdouble.matrix}
\alias{to_ergm_Cdouble.rlebdm}
\alias{to_ergm_Cdouble}
\title{Methods to serialize objects into numeric vectors for passing to the C side.}
\usage{
\method{to_ergm_Cdouble}{character}(x, ...)

\method{to_ergm_Cdouble}{ergm_model}(x, ...)

\method{to_ergm_Cdouble}{network}(x, attrname = NULL, ...)

\method{to_ergm_Cdouble}{matrix}(x, prototype = NULL, ...)

\method{to_ergm_Cdouble}{rlebdm}(x, ...)

to_ergm_Cdouble(x, ...)
}
\arguments{
\item{x}{object to be serialized.}

\item{...}{arguments for methods.}

\item{attrname}{name of an edge attribute.}

\item{prototype}{A network whose relevant attributes (size,
directedness, bipartitedness, and presence of loops) are imposed
on the output edgelist if \code{x} is already an edgelist. (For
example, if the prototype is undirected, \code{ergm.Cprepare.el}
will ensure that \eqn{t < h}.)}
}
\value{
The \code{character} method takes a character vector of length 1 and returns a numerical vector encoding it. It concatenates the following:
\itemize{
\item the number of characters in the input string and
\item the characters in the input string encoded using ASCII encoding.
}

This is intended to be decoded by \code{unpack_str_as_double()} C routines.

The \code{ergm_model} method returns a numeric vector concatenating the following:
\itemize{
\item number of terms in the model;
\item length of and encoded string of term names;
\item length of and encoded string of library names; and
\item vector of intputs to the model.
}

This is intended to be decoded by \code{unpack_*Model_as_double()} C routines.

A \code{\link{double}}-precision vector.

For serialization of edge lists, this usually takes the form of a
\eqn{2 e + 1}-vector, whose first element is the number of edges
and whose subsequent elements are tails and heads of the edges,
sorted in tail-major order, with an optional third attribute
column.
}
\description{
Methods to serialize objects into numeric vectors for passing to the C side.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{ergm_model}: 

\item \code{network}: Method for \code{\link{network}} objects.

\item \code{matrix}: Method for \code{\link{matrix}} objects, assumed to be edgelists.

\item \code{rlebdm}: Method for \code{\link{rlebdm}} objects, serialized as number of dyads, then
number of runs of nonzeros, starting points positions of the runs,
and cumulative lenght of the runs.
}}

