% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_ergm.R
\name{vif_ergm}
\alias{vif_ergm}
\title{Variance Inflation Factors for ERGMs}
\usage{
vif_ergm(object, ...)
}
\arguments{
\item{model}{model object as returned by \code{\link[=ergm]{ergm()}}}
}
\value{

}
\description{
Variance Inflation Factors for ERGMs
}
\references{
Duxbury, S. W. (2018). Diagnosing multicollinearity in exponential random
graph models. Sociological Methods & Research \url{doi:10.1177/0049124118782543}
}
\author{
Based on original code by Scott W. Duxbury
}
